==> 2_EpidemicModel-Stochastic.Rmd <==
# Simulating an Epidemic (Stochastic)

## Required Packages

```{r, include = FALSE}
library(deSolve)
library(FME)
library(GillespieSSA)
require(gdata)
require(doBy)
require(ecp)
require(strucchange)
```

# Set-Up Objects

```{r}
minTemp <- 273.15 # The minimum temperature for which S_c is calculated
maxTemp <- 313.15 # The maximum temperature for which S_c is calculated
resol <- 0.01 # Resolution
temp <- seq(from = minTemp, to = maxTemp, by = resol)

# Initialize vectors for each of the temperature-dependent parameters that determine S_c
mu <- vector(mode = "numeric", length(temp))
beta <- vector(mode = "numeric", length(temp))
alpha <- vector(mode = "numeric", length(temp)) 
chi <- vector(mode = "numeric", length(temp))
lambda <- vector(mode = "numeric", length(temp))
sigma <- vector(mode = "numeric", length(temp))
theta <- vector(mode = "numeric", length(temp))
omega <- vector(mode = "numeric", length(temp))
equil.abundance <- vector(mode = "numeric", length(temp))
```

# Defining MTE Parameters

Note when parameterizing the *PLOS Biol.* paper, Devin used 15C as T~0~, and in *AmNat* paper, he used 20C as T~0~. `cs` is contact allometric scaling; `is` is infection rate allometric scaling.

```{r}
# Normalization constant (i.e.. value at reference temperature)
chi0 <- 1.255755e+02 ## DET. SIM.: 1.304617e+02 
sigma0 <- 5.684252e-09
omega0 <- 131
beta0 <- 2.281060732
mu0 <- 0.007507025                       

# Activation energy (THIS IS WHAT I WANTED TO PLAY AROUND WITH) 
E_chi <- 3.073807e-01 ## DET. SIM.: 2.90155e-01
E_sigma <- 1.106441     
E_omega <- -0.295
E_beta <- 0.390532437  
E_mu <- 0.801482243 

# 'Inactivation' energy at lower threshold (taken as per Molnar et al. 2013)         
E_sigmaL <- 6.766889
E_omegaL <- 357                            
E_muL <- E_mu*5

# 'Inactivation' energy at upper threshold (taken as per Molnar et al. 2013) 
E_chiH <- 4.196706e+00 ## DET. SIM.: 4.129647
E_sigmaH <- 5.661312
E_omegaH <- 6.32
E_betaH <- 1.472373114
E_muH <- 4.239515185          

# Reference temperatures (taken as 20C for contact and infection, 15 for mu and omega)   
T0_chi <- 293.15
T0_sigma <- 293.15
T0_omega <- 288.15
T0_beta <- 288.15
T0_mu <- 288.15                      

# Lower thresholds      
T_sigmaL <- 273.2 + 14.14539
T_omegaL <- 273.2 + 9.9
T_muL <- 8.957624193 + 273.15
# Upper thresholds 
T_chiH <- 273.15 + 3.455977e+01 ## DET. SIM.: + 3.513025e+01
T_sigmaH <- 273.15 + 23.43956
T_omegaH <- 273.15 + 23.9
T_betaH <- 28.8419577 + 273.15
T_muH <- 30.23264714 + 273.15                        

cs <- 0.2574784
is <- -1.181799

k <- 8.62*10^(-5)
```

# Assumptions

1.  Within-host parasite abundance as a function of equilibrium abundance:

```{r}
abundance.prop <- 0.182
```

2.  How long clusters take to burst:

```{r}
burst.time <- 4.5 ## DET. SIM.: 7
```

3.  Spores per cluster:

```{r}
spores.per.cluster <- 12 ## DET. SIM.: 24
```

4.  How many spores are released into the environment from each burst:

```{r}
burst.size <- 0.5
```

5.  Daphnia size and mass:

```{r}
size <- 2700
mass <- (0.009*(size*0.001)^2.63)*0.001
```

# Parameter Values

## Temperature-Dependent Parameters Along TPCs

Note: In this loop, `sigma` is the infection rate; we'll calculate infection probability in the next step.

```{r}
for(t in seq(temp) ){
  chi[t] <- ((((chi0*(mass^cs)*exp((-E_chi/k)*(1/temp[t] - 1/T0_chi))/(1 + exp(E_chiH/k*(-1/temp[t] + 1/T_chiH)))))/3.5E7)*1440) # Values at the end are to standardize to the volume and time of the experiment
  
  sigma[t] <- sigma0*(mass^is)*exp((-E_sigma/k)*(1/temp[t] - 1/T0_sigma))*(1 + exp(E_sigmaL/k*(1/temp[t] - 1/T_sigmaL)) + exp(E_sigmaH/k*(-1/temp[t] + 1/T_sigmaH)))^(-1) 
  
  equil.abundance[t] <-  omega0*exp((-E_omega/k)*(1/temp[t] - 1/T0_omega))*(1 + exp(E_omegaL/k*(1/temp[t] - 1/T_omegaL)) + exp(E_omegaH/k*(-1/temp[t] + 1/T_omegaH)))^(-1)
  
  omega[t] <- (equil.abundance[t]*abundance.prop)*spores.per.cluster
  
  alpha[t] <- (equil.abundance[t]*abundance.prop)*5.12E-6; # Parasite intensity multiplied by per parasite virulence
  
  lambda[t] <- (equil.abundance[t]*abundance.prop)*(spores.per.cluster*burst.size)/burst.time
  
  mu[t] <- mu0*exp((-E_mu/k)*(1/temp[t] - 1/T0_mu))*(1 + exp(E_muL/k*(1/temp[t] - 1/T_muL)) + exp(E_muH/k*(-1/temp[t] + 1/T_muH)))
  
  beta[t] <- beta0*exp((-E_beta/k)*((1/temp[t]) - (1/T0_beta)))*((1 + exp((E_betaH/k)*((-1/temp[t]) + (1/T_betaH))))^(-1))
}
```

### Calculating Infection Probability

To calculate infection probability, we need to calculate contact in microliters per minute, to calculate gut residence time.

Contact rate is measured in microliters/minute, but we need to convert to 35L/24 hours to standardize to the volume of the experiment. There are $3.5 \cdot 10^7$ microliters in 35L, and 1440 minutes in a day.

```{r}
standardized.contact <- c()
for(i in 1:length(temp)){
  standardized.contact[i] <- ((chi0*(mass^cs)*exp((-E_chi/k)*(1/temp[i] - 1/T0_chi))/(1 + exp(E_chiH/k*(-1/temp[i] + 1/T_chiH)))))
}
```

Next, calculate ingestion rate (cells/hour), and we have 10 algal cells per microliter:

```{r}
ing.cells.rate <- ((standardized.contact)*10)*60
```

From Kooijman 1993 and Evers and Kooijman: $gut.res.time = \frac{gv * length^3}{ing.cells.rate}$, in hours. Note that the concentration of algae here was much lower than in Kirk et al. (2019). Therefore, ingestion of cells is lower and gut residence time is longer.

```{r}
gut.res.time <- (9900*((size/1000)^3))/ing.cells.rate; head(gut.res.time)
```

And finally, let's calculate the probability of infection:

```{r}
sigma.prob <- 1 - exp(-sigma*gut.res.time); head(sigma.prob)
```

## Temperature-Dependent Parameters at Experimental Temperatures

The following parameters are given values across a range of temperatures by the MTE models above:

-   `chi`

-   `sigma`

-   `omega`

-   `alpha`

-   `lambda`

-   `mu`

-   `beta`

Note that `[1001]` refers to the parameter's value at 10.0ºC, while `[1351]` refers to the parameter's value at 13.5ºC.

```{r}
indices <- c(1001, 1051, 1101, 1151, 1201, 1251, 1301, 1351)

# Contact rate 
# Temperature dependent, calculate from AmNat paper 
chi.t <- c(chi[indices])
  
# Probability of infection
# Temperature dependent, calculate from AmNat paper 
sigma.t <- c(sigma.prob[indices])

# Parasite virulence rate 
alpha.t <- c(alpha[indices])

# Parasites released at death 
omega.t <- c(omega[indices])

# Parasite shedding rate 
lambda.t <- c(lambda[indices])
  
# Natural mortality rate
# Temperature dependent, calculate from PLoS Biol. paper

# Temp[101] = 10C, so we want 101, 106, 111, 116, 121, 126, 131, 136 to get to 13.5C

for(i in indices){
  mu.sim <- mu[i]
  beta.sim <- beta[i]
  
  U <- 1; # Starting conditions
  N0 <- c(U)
  
  TT <- seq(0.1, 400, 0.1) # Set amount of time steps to run in simulation
  step <- 0.1
  
  parms <- c(beta.sim, mu.sim)
  
  survival <- function(t, y, p){
    beta <- p[1]; mu <- p[2];
    U <- y[1];
    
    dU <- -beta*mu^(beta)*(t^(beta - 1))*U
    
    list(c(dU))
  }
  # Run lsoda	
  predictions <- lsoda(N0, TT, survival, parms)	
  
  nam <- paste("predictions", i, sep = "")
  assign(nam, predictions)
}

# Point at which survival of infectids is 50%
surv.10 <- predictions1001[,2]
surv.105 <- predictions1051[,2]
surv.11 <- predictions1101[,2]
surv.115 <- predictions1151[,2]
surv.12 <- predictions1201[,2]
surv.125 <- predictions1251[,2]
surv.13 <- predictions1301[,2]
surv.135 <- predictions1351[,2]

# 50% survival between index 1226 and index 1227
# Crosses 1% survival between index 3764 and 3765 

# 1 / lifespan gives mortality rate 
# Will take lifespan as the lifespan of the average individual 
# Which will be the first timepoint where survival has crossed or equals 50% 

hazard.10 <- 1/predictions1001[which(surv.10 <= 0.5)[[1]], 1]
hazard.105 <- 1/predictions1051[which(surv.105 <= 0.5)[[1]], 1]
hazard.11 <- 1/predictions1101[which(surv.11 <= 0.5)[[1]], 1]
hazard.115 <- 1/predictions1151[which(surv.115 <= 0.5)[[1]], 1]
hazard.12 <- 1/predictions1201[which(surv.12 <= 0.5)[[1]], 1]
hazard.125 <- 1/predictions1251[which(surv.125 <= 0.5)[[1]], 1]
hazard.13 <- 1/predictions1301[which(surv.13 <= 0.5)[[1]], 1]
hazard.135 <- 1/predictions1351[which(surv.135 <= 0.5)[[1]], 1]

# This is the "new" mu, which is the mortality rate assuming exponential hazard 
mu.t <- as.numeric(c(hazard.10, hazard.105, hazard.11, hazard.115, 
                     hazard.12, hazard.125, hazard.13, hazard.135))
```

## Temperature-Independent Parameters

Note: Many parameters will be vectors of length eight (one for each of eight temperature treatments, from 10ºC to 13.5ºC).

Input from stocks:

-   Uninfected: 0.5350318

-   Infected: 0.4649682

"Input S" would be $(9+0.5350318)\cdot 3$. "Input I" would be $0.4649682 \cdot 3$. We can convert these to a daily metric by dividing by 3.

```{r}
inputS = (3 + 0.5350318)*3/3
inputI = 0.4649682*3/3

# Maximum recruitment rate 
# Can be temperature dependent but probably doesn't need to be if sufficiently high
recruit = 1.33
  
# Carrying capacity #
# Doesn't need to be temperature dependent
# This is the mean abundance across all populations throughout the experiment
kappa = 170

# Harvesting rate 
# 12 every three days, so 4 per day, but per capita would be 0.025 because 4/170 = 0.0235
h = 0.0235

# Degradation of dead infecteds rate 
theta = 0.1

# Environmental parasite mortality rate (induced via changing the medium)
# We take 3L out of the 35L every 3 days (equivalent of 1L per day, or 0.02857143 mortality for spores) 
# For now we will assume this is their natural mortality rate (i.e. it swamps out any other environmental mortality)
gamma <-  0.02857143
```

# Organizing Parameters by Temperature Treatment

```{r}
parms10 <- c(recruit = recruit, chi.t = chi.t[1], sigma.t = sigma.t[1], mu.t = mu.t[1], alpha.t = alpha.t[1], theta = theta, lambda.t = lambda.t[1], omega.t = omega.t[1], gamma = gamma, h = h, inputS = inputS, inputI = inputI)

parms105 <- c(recruit = recruit, chi.t = chi.t[2], sigma.t = sigma.t[2], mu.t = mu.t[2], alpha.t = alpha.t[2], theta = theta, lambda.t = lambda.t[2], omega.t = omega.t[2], gamma = gamma, h = h, inputS = inputS, inputI = inputI)

parms11 <- c(recruit = recruit, chi.t = chi.t[3], sigma.t = sigma.t[3], mu.t = mu.t[3], alpha.t = alpha.t[3], theta = theta, lambda.t = lambda.t[3], omega.t = omega.t[3], gamma = gamma, h = h, inputS = inputS, inputI = inputI)

parms115 <- c(recruit = recruit, chi.t = chi.t[4], sigma.t = sigma.t[4], mu.t = mu.t[4], alpha.t = alpha.t[4], theta = theta, lambda.t = lambda.t[4], omega.t = omega.t[4], gamma = gamma, h = h, inputS = inputS, inputI = inputI)

parms12 <- c(recruit = recruit, chi.t = chi.t[5], sigma.t = sigma.t[5], mu.t = mu.t[5], alpha.t = alpha.t[5], theta = theta, lambda.t = lambda.t[5], omega.t = omega.t[5], gamma = gamma, h = h, inputS = inputS, inputI = inputI)

parms125 <- c(recruit = recruit, chi.t = chi.t[6], sigma.t = sigma.t[6], mu.t = mu.t[6], alpha.t = alpha.t[6], theta = theta, lambda.t = lambda.t[6], omega.t = omega.t[6], gamma = gamma, h = h, inputS = inputS, inputI = inputI)

parms13 <- c(recruit = recruit, chi.t = chi.t[7], sigma.t = sigma.t[7], mu.t = mu.t[7], alpha.t = alpha.t[7], theta = theta, lambda.t = lambda.t[7], omega.t = omega.t[7], gamma = gamma, h = h, inputS = inputS, inputI = inputI)

parms135 <- c(recruit = recruit, chi.t = chi.t[8], sigma.t = sigma.t[8], mu.t = mu.t[8], alpha.t = alpha.t[8], theta = theta, lambda.t = lambda.t[8], omega.t = omega.t[8], gamma = gamma, h = h, inputS = inputS, inputI = inputI)
```

# Implementing the Stochastic Model

## Set-Up

Writing the state change matrix and model for stochastic implementation:

Note: Could we make the sampling and input parameters deterministic?

```{r}
nu <- matrix(c(+1, -1, -1, 0, 0, 0, 0, 0, 0, -1, 0, +1, 0, 
               0, +1, 0, -1, -1, 0, 0, 0, 0, 0, -1, 0, +1, 
               0, 0, 0, +1, +1, -1, 0, 0, 0, 0, 0, 0, 0, 
               0, 0, 0, 0, 0, 0, +1, +1, -1, 0, 0, 0, 0), 
             nrow = 4, byrow = TRUE)

a <- c("recruit",              # Density-independent recruitment
       "chi.t*sigma.t*S*E",    # Transmission
       "mu.t*S",               # Natural death of susceptible
       "mu.t*I",               # Natural death of infected
       "alpha.t*I",            # Parasite-induced death of infected
       "theta*D",              # Decay and loss of infected corpse
       "lambda.t*I",           # Parasite spore shed by infected
       "omega.t*theta*D",      # Parasite spore shed by corpse
       "gamma*E",              # Parasite death
       "h*(S+I)*(S/(S+I))",    # Sampling uninfected
       "h*(S+I)*(I/(S+I))",    # Sampling infected
       "inputS",               # Inputting susceptible
       "inputI")               # Inputting infected

number.of.sims = 4
```

## Constant Temperature Treatment

### 10ºC

```{r}
x10 <- c(S = 169, I = 0, D = 0, E = 0)
constant <- NULL
set.seed <- (1234)
for(i in 1:number.of.sims){
  constant[[i]] <- ssa(x10, a, nu, parms10, tf = 120, method = "ETL",
                       simName = "SIDE Model",
                       verbose = FALSE,
                       consoleInterval = Inf,
                       tau = 0.001)
}
```

#### Quick Plotting

```{r}
constant_1 <- as.data.frame(constant[[1]][["data"]]); max <- max(constant_1[ , 2:4])
par(mar = c(5, 4, 4, 9), xpd = TRUE)
plot(constant_1$t, constant_1$S, type = "l", ylim = c(0, max), 
     main = "Constant Treatment (10C): Population 1", xlab = "Day", ylab = "Abundance",
     col = "red3")
lines(constant_1$t, constant_1$I, col = "orange1")
lines(constant_1$t, constant_1$D, col = "orchid")
legend("topright", inset = c(-0.4, 0), legend = c("Susceptible", "Infected", "Dead"), 
       col = c("red3", "orange1", "orchid"), lty = 1, lwd = 2)

constant_2 <- as.data.frame(constant[[2]][["data"]]); max <- max(constant_2[ , 2:4])
par(mar = c(5, 4, 4, 9), xpd = TRUE)
plot(constant_2$t, constant_2$S, type = "l", ylim = c(0, max), 
     main = "Constant Treatment (10C): Population 2", xlab = "Day", ylab = "Abundance",
     col = "red3")
lines(constant_2$t, constant_2$I, col = "orange1")
lines(constant_2$t, constant_2$D, col = "orchid")
legend("topright", inset = c(-0.4, 0), legend = c("Susceptible", "Infected", "Dead"), 
       col = c("red3", "orange1", "orchid"), lty = 1, lwd = 2)

constant_3 <- as.data.frame(constant[[3]][["data"]]); max <- max(constant_3[ , 2:4])
par(mar = c(5, 4, 4, 9), xpd = TRUE)
plot(constant_3$t, constant_3$S, type = "l", ylim = c(0, max), 
     main = "Constant Treatment (10C): Population 3", xlab = "Day", ylab = "Abundance",
     col = "red3")
lines(constant_3$t, constant_3$I, col = "orange1")
lines(constant_3$t, constant_3$D, col = "orchid")
legend("topright", inset = c(-0.4, 0), legend = c("Susceptible", "Infected", "Dead"), 
       col = c("red3", "orange1", "orchid"), lty = 1, lwd = 2)

constant_4 <- as.data.frame(constant[[4]][["data"]]); max <- max(constant_4[ , 2:4])
par(mar = c(5, 4, 4, 9), xpd = TRUE)
plot(constant_4$t, constant_4$S, type = "l", ylim = c(0, max), 
     main = "Constant Treatment (10C): Population 4", xlab = "Day", ylab = "Abundance",
     col = "red3")
lines(constant_4$t, constant_4$I, col = "orange1")
lines(constant_4$t, constant_4$D, col = "orchid")
legend("topright", inset = c(-0.4, 0), legend = c("Susceptible", "Infected", "Dead"), 
       col = c("red3", "orange1", "orchid"), lty = 1, lwd = 2)
```

## Warming Temperature Treatment

Making empty vectors for outputs:

```{r}
out10 <- NULL
out105 <- NULL
out11 <- NULL
out115 <- NULL
out12 <- NULL
out125 <- NULL
out13 <- NULL
out135 <- NULL

x10 <- NULL
x105 <- NULL
x11 <- NULL
x115 <- NULL
x12 <- NULL
x125 <- NULL
x13 <- NULL
x135 <- NULL
```

### 10.0ºC

```{r}
x10 <- c(S = 169, I = 0, D = 0, E = 0)
set.seed <- (1234)
for(i in 1:number.of.sims){
out10[[i]] <- ssa(x10, a, nu, parms10, tf = 15, method = "ETL",
                   simName = "SIDE Model",
                   verbose = FALSE,
                   consoleInterval = Inf,
                   tau  =0.001)
}
```

### 10.5ºC

```{r}
for(i in 1:number.of.sims){
  x105[[i]] <- c(out10[[i]]$data[15003, 2],
                 out10[[i]]$data[15003, 3],
                 out10[[i]]$data[15003, 4],
                 out10[[i]]$data[15003, 5])
}

set.seed <- (1234)
for(i in 1:number.of.sims){
  out105[[i]] <- ssa(x105[[i]], a, nu, parms105, tf = 15, method = "ETL",
                    simName = "SIDE Model",
                    verbose = FALSE,
                    consoleInterval = Inf,
                    tau = 0.001)
}
```

### 11.0ºC

```{r}
for(i in 1:number.of.sims){
  x11[[i]] <- c(out105[[i]]$data[15003, 2],
                 out105[[i]]$data[15003, 3],
                 out105[[i]]$data[15003, 4],
                 out105[[i]]$data[15003, 5])
}

set.seed <- (1234)
for(i in 1:number.of.sims){
  out11[[i]]<- ssa(x11[[i]], a, nu, parms11, tf = 15, method = "ETL",
                   simName = "SIDE Model",
                   verbose = FALSE,
                   consoleInterval = Inf,
                   tau = 0.001)
}
```

### 11.5ºC

```{r}
for(i in 1:number.of.sims){
  x115[[i]] <- c(out11[[i]]$data[15003, 2],
                 out11[[i]]$data[15003, 3],
                 out11[[i]]$data[15003, 4],
                 out11[[i]]$data[15003, 5])
}

set.seed <- (1234)
for(i in 1:number.of.sims){
  out115[[i]]<- ssa(x115[[i]], a, nu, parms115, tf = 15, method = "ETL",
                   simName = "SIDE Model",
                   verbose = FALSE,
                   consoleInterval = Inf,
                   tau = 0.001)
}
```

### 12.0ºC

```{r}
for(i in 1:number.of.sims){
  x12[[i]] <- c(out115[[i]]$data[15003, 2],
                 out115[[i]]$data[15003, 3],
                 out115[[i]]$data[15003, 4],
                 out115[[i]]$data[15003, 5])
}

set.seed <- (1234)
for(i in 1:number.of.sims){
  out12[[i]]<- ssa(x12[[i]], a, nu, parms12, tf = 15, method = "ETL",
                   simName = "SIDE Model",
                   verbose = FALSE,
                   consoleInterval = Inf,
                   tau = 0.001)
}
```

### 12.5ºC

```{r}
for(i in 1:number.of.sims){
  x125[[i]] <- c(out12[[i]]$data[15003, 2],
                 out12[[i]]$data[15003, 3],
                 out12[[i]]$data[15003, 4],
                 out12[[i]]$data[15003, 5])
}

set.seed <- (1234)
for(i in 1:number.of.sims){
  out125[[i]]<- ssa(x125[[i]], a, nu, parms125, tf = 15, method = "ETL",
                   simName = "SIDE Model",
                   verbose = FALSE,
                   consoleInterval = Inf,
                   tau = 0.001)
}
```

### 13.0ºC

```{r}
for(i in 1:number.of.sims){
  x13[[i]] <- c(out125[[i]]$data[15003, 2],
                 out125[[i]]$data[15003, 3],
                 out125[[i]]$data[15003, 4],
                 out125[[i]]$data[15003, 5])
}

set.seed <- (1234)
for(i in 1:number.of.sims){
  out13[[i]]<- ssa(x13[[i]], a, nu, parms13, tf = 15, method = "ETL",
                   simName = "SIDE Model",
                   verbose = FALSE,
                   consoleInterval = Inf,
                   tau = 0.001)
}
```

### 13.5ºC

```{r}
for(i in 1:number.of.sims){
  x135[[i]] <- c(out13[[i]]$data[15003, 2],
                 out13[[i]]$data[15003, 3],
                 out13[[i]]$data[15003, 4],
                 out13[[i]]$data[15003, 5])
}

set.seed <- (1234)
for(i in 1:number.of.sims){
  out135[[i]]<- ssa(x135[[i]], a, nu, parms135, tf = 15, method = "ETL",
                   simName = "SIDE Model",
                   verbose = FALSE,
                   consoleInterval = Inf,
                   tau = 0.001)
}
```

### Organize Outputs

```{r}
warming <- NULL

for(i in 1:number.of.sims){
  warming[[i]] <- rbind(out10[[i]]$data,
                        out105[[i]]$data,
                        out11[[i]]$data,
                        out115[[i]]$data,
                        out12[[i]]$data,
                        out125[[i]]$data,
                        out13[[i]]$data,
                        out135[[i]]$data)
  
  warming[[i]][, 1] <- seq(0.000, 120.023, 0.001)
  colnames(warming[[i]]) <- c("Time", "S", "I", "D", "E")
}
```

#### Quick Plotting

```{r}
warming_1 <- as.data.frame(warming[[1]]); max <- max(warming_1[ , 2:4])
par(mar = c(5, 4, 4, 9), xpd = TRUE)
plot(warming_1$Time, warming_1$S, type = "l", ylim = c(0, max), 
     main = "Warming Treatment: Population 1", xlab = "Day", ylab = "Abundance",
     col = "red3")
lines(warming_1$Time, warming_1$I, col = "orange1")
lines(warming_1$Time, warming_1$D, col = "orchid")
legend("topright", inset = c(-0.4, 0), legend = c("Susceptible", "Infected", "Dead"), 
       col = c("red3", "orange1", "orchid"), lty = 1, lwd = 2)

warming_2 <- as.data.frame(warming[[2]]); max <- max(warming_2[ , 2:4])
par(mar = c(5, 4, 4, 9), xpd = TRUE)
plot(warming_2$Time, warming_2$S, type = "l", ylim = c(0, max), 
     main = "Warming Treatment: Population 2", xlab = "Day", ylab = "Abundance",
     col = "red3")
lines(warming_2$Time, warming_2$I, col = "orange1")
lines(warming_2$Time, warming_2$D, col = "orchid")
legend("topright", inset = c(-0.4, 0), legend = c("Susceptible", "Infected", "Dead"), 
       col = c("red3", "orange1", "orchid"), lty = 1, lwd = 2)

warming_3 <- as.data.frame(warming[[3]]); max <- max(warming_3[ , 2:4])
par(mar = c(5, 4, 4, 9), xpd = TRUE)
plot(warming_3$Time, warming_3$S, type = "l", ylim = c(0, max), 
     main = "Warming Treatment: Population 3", xlab = "Day", ylab = "Abundance",
     col = "red3")
lines(warming_3$Time, warming_3$I, col = "orange1")
lines(warming_3$Time, warming_3$D, col = "orchid")
legend("topright", inset = c(-0.4, 0), legend = c("Susceptible", "Infected", "Dead"), 
       col = c("red3", "orange1", "orchid"), lty = 1, lwd = 2)

warming_4 <- as.data.frame(warming[[4]]); max <- max(warming_4[ , 2:4])
par(mar = c(5, 4, 4, 9), xpd = TRUE)
plot(warming_4$Time, warming_4$S, type = "l", ylim = c(0, max), 
     main = "Warming Treatment: Population 4", xlab = "Day", ylab = "Abundance",
     col = "red3")
lines(warming_4$Time, warming_4$I, col = "orange1")
lines(warming_4$Time, warming_4$D, col = "orchid")
legend("topright", inset = c(-0.4, 0), legend = c("Susceptible", "Infected", "Dead"), 
       col = c("red3", "orange1", "orchid"), lty = 1, lwd = 2)
```

## Plotting

```{r}
par(mar = c(5, 5, 3, 2))
plot(constant[[1]]$data[, 1], (constant[[1]]$data[, 3]/(constant[[1]]$data[, 2] + constant[[1]]$data[, 3])), 
     type = 'n', xlim = c(0, 120), ylim = c(0, 0.5), 
     main = NA, xlab = "Days", ylab = "Prevalence")

for(i in 1:length(constant)){
  lines(constant[[i]]$data[, 1], (constant[[i]]$data[, 3]/(constant[[i]]$data[, 2] + constant[[i]]$data[, 3])), 
        col = "cornflowerblue", lwd = 2)
}

for(i in 1:length(warming)){
  lines(warming[[i]][, 1], (warming[[i]][, 3]/(warming[[i]][, 2] + warming[[i]][, 3])), 
        col = "red3", lwd = 2)
}

legend("topright", legend = c("Constant Temperature Treatment", "Warming Temperature Treatment"),
       col = c("cornflowerblue", "red3"), lwd = 2)
```

# Computing 95% CIs

```{r}
constant.prevalences <- matrix(nrow = 120002, ncol = 4)
warming.prevalences <- matrix(nrow = 120024, ncol = 4)

for(i in 1:4){
  constant.prevalences[,i] <- as.numeric(constant[[i]]$data[, 3])/(as.numeric(constant[[i]]$data[, 3]) + as.numeric(constant[[i]]$data[, 2]))
}
for(i in 1:4){
  warming.prevalences[,i] <- warming[[i]][, 3]/(warming[[i]][, 3]+warming[[i]][, 2])
}

# Get medians and confidence intervals for MTE predictions #
median.constant.prev <- c()
lower.constant.prev <- c()
upper.constant.prev <- c()
for(i in 1:dim(constant.prevalences)[[1]]){
  median.constant.prev[i] <- median(constant.prevalences[i, ])
}
for(i in 1:dim(constant.prevalences)[[1]]){
  lower.constant.prev[i] <- quantile(constant.prevalences[i, ], probs = 0.025)
}
for(i in 1:dim(constant.prevalences)[[1]]){
  upper.constant.prev[i] <- quantile(constant.prevalences[i, ], probs = 0.975)
}

median.warming.prev <- c()
lower.warming.prev <- c()
upper.warming.prev <- c()
for(i in 1:dim(warming.prevalences)[[1]]){
  median.warming.prev[i] <- median(warming.prevalences[i, ])
}
for(i in 1:dim(warming.prevalences)[[1]]){
  lower.warming.prev[i] <- quantile(warming.prevalences[i, ], probs = 0.025)
}
for(i in 1:dim(warming.prevalences)[[1]]){
  upper.warming.prev[i] <- quantile(warming.prevalences[i, ], probs = 0.975)
}
```

# Plotting Simulated and Empirical Data

```{r}
data <- read.csv("Kirk_Data/Epidemics_Data_Dryad/Kirk_et_al_epidemics_sampling_data.csv")
data.sum <- summaryBy(status ~ population + treatment + day, data = data, FUN = sum)

sub1 <- subset(data.sum, population == 1)
sub2 <- subset(data.sum, population == 2)
sub3 <- subset(data.sum, population == 3)
sub4 <- subset(data.sum, population == 4)
sub5 <- subset(data.sum, population == 5)
sub6 <- subset(data.sum, population == 6)
sub7 <- subset(data.sum, population == 7)
sub8 <- subset(data.sum, population == 8)

constant.time.axis <- c(seq(0.001, 120.002, 0.001))
warming.time.axis <- c(seq(0.001, 120.024, 0.001))
```

Plot the experimental populations:

```{r}
par(mfrow = c(2, 4))
par(mar = c(1, 1, 1, 1))
par(oma = c(6, 6, 0, 0))

plot(sub1$status.sum/12 ~ sub1$day, ylim = c(0, 1), xlim = c(0, 120), col = "cornflowerblue", type = "l", lty = 1,
     xlab = NA, ylab = NA, lwd = 4, pch = 19)
text(x = max(sub1$day)/2, y = 0.9, labels = "10C")
lines(constant.time.axis, lower.constant.prev, lwd = 2)
lines(constant.time.axis, upper.constant.prev, lwd = 2)
lines(constant.time.axis, median.constant.prev, lwd = 3)

plot(sub2$status.sum/12 ~ sub2$day, ylim = c(0, 1), xlim = c(0, 120),col = "cornflowerblue", type = "l", lty = 1,
     xlab = NA, ylab = NA, lwd = 4, pch = 19)
text(x = max(sub1$day)/2, y = 0.9, labels = "10C")
lines(constant.time.axis, lower.constant.prev, lwd = 2)
lines(constant.time.axis, upper.constant.prev, lwd = 2)
lines(constant.time.axis, median.constant.prev, lwd = 3)

plot(sub3$status.sum/12 ~ sub3$day, ylim = c(0, 1), xlim = c(0, 120),col = "cornflowerblue", type = "l", lty = 1,
     xlab = NA, ylab = NA, lwd = 4, pch = 19)
text(x = max(sub1$day)/2, y = 0.9, labels = "10C")
lines(constant.time.axis, lower.constant.prev, lwd = 2)
lines(constant.time.axis, upper.constant.prev, lwd = 2)
lines(constant.time.axis, median.constant.prev, lwd = 3)

plot(sub4$status.sum/12 ~ sub4$day, ylim = c(0, 1), xlim = c(0, 120),col = "cornflowerblue", type = "l", lty = 1,
     xlab = NA, ylab = NA, lwd = 4, pch = 19)
text(x = max(sub1$day)/2, y = 0.9, labels = "10C")
lines(constant.time.axis, lower.constant.prev, lwd = 2)
lines(constant.time.axis, upper.constant.prev, lwd = 2)
lines(constant.time.axis, median.constant.prev, lwd = 3)

plot(sub5$status.sum/12 ~ sub5$day, ylim = c(0, 1), xlim = c(0, 120),col = "red3", type = "l", lty = 1,
     xlab = NA, ylab = NA, lwd = 4, pch = 19)
text(x = c(7.5,22.5,37.5, 52.5, 67.5, 82.5, 97.5, 112.5), y = 0.9, labels = c("10","10.5","11","11.5","12","12.5","13","13.5"))
lines(warming.time.axis, lower.warming.prev, lwd = 2)
lines(warming.time.axis, upper.warming.prev, lwd = 2)
lines(warming.time.axis, median.warming.prev, lwd = 3)
abline(v = c(15, 30, 45, 60, 75, 90, 105), lty = 3)

plot(sub6$status.sum/12 ~ sub6$day, ylim = c(0, 1), xlim = c(0, 120),col = "red3", type = "l", lty = 1,
     xlab = NA, ylab = NA, lwd = 4, pch = 19)
text(x  =  c(7.5,22.5,37.5, 52.5, 67.5, 82.5, 97.5, 112.5), y = 0.9, labels = c("10","10.5","11","11.5","12","12.5","13","13.5"))
lines(warming.time.axis, lower.warming.prev, lwd = 2)
lines(warming.time.axis, upper.warming.prev, lwd = 2)
lines(warming.time.axis, median.warming.prev, lwd = 3)
abline(v = c(15, 30, 45, 60, 75, 90, 105), lty = 3)

plot(sub7$status.sum/12 ~ sub7$day, ylim = c(0, 1), xlim = c(0, 120),col = "red3", type = "l", lty = 1,
     xlab = NA, ylab = NA, lwd = 4, pch = 19)
text(x  =  c(7.5,22.5,37.5, 52.5, 67.5, 82.5, 97.5, 112.5), y = 0.9, labels = c("10","10.5","11","11.5","12","12.5","13","13.5"))
lines(warming.time.axis, lower.warming.prev, lwd = 2)
lines(warming.time.axis, upper.warming.prev, lwd = 2)
lines(warming.time.axis, median.warming.prev, lwd = 3)
abline(v = c(15, 30, 45, 60, 75, 90, 105), lty = 3)

plot(sub8$status.sum/12 ~ sub8$day, ylim = c(0, 1), xlim = c(0, 120),col = "red3", type = "l", lty = 1,
     xlab = NA, ylab = NA, lwd = 4, pch = 19)
text(x  =  c(7.5,22.5,37.5, 52.5, 67.5, 82.5, 97.5, 112.5), y = 0.9, labels = c("10","10.5","11","11.5","12","12.5","13","13.5"))
lines(warming.time.axis, lower.warming.prev, lwd = 2)
lines(warming.time.axis, upper.warming.prev, lwd = 2)
lines(warming.time.axis, median.warming.prev, lwd = 3)
abline(v = c(15, 30, 45, 60, 75, 90, 105), lty = 3)

mtext("Prevalence", side = 2, line = 2.5, outer = TRUE)
mtext("Day", side = 1, line = 3.5, outer = TRUE)
```

# Saving Plots

```{r}
## Stochastic model, constant temperature treatment
constant_1 <- as.data.frame(constant[[1]][["data"]]); max <- max(constant_1[ , 2:4])
png("Plots/StochMod_ConstTemp_1.png", width = 800, height = 600)
par(mar = c(5, 4, 4, 9), xpd = TRUE)
plot(constant_1$t, constant_1$S, type = "l", ylim = c(0, max), 
     main = "Constant Treatment (10C): Population 1", xlab = "Day", ylab = "Abundance",
     col = "red3")
lines(constant_1$t, constant_1$I, col = "orange1")
lines(constant_1$t, constant_1$D, col = "orchid")
legend("topright", inset = c(-0.4, 0), legend = c("Susceptible", "Infected", "Dead"), 
       col = c("red3", "orange1", "orchid"), lty = 1, lwd = 2)
dev.off()

constant_2 <- as.data.frame(constant[[2]][["data"]]); max <- max(constant_2[ , 2:4])
png("Plots/StochMod_ConstTemp_2.png", width = 800, height = 600)
par(mar = c(5, 4, 4, 9), xpd = TRUE)
plot(constant_2$t, constant_2$S, type = "l", ylim = c(0, max), 
     main = "Constant Treatment (10C): Population 2", xlab = "Day", ylab = "Abundance",
     col = "red3")
lines(constant_2$t, constant_2$I, col = "orange1")
lines(constant_2$t, constant_2$D, col = "orchid")
legend("topright", inset = c(-0.4, 0), legend = c("Susceptible", "Infected", "Dead"), 
       col = c("red3", "orange1", "orchid"), lty = 1, lwd = 2)
dev.off()

constant_3 <- as.data.frame(constant[[3]][["data"]]); max <- max(constant_3[ , 2:4])
png("Plots/StochMod_ConstTemp_3.png", width = 800, height = 600)
par(mar = c(5, 4, 4, 9), xpd = TRUE)
plot(constant_3$t, constant_3$S, type = "l", ylim = c(0, max), 
     main = "Constant Treatment (10C): Population 3", xlab = "Day", ylab = "Abundance",
     col = "red3")
lines(constant_3$t, constant_3$I, col = "orange1")
lines(constant_3$t, constant_3$D, col = "orchid")
legend("topright", inset = c(-0.4, 0), legend = c("Susceptible", "Infected", "Dead"), 
       col = c("red3", "orange1", "orchid"), lty = 1, lwd = 2)
dev.off()

constant_4 <- as.data.frame(constant[[4]][["data"]]); max <- max(constant_4[ , 2:4])
png("Plots/StochMod_ConstTemp_4.png", width = 800, height = 600)
par(mar = c(5, 4, 4, 9), xpd = TRUE)
plot(constant_4$t, constant_4$S, type = "l", ylim = c(0, max), 
     main = "Constant Treatment (10C): Population 4", xlab = "Day", ylab = "Abundance",
     col = "red3")
lines(constant_4$t, constant_4$I, col = "orange1")
lines(constant_4$t, constant_4$D, col = "orchid")
legend("topright", inset = c(-0.4, 0), legend = c("Susceptible", "Infected", "Dead"), 
       col = c("red3", "orange1", "orchid"), lty = 1, lwd = 2)
dev.off()

## Stochastic model, warming temperature treatment
warming_1 <- as.data.frame(warming[[1]]); max <- max(warming_1[ , 2:4])
png("Plots/StochMod_WarmTemp_1.png", width = 800, height = 600)
par(mar = c(5, 4, 4, 9), xpd = TRUE)
plot(warming_1$Time, warming_1$S, type = "l", ylim = c(0, max), 
     main = "Warming Treatment: Population 1", xlab = "Day", ylab = "Abundance",
     col = "red3")
lines(warming_1$Time, warming_1$I, col = "orange1")
lines(warming_1$Time, warming_1$D, col = "orchid")
legend("topright", inset = c(-0.4, 0), legend = c("Susceptible", "Infected", "Dead"), 
       col = c("red3", "orange1", "orchid"), lty = 1, lwd = 2)
dev.off()

warming_2 <- as.data.frame(warming[[2]]); max <- max(warming_2[ , 2:4])
png("Plots/StochMod_WarmTemp_2.png", width = 800, height = 600)
par(mar = c(5, 4, 4, 9), xpd = TRUE)
plot(warming_2$Time, warming_2$S, type = "l", ylim = c(0, max), 
     main = "Warming Treatment: Population 2", xlab = "Day", ylab = "Abundance",
     col = "red3")
lines(warming_2$Time, warming_2$I, col = "orange1")
lines(warming_2$Time, warming_2$D, col = "orchid")
legend("topright", inset = c(-0.4, 0), legend = c("Susceptible", "Infected", "Dead"), 
       col = c("red3", "orange1", "orchid"), lty = 1, lwd = 2)
dev.off()

warming_3 <- as.data.frame(warming[[3]]); max <- max(warming_3[ , 2:4])
png("Plots/StochMod_WarmTemp_3.png", width = 800, height = 600)
par(mar = c(5, 4, 4, 9), xpd = TRUE)
plot(warming_3$Time, warming_3$S, type = "l", ylim = c(0, max), 
     main = "Warming Treatment: Population 3", xlab = "Day", ylab = "Abundance",
     col = "red3")
lines(warming_3$Time, warming_3$I, col = "orange1")
lines(warming_3$Time, warming_3$D, col = "orchid")
legend("topright", inset = c(-0.4, 0), legend = c("Susceptible", "Infected", "Dead"), 
       col = c("red3", "orange1", "orchid"), lty = 1, lwd = 2)
dev.off()

warming_4 <- as.data.frame(warming[[4]]); max <- max(warming_4[ , 2:4])
png("Plots/StochMod_WarmTemp_4.png", width = 800, height = 600)
par(mar = c(5, 4, 4, 9), xpd = TRUE)
plot(warming_4$Time, warming_4$S, type = "l", ylim = c(0, max), 
     main = "Warming Treatment: Population 4", xlab = "Day", ylab = "Abundance",
     col = "red3")
lines(warming_4$Time, warming_4$I, col = "orange1")
lines(warming_4$Time, warming_4$D, col = "orchid")
legend("topright", inset = c(-0.4, 0), legend = c("Susceptible", "Infected", "Dead"), 
       col = c("red3", "orange1", "orchid"), lty = 1, lwd = 2)
dev.off()

## Stochastic model, constant and warming treatments
png("Plots/StochMod_CompTreatments.png", width = 800, height = 600)
par(mar = c(5, 5, 3, 2))
plot(constant[[1]]$data[, 1], (constant[[1]]$data[, 3]/(constant[[1]]$data[, 2] + constant[[1]]$data[, 3])), 
     type = 'n', xlim = c(0, 120), ylim = c(0, 0.5), 
     main = NA, xlab = "Days", ylab = "Prevalence")
for(i in 1:length(constant)){
  lines(constant[[i]]$data[, 1], (constant[[i]]$data[, 3]/(constant[[i]]$data[, 2] + constant[[i]]$data[, 3])), 
        col = "cornflowerblue", lwd = 2)
}
for(i in 1:length(warming)){
  lines(warming[[i]][, 1], (warming[[i]][, 3]/(warming[[i]][, 2] + warming[[i]][, 3])), 
        col = "red3", lwd = 2)
}
legend("topright", legend = c("Constant Temperature Treatment", "Warming Temperature Treatment"),
       col = c("cornflowerblue", "red3"), lwd = 2)
dev.off()

## Stochastic model and experimental data
png("Plots/StochMod_ExpData.png", width = 1600, height = 1200)
par(mfrow = c(2, 4))
par(mar = c(1, 1, 1, 1))
par(oma = c(6, 6, 0, 0))
plot(sub1$status.sum/12 ~ sub1$day, ylim = c(0, 1), xlim = c(0, 120), col = "cornflowerblue", type = "l", lty = 1,
     xlab = NA, ylab = NA, lwd = 4, pch = 19)
text(x = max(sub1$day)/2, y = 0.9, labels = "10C")
lines(constant.time.axis, lower.constant.prev, lwd = 2)
lines(constant.time.axis, upper.constant.prev, lwd = 2)
lines(constant.time.axis, median.constant.prev, lwd = 3)
plot(sub2$status.sum/12 ~ sub2$day, ylim = c(0, 1), xlim = c(0, 120),col = "cornflowerblue", type = "l", lty = 1,
     xlab = NA, ylab = NA, lwd = 4, pch = 19)
text(x = max(sub1$day)/2, y = 0.9, labels = "10C")
lines(constant.time.axis, lower.constant.prev, lwd = 2)
lines(constant.time.axis, upper.constant.prev, lwd = 2)
lines(constant.time.axis, median.constant.prev, lwd = 3)
plot(sub3$status.sum/12 ~ sub3$day, ylim = c(0, 1), xlim = c(0, 120),col = "cornflowerblue", type = "l", lty = 1,
     xlab = NA, ylab = NA, lwd = 4, pch = 19)
text(x = max(sub1$day)/2, y = 0.9, labels = "10C")
lines(constant.time.axis, lower.constant.prev, lwd = 2)
lines(constant.time.axis, upper.constant.prev, lwd = 2)
lines(constant.time.axis, median.constant.prev, lwd = 3)
plot(sub4$status.sum/12 ~ sub4$day, ylim = c(0, 1), xlim = c(0, 120),col = "cornflowerblue", type = "l", lty = 1,
     xlab = NA, ylab = NA, lwd = 4, pch = 19)
text(x = max(sub1$day)/2, y = 0.9, labels = "10C")
lines(constant.time.axis, lower.constant.prev, lwd = 2)
lines(constant.time.axis, upper.constant.prev, lwd = 2)
lines(constant.time.axis, median.constant.prev, lwd = 3)
plot(sub5$status.sum/12 ~ sub5$day, ylim = c(0, 1), xlim = c(0, 120),col = "red3", type = "l", lty = 1,
     xlab = NA, ylab = NA, lwd = 4, pch = 19)
text(x = c(7.5,22.5,37.5, 52.5, 67.5, 82.5, 97.5, 112.5), y = 0.9, labels = c("10","10.5","11","11.5","12","12.5","13","13.5"))
lines(warming.time.axis, lower.warming.prev, lwd = 2)
lines(warming.time.axis, upper.warming.prev, lwd = 2)
lines(warming.time.axis, median.warming.prev, lwd = 3)
abline(v = c(15, 30, 45, 60, 75, 90, 105), lty = 3)
plot(sub6$status.sum/12 ~ sub6$day, ylim = c(0, 1), xlim = c(0, 120),col = "red3", type = "l", lty = 1,
     xlab = NA, ylab = NA, lwd = 4, pch = 19)
text(x  =  c(7.5,22.5,37.5, 52.5, 67.5, 82.5, 97.5, 112.5), y = 0.9, labels = c("10","10.5","11","11.5","12","12.5","13","13.5"))
lines(warming.time.axis, lower.warming.prev, lwd = 2)
lines(warming.time.axis, upper.warming.prev, lwd = 2)
lines(warming.time.axis, median.warming.prev, lwd = 3)
abline(v = c(15, 30, 45, 60, 75, 90, 105), lty = 3)
plot(sub7$status.sum/12 ~ sub7$day, ylim = c(0, 1), xlim = c(0, 120),col = "red3", type = "l", lty = 1,
     xlab = NA, ylab = NA, lwd = 4, pch = 19)
text(x  =  c(7.5,22.5,37.5, 52.5, 67.5, 82.5, 97.5, 112.5), y = 0.9, labels = c("10","10.5","11","11.5","12","12.5","13","13.5"))
lines(warming.time.axis, lower.warming.prev, lwd = 2)
lines(warming.time.axis, upper.warming.prev, lwd = 2)
lines(warming.time.axis, median.warming.prev, lwd = 3)
abline(v = c(15, 30, 45, 60, 75, 90, 105), lty = 3)
plot(sub8$status.sum/12 ~ sub8$day, ylim = c(0, 1), xlim = c(0, 120),col = "red3", type = "l", lty = 1,
     xlab = NA, ylab = NA, lwd = 4, pch = 19)
text(x  =  c(7.5,22.5,37.5, 52.5, 67.5, 82.5, 97.5, 112.5), y = 0.9, labels = c("10","10.5","11","11.5","12","12.5","13","13.5"))
lines(warming.time.axis, lower.warming.prev, lwd = 2)
lines(warming.time.axis, upper.warming.prev, lwd = 2)
lines(warming.time.axis, median.warming.prev, lwd = 3)
abline(v = c(15, 30, 45, 60, 75, 90, 105), lty = 3)
mtext("Prevalence", side = 2, line = 2.5, outer = TRUE)
mtext("Day", side = 1, line = 3.5, outer = TRUE)
dev.off()
```
