==> 5_Detrending.Rmd <==
# Detrending Data

Lots of people detrend simulated and empirical data before looking for EWS to remove long-term trends unrelated to the system's response to perturbation.

# From the Literature

As a quick summary, here a few works that have discussed/performed pre-processing data by detrending:

## Dakos et al. 2008

-   They subtracted a **Gaussian kernel smoothing function** from the data and used the remaining residuals for the estimation of the autoregressive coefficient at lag 1

## Lenton 2011

-   Used detrended fluctuation analysis (DFA) to identify correlations over long time scales

-   Used **Gaussian filtering** to detrend the data

-   Warns that when using Gaussian filtering, filtering bandwidth and sliding-window length should be carefully chosen to remove any long-term trends whilst retaining the fluctuations pertinent to diagnosing slowing down

## Dakos et al. 2012

-   For all metrics that were estimated within rolling windows, they removed trends or filtered out high frequencies using **Gaussian smoothing** (autocorrelation, variance, skewness), **simple linear detrending** (DFA), or by **fitting linear autoregressive models** (conditional heteroskedasticity)
-   When applying these or any other type of detrending or filtering (i.e. **first-differences, removing running means, loess smoothing**), care should be taken to not over-fit or filter out the slow dynamics (of interest) from the dataset
-   Alternatively, one could also detrend within the rolling windows rather than the entire dataset

## O'Regan and Drake 2013

-   Used **Van Kampen detrending** and **Gaussian filtering**

-   They found no qualitative difference, but did find that Van Kampen detrending made it easier to detect EWS

## Gsell et al. 2016

-   Used a **Gaussian smoother**

## Miller et al. 2017

-   "For our model, estimating and removing a periodic trend prior to EWS analysis did not improve prediction uniformly among statistics. This was not entirely surprising because the **seasonal signal was not apparent in many time series... Therefore, periodic detrending and differencing introduced artificial patterns in the time series.**"

## Pananos et al. 2017

-   They smoothed the time series for both the stochastic model and the empirical data using a **Nadarya Watson estimator with Gaussian kernel at a bandwidth of 10%, selected based on Silverman's rule of thumb, and then subtracted the smoothed time series from the raw time series to generate a detrended (residual) time series.**

## Dessavre et al. 2019

-   Compares methods, mainly focusing on **simulation detrending (over multiple realisations of the process) and windowed detrending (applied to empirical data)**

-   Also mentions: **Gaussian detrending, windowed linear regression, windowed quadratic regression, and wavelets**

-   They propose a **spatial detrending approach** that could me applied to meta-population data

    -   This approach removes the mean over multiple populations to obtain the fluctuations, and unlike Gaussian detrending, does not require any hyperparamaters to be inputted; however, it does **assume spatial ergodicity**, i.e. all subpopulations are similar

## Harris et al. 2020

-   Used a **Gaussian kernel**

## Bury et al. 2021

-   Data were detrended using **Lowess smoothing** with a span of 0.2 and degree 1

-   They used the same data preprocessing as Dakos et al. 2012, which involved using **linear interpolation to make the data equidistant**, and **detrending with a Gaussian kernel smoothing function**. Bandwidth of the kernel was specified for each time series to remove long-term trends while not overfitting

## Southall et al. 2021

-   "In simulation studies, [detrending] can be done by removing the average over replicate realizations, and it has been shown that stochastic simulations produced using the Gillespie algorithm match the theoretical predictions of EWSs"

-   "In practice, without the availability of true replicates, **Gaussian detrending** is often implemented. **Gaussian detrending is a moving average technique, which removes a weighted mean over a selected window size, where the weights are taken from a Gaussian kernel with a user-inputted standard deviation.**"

## Dablander et al. 2022

-   Used R package, *spaero*, which detrends with a **Gaussian kernel**

## Delecroix et al. 2022 (medrxiv)

-   Essentially cites Miller et al. 2017 and Dessavre et al. 2019 to say that **detrending is very important when dynamics are subject to seasonal forcing**

## Summary:

-   Gaussian smoothing/filtering/detrending seems to be the most popular approach

-   However, per Southall et al. 2021: "This method not only **requires the user to select a suitable choice of window size and standard deviation**, but also **makes the assumption that the data are ergodic**. This raises a key challenge with this technique, as **ergodicity only holds for stationary time series; however, these methods will be implemented on data that are believed to be approaching a critical transition**. For this reason, the choice of the window size and the speed a disease is approaching a critical transition are interlinked when deciding if the assumptions of Gaussian detrending are appropriate."

    -   Ergodicity: a stochastic process is said to be in an ergodic regime if an observable's ensemble average equals the time average. In this regime, **any collection of random samples from a process must represent the average statistical properties of the entire regime**.

-   How to actually do it:

    -   May find more specific methods in Dakos et al. 2012, Pananos et al. 2017

    -   Could also take a look at the meta-population approach in Dessavre et al. 2019...

    -   From [Towards Data Science](https://towardsdatascience.com/gaussian-smoothing-in-time-series-data-c6801f8a4dc3):

        -   Kernels define the shape of the function used to take the average of the neighboring points. The Gaussian kernel has the shape of the Gaussian curve. As seen in the Gaussian curve, the near points (around 0 in the above curve) will be weighted higher and the farther points will be weighted lower.

        -   Expression for the Gaussian kernel: $K(x^*,x_i)=exp(-\frac{(x^*-x_i)^2}{2b^2})$, where $b$ defines the width of the kernel (or, more simply, the number of data points included in the window; as such, smaller widths risk maintining the noise, and larger widths risk removing the signal)

        -   For doing the smoothing, we proceed data point by point. For each point, we calculate the kernel function values and calculate the weighted average of data points weighted by kernel function values.

        -   We standardize the kernel values so that we do not scale up the new number of cases. We do it by dividing the Gaussian kernel values by sum of all the Gaussian kernel values. Then, we do element-wise multiplication of *new cases* column with *Gaussian kernel values* column and sum them to get the *smoothed* number of cases.

# Troubleshooting

## Using R Package: `smoother`

Note: You can also implement Gaussian filtering using `library(spaero)` or `library(KRLS)`.

### COVID Data

```{r}
library(smoother)
# smth.gaussian(x, window, alpha)
## Where:
### x: data
### window: the length of the smoothing window, if an integer, represents number of items, else, if a value between 0 and 1, represents the proportion of the input vector
### alpha: parameter to determine the breadth of the gaussian window, yielding more or less sensitive smoothing characteristics

data <- read.csv("https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/owid-covid-data.csv")
npl <- subset(data, iso_code == "NPL")
smooth.ts <- smth.gaussian(npl$new_cases)
days <- seq(1, nrow(npl), 1)
plot(days, npl$new_cases, type = "l", main = "Default 'window' and 'alpha' Values", xlab = "Day", ylab = "New Cases", col = "grey", lwd = 1)
lines(days, smooth.ts, col = "orchid", lwd = 2)
lines(days, npl$new_cases_smoothed, col = "orange", lwd = 2)
legend("topright", legend = c("Raw Data", "Auto-Smoothed", "Given Smoothed"),
       col = c("grey", "orchid", "orange"), lwd = 2)
# With adjusted window and alpha values
smooth.ts <- smth.gaussian(npl$new_cases, window = 8, alpha = 1)
plot(days, npl$new_cases, type = "l", main = "Bandwidth = 8", xlab = "Day", ylab = "New Cases", col = "grey", lwd = 1)
lines(days, smooth.ts, col = "orchid", lwd = 2)
lines(days, npl$new_cases_smoothed, col = "orange", lwd = 2)
legend("topright", legend = c("Raw Data", "Auto-Smoothed", "Given Smoothed"),
       col = c("grey", "orchid", "orange"), lwd = 2)
```

### Devin's Data

```{r}
require(doBy); library(smoother)
# smth.gaussian(x, window, alpha)
## Where:
### x: data
### window: the length of the smoothing window, if an integer, represents number of items, else, if a value between 0 and 1, represents the proportion of the input vector
### alpha: parameter to determine the breadth of the gaussian window, yielding more or less sensitive smoothing characteristics

data <- read.csv("Kirk_Data/Epidemics_Data_Dryad/Kirk_et_al_epidemics_sampling_data.csv")
data.sum <- summaryBy(status ~ population + treatment + day, data = data, FUN = sum)
sub8 <- subset(data.sum, population == 8); sub8_ts <- sub8$status.sum
smooth.ts <- smth.gaussian(sub8$status.sum); # smth.gaussian(x, window, alpha)
plot(sub8$day, sub8$status.sum, type = "l", main = "Default 'window' and 'alpha' Values", xlab = "Day", ylab = "Infecteds", col = "grey", lwd = 2)
lines(sub8$day, smooth.ts, col = "orchid", lwd = 2)
legend("topleft", legend = c("Raw Data", "Auto-Smoothed"),
       col = c("grey", "orchid"), lwd = 2)
# With adjusted window and alpha values
smooth.ts <- smth.gaussian(sub8$status.sum, window = 3, alpha = 1)
plot(sub8$day, sub8$status.sum, type = "l", main = "Bandwidth = 8", xlab = "Day", ylab = "Infecteds", col = "grey", lwd = 2)
lines(sub8$day, smooth.ts, col = "orchid", lwd = 2)
legend("topleft", legend = c("Raw Data", "Auto-Smoothed"),
       col = c("grey", "orchid"), lwd = 2)
```

## Manually

### COVID Data

```{r}
data <- read.csv("https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/owid-covid-data.csv")
npl <- subset(data, iso_code == "NPL")
npl.dates <- as.Date(npl$date)
## Setting up the df for the top half of the expression
day.nums <- seq(0, length(npl.dates) - 1, 1)
for (i in 2:(length(npl.dates) - 1)){
  temp <- c(seq(-i + 1, -1, 1), 0, seq(1, length(npl.dates) - i, 1))
  day.nums <- cbind(day.nums, temp)
}
day.nums <- cbind(day.nums, rev(seq(0, length(npl.dates) - 1, 1))*(-1))
## Calculating the Gaussian kernel
temp <- c(); width <- 8; smoothed <- c()
for (i in 1:length(npl.dates)){
  temp <- exp(-(day.nums[ , i]^2) / (2*(width^2)))
  temp <- temp/sum(temp)
  smoothed[i] <- sum(npl$new_cases*temp)
}
## Plotting
plot(smoothed, type = "l")
# Let'c compare it to (1) the raw data, and (2) the automatically smoothed data:
smooth.ts <- smth.gaussian(npl$new_cases, window = 8, alpha = 1)
days <- seq(1, nrow(npl), 1)
plot(days, npl$new_cases, type = "l", main = "Bandwidth = 8", xlab = "Day", ylab = "New Cases", col = "grey", lwd = 1)
lines(days, smoothed, col = "red3", lwd = 2)
lines(days, smooth.ts, col = "orchid", lwd = 2, lty = 2)
lines(days, npl$new_cases_smoothed, col = "orange", lwd = 2, lty = 3)
legend("topright", legend = c("Raw Data", "Manually Smoothed", "Auto-Smoothed", "Given Smoothed"),
       col = c("grey", "red3", "orchid", "orange"), lwd = 2, lty = c(1, 1, 2, 3))
## Summing total cases
sum(npl$new_cases); sum(smoothed) # They're the same!
```

### Devin's Data

```{r}
## Data manipulation
data <- read.csv("Kirk_Data/Epidemics_Data_Dryad/Kirk_et_al_epidemics_sampling_data.csv")
data.sum <- summaryBy(status ~ population + treatment + day, data = data, FUN = sum)
sub8 <- subset(data.sum, population == 8)
sub8 <- data.frame(sub8$day, sub8$status.sum); colnames(sub8) <- c("day", "infecteds")

## EVERYDAY
# Adding missing rows
all.days <- seq(1, 120, 1); exp.days <- seq(3, 120, 3); missing.days <- setdiff(all.days, exp.days)
missing.days <- data.frame(cbind(missing.days, c(rep(0, length(missing.days))))); colnames(missing.days) <- c("day", "infecteds")
sub8.mod <- rbind(sub8, missing.days)
sub8.mod <- sub8.mod[order(sub8.mod$day), ]
# Setting up the df for the top half of the expression
day.nums <- seq(0, length(all.days) - 1, 1)
for (i in 2:(length(all.days) - 1)){
  temp <- c(seq(-i + 1, -1, 1), 0, seq(1, length(all.days) - i, 1))
  day.nums <- cbind(day.nums, temp)
}
day.nums <- cbind(day.nums, rev(seq(0, length(all.days) - 1, 1))*(-1))
# Calculating the Gaussian kernel
temp <- c(); width <- 3; smoothed <- c()
for (i in 1:length(all.days)){
  temp <- exp(-(day.nums[ , i]^2) / (2*(width^2)))
  temp <- temp/sum(temp)
  smoothed[i] <- sum(sub8.mod$infecteds*temp)
}
# Plotting
plot(smoothed, type = "l")
# Let'c compare it to (1) the raw data, and (2) the automatically smoothed data:
smooth.ts.1 <- smth.gaussian(sub8$infecteds, window = 3, alpha = 1)
smooth.ts.2 <- smth.gaussian(sub8.mod$infecteds, window = 3, alpha = 1)
plot(sub8$day, sub8$infecteds, type = "l", main = "Bandwidth = 3", xlab = "Day", ylab = "New Cases", col = "grey", lwd = 1)
lines(sub8.mod$day, smoothed, col = "red3", lwd = 2)
lines(sub8$day, smooth.ts.1, col = "orchid", lwd = 2, lty = 2)
lines(sub8.mod$day, smooth.ts.2, col = "hotpink", lwd = 2, lty = 2)
legend("topleft", legend = c("Raw Data", "Manually Smoothed (Everyday)", "Auto-Smoothed (Every 3 Days)", "Auto-Smoothed (Everyday)"),
       col = c("grey", "red3", "orchid", "hotpink"), lwd = 2, lty = c(1, 1, 2, 2))

## EVERY 3 DAYS
# Setting up the df for the top half of the expression
interval <- 3
day.nums <- seq(0, sub8$day[length(sub8$day)] - interval, 3)
suppressWarnings(
for (i in sub8$day[1:39]){
  temp <- c(seq(-i, -interval, interval), 0, seq(interval, sub8$day[length(sub8$day)] - interval*2, interval))
  day.nums <- cbind(day.nums, temp)
})
# Calculating the Gaussian kernel
temp <- c(); width <- 3; smoothed_3days <- c()
for (i in 1:length(sub8$day)){
  temp <- exp(-(day.nums[ , i]^2) / (2*(width^2)))
  temp <- temp/sum(temp)
  smoothed_3days[i] <- sum(sub8$infecteds*temp)
}
# Plotting
plot(smoothed_3days, type = "l")
# Let'c compare it to (1) the raw data, (2) the manually smoothed over 120 days data, and (3) the automatically smoothed data:
plot(sub8$day, sub8$infecteds, type = "l", main = "Bandwidth = 3", xlab = "Day", ylab = "New Cases", col = "grey", lwd = 1)
lines(sub8.mod$day, smoothed, col = "red3", lwd = 2)
lines(sub8$day, smooth.ts.1, col = "orchid", lwd = 2, lty = 2)
lines(sub8.mod$day, smooth.ts.2, col = "hotpink", lwd = 2, lty = 2)
lines(sub8$day, smoothed_3days, col = "goldenrod1", lwd = 2)
legend("topleft", legend = c("Raw Data", "Manually Smoothed (Everyday)", "Manually Smoothed (Every 3 Days)", "Auto-Smoothed (Every 3 Days)", "Auto-Smoothed (Everyday)"),
       col = c("grey", "red3", "goldenrod1", "orchid", "hotpink"), lwd = 2, lty = c(1, 1, 1, 2, 2))
```

So, given the experimental design and data structure, I feel like we shouldn't insert the missing days, so we'll go with the "Every 3 Days" set-up.

# Playing with Bandwidth

```{r}
interval <- 3
day.nums <- seq(0, sub8$day[length(sub8$day)] - interval, 3)
suppressWarnings(
for (i in sub8$day[1:39]){
  temp <- c(seq(-i, -interval, interval), 0, seq(interval, sub8$day[length(sub8$day)] - interval*2, interval))
  day.nums <- cbind(day.nums, temp)
})

widths <- seq(1, 8, 1); df <- matrix(nrow = length(sub8$day), ncol = length(widths))
for (j in widths){
  temp <- c(); smoothed <- c()
  for (i in 1:length(sub8$day)){
    temp <- exp(-(day.nums[ , i]^2) / (2*(widths[j]^2)))
    temp <- temp/sum(temp)
    smoothed[i] <- sum(sub8$infecteds*temp)
  }
  df[ , j] <- smoothed
}

cols <- c("red3", "orange", "goldenrod1", "forestgreen", "cornflowerblue", "orchid")
plot(sub8$day, sub8$infecteds, type = "l",
     main = "Detrended Experimental Data (Population 8), Varying Bandwidths", xlab = "Day", ylab = "Infecteds",
     col = "grey", lwd = 2)
for (i in 1:length(widths)){
  lines(sub8$day, df[ , i], col = cols[i], lwd = 2, lty = 2)
}
legend("topleft", 
       legend = c("Raw Data", "Bandwidth = 1", "Bandwidth = 2", "Bandwidth = 3", "Bandwidth = 4", "Bandwidth = 5", "Bandwidth = 6"), 
       lwd = 2, col = c("grey", cols))
```

Looking at the smoothed time series above, it looks like we should go with a **bandwidth of 2 or 3**.

# Smoothing: Bandwidth = 1

## Simulated Data

```{r}
constant1 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.1.csv")
constant2 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.2.csv")
constant3 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.3.csv")
constant4 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.4.csv")
constant5 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.5.csv")
constant6 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.6.csv")
constant7 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.7.csv")
constant8 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.8.csv")
constant9 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.9.csv")
constant10 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.10.csv")
warming1 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.1.csv")
warming2 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.2.csv")
warming3 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.3.csv")
warming4 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.4.csv")
warming5 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.5.csv")
warming6 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.6.csv")
warming7 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.7.csv")
warming8 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.8.csv")
warming9 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.9.csv")
warming10 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.10.csv")
const.df <- data.frame(cbind(constant1[ , 2:101],
                             constant2[ , 2:101],
                             constant3[ , 2:101],
                             constant4[ , 2:101],
                             constant5[ , 2:101],
                             constant6[ , 2:101],
                             constant7[ , 2:101],
                             constant8[ , 2:101],
                             constant9[ , 2:101],
                             constant10[ , 2:101]))
warm.df <- data.frame(cbind(warming1[ , 2:101],
                             warming2[ , 2:101],
                             warming3[ , 2:101],
                             warming4[ , 2:101],
                             warming5[ , 2:101],
                             warming6[ , 2:101],
                             warming7[ , 2:101],
                             warming8[ , 2:101],
                             warming9[ , 2:101],
                             warming10[ , 2:101]))

all.days <- seq(1, 120, 1)
day.nums <- seq(0, length(all.days) - 1, 1)
for (i in 2:(length(all.days) - 1)){
  temp <- c(seq(-i + 1, -1, 1), 0, seq(1, length(all.days) - i, 1))
  day.nums <- cbind(day.nums, temp)
}
day.nums <- cbind(day.nums, rev(seq(0, length(all.days) - 1, 1))*(-1))

width <- 1
smooth.sim.const.b1 <- matrix(nrow = length(all.days), ncol = ncol(const.df)) 
smooth.sim.warm.b1 <- matrix(nrow = length(all.days), ncol = ncol(warm.df))
for (j in 1:ncol(const.df)){
  temp <- c(); smoothed.c <- c(); smoothed.w <- c()
  for (i in 1:length(all.days)){
    temp <- exp(-(day.nums[ , i]^2) / (2*(width^2)))
    temp <- temp/sum(temp)
    smoothed.c[i] <- sum(const.df[ , j]*temp)
    smoothed.w[i] <- sum(warm.df[ , j]*temp)
  }
  smooth.sim.const.b1[ , j] <- smoothed.c
  smooth.sim.warm.b1[ , j] <- smoothed.w
}

par(mfrow = c(1, 2))
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Constant Temperature", xlab = "Day", ylab = "Infecteds")
for (i in 1:1000){
  lines(1:120, const.df[ , i], col = "grey")
  lines(1:120, smooth.sim.const.b1[ , i], col = "cornflowerblue", lty = 2)
}
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Rising Temperature", xlab = "Day", ylab = "Infecteds")
for (i in 1:1000){
  lines(1:120, warm.df[ , i], col = "grey")
  lines(1:120, smooth.sim.warm.b1[ , i], col = "red3", lty = 2)
}

set.seed(123); sample <- sample(1:1000, 20)
par(mfrow = c(1, 2))
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Constant Temperature (Select Implementations)", xlab = "Day", ylab = "Infecteds")
for (i in sample){
  lines(1:120, const.df[ , i], col = "grey")
  lines(1:120, smooth.sim.const.b1[ , i], col = "cornflowerblue", lty = 2)
}
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Rising Temperature (Select Implementations)", xlab = "Day", ylab = "Infecteds")
for (i in sample){
  lines(1:120, warm.df[ , i], col = "grey")
  lines(1:120, smooth.sim.warm.b1[ , i], col = "red3", lty = 2)
}
```

## Empirical Data

```{r}
data <- read.csv("Kirk_Data/Epidemics_Data_Dryad/Kirk_et_al_epidemics_sampling_data.csv")
data.sum <- summaryBy(status ~ population + treatment + day, data = data, FUN = sum)
sub1 <- subset(data.sum, population == 1) 
sub1 <- cbind(sub1$day, sub1$status.sum); colnames(sub1) <- c("day", "infecteds")
sub2 <- subset(data.sum, population == 2) 
sub2 <- cbind(sub2$day, sub2$status.sum); colnames(sub2) <- c("day", "infecteds")
sub3 <- subset(data.sum, population == 3) 
sub3 <- cbind(sub3$day, sub3$status.sum); colnames(sub3) <- c("day", "infecteds")
sub4 <- subset(data.sum, population == 4) 
sub4 <- cbind(sub4$day, sub4$status.sum); colnames(sub4) <- c("day", "infecteds")
sub5 <- subset(data.sum, population == 5) 
sub5 <- cbind(sub5$day, sub5$status.sum); colnames(sub5) <- c("day", "infecteds")
sub6 <- subset(data.sum, population == 6) 
sub6 <- cbind(sub6$day, sub6$status.sum); colnames(sub6) <- c("day", "infecteds")
sub7 <- subset(data.sum, population == 7) 
sub7 <- cbind(sub7$day, sub7$status.sum); colnames(sub7) <- c("day", "infecteds")
sub8 <- subset(data.sum, population == 8) 
sub8 <- cbind(sub8$day, sub8$status.sum); colnames(sub8) <- c("day", "infecteds")

subpops.list <- list(sub1, sub2, sub3, sub4, sub5, sub6, sub7, sub8)

interval <- 3
day.nums <- seq(0, sub1[ , 1][length(sub1[ , 1])] - interval, 3)
suppressWarnings(
for (i in sub1[ , 1][1:39]){
  temp <- c(seq(-i, -interval, interval), 0, seq(interval, sub1[ , 1][length(sub1[ , 1])] - interval*2, interval))
  day.nums <- cbind(day.nums, temp)
})

width <- 1; subpops <- seq(1, 8, 1); smooth.emp.b1 <- matrix(nrow = nrow(sub1), ncol = length(widths))
for (j in subpops){
  temp <- c(); smoothed <- c()
  for (i in 1:nrow(sub1)){
    temp <- exp(-(day.nums[ , i]^2) / (2*(width^2)))
    temp <- temp/sum(temp)
    smoothed[i] <- sum(subpops.list[[j]][ , 2]*temp)
  }
  smooth.emp.b1[ , j] <- smoothed
}
cols <- c("navy", "cornflowerblue", "steelblue2", "lightblue2")
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 5),
     main = "Constant Temperature", xlab = "Day", ylab = "Infecteds")
for (i in 1:4){
  lines(sub1[ , 1], subpops.list[[i]][ , 2], col = "grey", lwd = 2)
  lines(sub1[ , 1], smooth.emp.b1[ , i], col = cols[i], lwd = 2)
}
cols <- c("darkred", "firebrick", "red2", "tomato")
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 5),
     main = "Rising Temperature", xlab = "Day", ylab = "Infecteds")
for (i in 5:8){
  lines(sub1[ , 1], subpops.list[[i]][ , 2], col = "grey", lwd = 2)
  lines(sub1[ , 1], smooth.emp.b1[ , i], col = cols[i - 4], lwd = 2)
}
```

### Plotting

```{r}
png("Plots/EmpDat_Detrend_B1.png", width = 2000, height = 800)
par(mfrow = c(2, 4))
plot(sub1[ , 1], subpops.list[[1]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 1 (Constant) | Bandwidth = 4", xlab = "Day", ylab = "Infecteds")
lines(sub1[ , 1], smooth.emp.b1[ , 1], col = "cornflowerblue", lwd = 2, lty = 2)
plot(sub2[ , 1], subpops.list[[2]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 2 (Constant) | Bandwidth = 4", xlab = "Day", ylab = "Infecteds")
lines(sub2[ , 1], smooth.emp.b1[ , 2], col = "cornflowerblue", lwd = 2, lty = 2)
plot(sub4[ , 1], subpops.list[[3]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 3 (Constant) | Bandwidth = 4", xlab = "Day", ylab = "Infecteds")
lines(sub4[ , 1], smooth.emp.b1[ , 3], col = "cornflowerblue", lwd = 2, lty = 2)
plot(sub4[ , 1], subpops.list[[4]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 4 (Constant) | Bandwidth = 4", xlab = "Day", ylab = "Infecteds")
lines(sub4[ , 1], smooth.emp.b1[ , 4], col = "cornflowerblue", lwd = 2, lty = 2)
plot(sub5[ , 1], subpops.list[[5]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 5 (Warming) | Bandwidth = 4", xlab = "Day", ylab = "Infecteds")
lines(sub5[ , 1], smooth.emp.b1[ , 5], col = "red3", lwd = 2, lty = 2)
plot(sub6[ , 1], subpops.list[[6]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 6 (Warming) | Bandwidth = 4", xlab = "Day", ylab = "Infecteds")
lines(sub6[ , 1], smooth.emp.b1[ , 6], col = "red3", lwd = 2, lty = 2)
plot(sub7[ , 1], subpops.list[[7]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 7 (Warming) | Bandwidth = 4", xlab = "Day", ylab = "Infecteds")
lines(sub7[ , 1], smooth.emp.b1[ , 7], col = "red3", lwd = 2, lty = 2)
plot(sub8[ , 1], subpops.list[[8]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 8 (Warming) | Bandwidth = 4", xlab = "Day", ylab = "Infecteds")
lines(sub8[ , 1], smooth.emp.b1[ , 8], col = "red3", lwd = 2, lty = 2)
dev.off()
```

# Smoothing: Bandwidth = 2

## Simulated Data

```{r}
all.days <- seq(1, 120, 1)
day.nums <- seq(0, length(all.days) - 1, 1)
for (i in 2:(length(all.days) - 1)){
  temp <- c(seq(-i + 1, -1, 1), 0, seq(1, length(all.days) - i, 1))
  day.nums <- cbind(day.nums, temp)
}
day.nums <- cbind(day.nums, rev(seq(0, length(all.days) - 1, 1))*(-1))

width <- 2
smooth.sim.const.b2 <- matrix(nrow = length(all.days), ncol = ncol(const.df)) 
smooth.sim.warm.b2 <- matrix(nrow = length(all.days), ncol = ncol(warm.df))
for (j in 1:ncol(const.df)){
  temp <- c(); smoothed.c <- c(); smoothed.w <- c()
  for (i in 1:length(all.days)){
    temp <- exp(-(day.nums[ , i]^2) / (2*(width^2)))
    temp <- temp/sum(temp)
    smoothed.c[i] <- sum(const.df[ , j]*temp)
    smoothed.w[i] <- sum(warm.df[ , j]*temp)
  }
  smooth.sim.const.b2[ , j] <- smoothed.c
  smooth.sim.warm.b2[ , j] <- smoothed.w
}

par(mfrow = c(1, 2))
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Constant Temperature", xlab = "Day", ylab = "Infecteds")
for (i in 1:1000){
  lines(1:120, const.df[ , i], col = "grey")
  lines(1:120, smooth.sim.const.b2[ , i], col = "cornflowerblue", lty = 2)
}
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Rising Temperature", xlab = "Day", ylab = "Infecteds")
for (i in 1:1000){
  lines(1:120, warm.df[ , i], col = "grey")
  lines(1:120, smooth.sim.warm.b2[ , i], col = "red3", lty = 2)
}

set.seed(123); sample <- sample(1:1000, 20)
par(mfrow = c(1, 2))
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Constant Temperature (Select Implementations)", xlab = "Day", ylab = "Infecteds")
for (i in sample){
  lines(1:120, const.df[ , i], col = "grey")
  lines(1:120, smooth.sim.const.b2[ , i], col = "cornflowerblue", lty = 2)
}
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Rising Temperature (Select Implementations)", xlab = "Day", ylab = "Infecteds")
for (i in sample){
  lines(1:120, warm.df[ , i], col = "grey")
  lines(1:120, smooth.sim.warm.b2[ , i], col = "red3", lty = 2)
}
```

## Empirical Data

```{r}
interval <- 3
day.nums <- seq(0, sub1[ , 1][length(sub1[ , 1])] - interval, 3)
suppressWarnings(
for (i in sub1[ , 1][1:39]){
  temp <- c(seq(-i, -interval, interval), 0, seq(interval, sub1[ , 1][length(sub1[ , 1])] - interval*2, interval))
  day.nums <- cbind(day.nums, temp)
})

width <- 2; subpops <- seq(1, 8, 1); smooth.emp.b2 <- matrix(nrow = nrow(sub1), ncol = length(widths))
for (j in subpops){
  temp <- c(); smoothed <- c()
  for (i in 1:nrow(sub1)){
    temp <- exp(-(day.nums[ , i]^2) / (2*(width^2)))
    temp <- temp/sum(temp)
    smoothed[i] <- sum(subpops.list[[j]][ , 2]*temp)
  }
  smooth.emp.b2[ , j] <- smoothed
}
cols <- c("navy", "cornflowerblue", "steelblue2", "lightblue2")
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 5),
     main = "Constant Temperature", xlab = "Day", ylab = "Infecteds")
for (i in 1:4){
  lines(sub1[ , 1], subpops.list[[i]][ , 2], col = "grey", lwd = 2)
  lines(sub1[ , 1], smooth.emp.b2[ , i], col = cols[i], lwd = 2)
}
cols <- c("darkred", "firebrick", "red2", "tomato")
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 5),
     main = "Rising Temperature", xlab = "Day", ylab = "Infecteds")
for (i in 5:8){
  lines(sub1[ , 1], subpops.list[[i]][ , 2], col = "grey", lwd = 2)
  lines(sub1[ , 1], smooth.emp.b2[ , i], col = cols[i - 4], lwd = 2)
}
```

### Plotting

```{r}
png("Plots/EmpDat_Detrend_B2.png", width = 2000, height = 800)
par(mfrow = c(2, 4))
plot(sub1[ , 1], subpops.list[[1]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 1 (Constant) | Bandwidth = 2", xlab = "Day", ylab = "Infecteds")
lines(sub1[ , 1], smooth.emp.b2[ , 1], col = "cornflowerblue", lwd = 2)
plot(sub2[ , 1], subpops.list[[2]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 2 (Constant) | Bandwidth = 2", xlab = "Day", ylab = "Infecteds")
lines(sub2[ , 1], smooth.emp.b2[ , 2], col = "cornflowerblue", lwd = 2)
plot(sub2[ , 1], subpops.list[[3]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 3 (Constant) | Bandwidth = 2", xlab = "Day", ylab = "Infecteds")
lines(sub2[ , 1], smooth.emp.b2[ , 3], col = "cornflowerblue", lwd = 2)
plot(sub4[ , 1], subpops.list[[4]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 4 (Constant) | Bandwidth = 2", xlab = "Day", ylab = "Infecteds")
lines(sub4[ , 1], smooth.emp.b2[ , 4], col = "cornflowerblue", lwd = 2)
plot(sub5[ , 1], subpops.list[[5]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 5 (Warming) | Bandwidth = 2", xlab = "Day", ylab = "Infecteds")
lines(sub5[ , 1], smooth.emp.b2[ , 5], col = "red3", lwd = 2)
plot(sub6[ , 1], subpops.list[[6]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 6 (Warming) | Bandwidth = 2", xlab = "Day", ylab = "Infecteds")
lines(sub6[ , 1], smooth.emp.b2[ , 6], col = "red3", lwd = 2)
plot(sub7[ , 1], subpops.list[[7]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 7 (Warming) | Bandwidth = 2", xlab = "Day", ylab = "Infecteds")
lines(sub7[ , 1], smooth.emp.b2[ , 7], col = "red3", lwd = 2)
plot(sub8[ , 1], subpops.list[[8]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 8 (Warming) | Bandwidth = 2", xlab = "Day", ylab = "Infecteds")
lines(sub8[ , 1], smooth.emp.b2[ , 8], col = "red3", lwd = 2)
dev.off()
```

# Smoothing: Bandwidth = 3

## Simulated Data

```{r}
constant1 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.1.csv")
constant2 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.2.csv")
constant3 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.3.csv")
constant4 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.4.csv")
constant5 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.5.csv")
constant6 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.6.csv")
constant7 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.7.csv")
constant8 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.8.csv")
constant9 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.9.csv")
constant10 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.10.csv")
warming1 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.1.csv")
warming2 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.2.csv")
warming3 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.3.csv")
warming4 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.4.csv")
warming5 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.5.csv")
warming6 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.6.csv")
warming7 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.7.csv")
warming8 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.8.csv")
warming9 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.9.csv")
warming10 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.10.csv")
const.df <- data.frame(cbind(constant1[ , 2:101],
                             constant2[ , 2:101],
                             constant3[ , 2:101],
                             constant4[ , 2:101],
                             constant5[ , 2:101],
                             constant6[ , 2:101],
                             constant7[ , 2:101],
                             constant8[ , 2:101],
                             constant9[ , 2:101],
                             constant10[ , 2:101]))
warm.df <- data.frame(cbind(warming1[ , 2:101],
                             warming2[ , 2:101],
                             warming3[ , 2:101],
                             warming4[ , 2:101],
                             warming5[ , 2:101],
                             warming6[ , 2:101],
                             warming7[ , 2:101],
                             warming8[ , 2:101],
                             warming9[ , 2:101],
                             warming10[ , 2:101]))

all.days <- seq(1, 120, 1)
day.nums <- seq(0, length(all.days) - 1, 1)
for (i in 2:(length(all.days) - 1)){
  temp <- c(seq(-i + 1, -1, 1), 0, seq(1, length(all.days) - i, 1))
  day.nums <- cbind(day.nums, temp)
}
day.nums <- cbind(day.nums, rev(seq(0, length(all.days) - 1, 1))*(-1))

width <- 3 
smooth.sim.const.b3 <- matrix(nrow = length(all.days), ncol = ncol(const.df)) 
smooth.sim.warm.b3 <- matrix(nrow = length(all.days), ncol = ncol(warm.df))
for (j in 1:ncol(const.df)){
  temp <- c(); smoothed.c <- c(); smoothed.w <- c()
  for (i in 1:length(all.days)){
    temp <- exp(-(day.nums[ , i]^2) / (2*(width^2)))
    temp <- temp/sum(temp)
    smoothed.c[i] <- sum(const.df[ , j]*temp)
    smoothed.w[i] <- sum(warm.df[ , j]*temp)
  }
  smooth.sim.const.b3[ , j] <- smoothed.c
  smooth.sim.warm.b3[ , j] <- smoothed.w
}

par(mfrow = c(1, 2))
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Constant Temperature", xlab = "Day", ylab = "Infecteds")
for (i in 1:1000){
  lines(1:120, const.df[ , i], col = "grey")
  lines(1:120, smooth.sim.const.b3[ , i], col = "cornflowerblue", lty = 2)
}
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Rising Temperature", xlab = "Day", ylab = "Infecteds")
for (i in 1:1000){
  lines(1:120, warm.df[ , i], col = "grey")
  lines(1:120, smooth.sim.warm.b3[ , i], col = "red3", lty = 2)
}

set.seed(123); sample <- sample(1:1000, 20)
par(mfrow = c(1, 2))
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Constant Temperature (Select Implementations)", xlab = "Day", ylab = "Infecteds")
for (i in sample){
  lines(1:120, const.df[ , i], col = "grey")
  lines(1:120, smooth.sim.const.b3[ , i], col = "cornflowerblue", lty = 2)
}
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Rising Temperature (Select Implementations)", xlab = "Day", ylab = "Infecteds")
for (i in sample){
  lines(1:120, warm.df[ , i], col = "grey")
  lines(1:120, smooth.sim.warm.b3[ , i], col = "red3", lty = 2)
}
```

## Empirical Data

```{r}
data <- read.csv("Kirk_Data/Epidemics_Data_Dryad/Kirk_et_al_epidemics_sampling_data.csv")
data.sum <- summaryBy(status ~ population + treatment + day, data = data, FUN = sum)
sub1 <- subset(data.sum, population == 1) 
sub1 <- cbind(sub1$day, sub1$status.sum); colnames(sub1) <- c("day", "infecteds")
sub2 <- subset(data.sum, population == 2) 
sub2 <- cbind(sub2$day, sub2$status.sum); colnames(sub2) <- c("day", "infecteds")
sub3 <- subset(data.sum, population == 3) 
sub3 <- cbind(sub3$day, sub3$status.sum); colnames(sub3) <- c("day", "infecteds")
sub4 <- subset(data.sum, population == 4) 
sub4 <- cbind(sub4$day, sub4$status.sum); colnames(sub4) <- c("day", "infecteds")
sub5 <- subset(data.sum, population == 5) 
sub5 <- cbind(sub5$day, sub5$status.sum); colnames(sub5) <- c("day", "infecteds")
sub6 <- subset(data.sum, population == 6) 
sub6 <- cbind(sub6$day, sub6$status.sum); colnames(sub6) <- c("day", "infecteds")
sub7 <- subset(data.sum, population == 7) 
sub7 <- cbind(sub7$day, sub7$status.sum); colnames(sub7) <- c("day", "infecteds")
sub8 <- subset(data.sum, population == 8) 
sub8 <- cbind(sub8$day, sub8$status.sum); colnames(sub8) <- c("day", "infecteds")

subpops.list <- list(sub1, sub2, sub3, sub4, sub5, sub6, sub7, sub8)

interval <- 3
day.nums <- seq(0, sub1[ , 1][length(sub1[ , 1])] - interval, 3)
suppressWarnings(
for (i in sub1[ , 1][1:39]){
  temp <- c(seq(-i, -interval, interval), 0, seq(interval, sub1[ , 1][length(sub1[ , 1])] - interval*2, interval))
  day.nums <- cbind(day.nums, temp)
})

width <- 3; subpops <- seq(1, 8, 1); smooth.emp.b3 <- matrix(nrow = nrow(sub1), ncol = length(widths))
for (j in subpops){
  temp <- c(); smoothed <- c()
  for (i in 1:nrow(sub1)){
    temp <- exp(-(day.nums[ , i]^2) / (2*(width^2)))
    temp <- temp/sum(temp)
    smoothed[i] <- sum(subpops.list[[j]][ , 2]*temp)
  }
  smooth.emp.b3[ , j] <- smoothed
}
cols <- c("navy", "cornflowerblue", "steelblue2", "lightblue2")
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 5),
     main = "Constant Temperature", xlab = "Day", ylab = "Infecteds")
for (i in 1:4){
  lines(sub1[ , 1], subpops.list[[i]][ , 2], col = "grey", lwd = 2)
  lines(sub1[ , 1], smooth.emp.b3[ , i], col = cols[i], lwd = 2)
}
cols <- c("darkred", "firebrick", "red2", "tomato")
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 5),
     main = "Rising Temperature", xlab = "Day", ylab = "Infecteds")
for (i in 5:8){
  lines(sub1[ , 1], subpops.list[[i]][ , 2], col = "grey", lwd = 2)
  lines(sub1[ , 1], smooth.emp.b3[ , i], col = cols[i - 4], lwd = 2)
}
```

### Plotting

```{r}
png("Plots/EmpDat_Detrend_B3.png", width = 2000, height = 800)
par(mfrow = c(2, 4))
plot(sub1[ , 1], subpops.list[[1]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 1 (Constant) | Bandwidth = 3", xlab = "Day", ylab = "Infecteds")
lines(sub1[ , 1], smooth.emp.b3[ , 1], col = "cornflowerblue", lwd = 2)
plot(sub2[ , 1], subpops.list[[2]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 2 (Constant) | Bandwidth = 3", xlab = "Day", ylab = "Infecteds")
lines(sub2[ , 1], smooth.emp.b3[ , 2], col = "cornflowerblue", lwd = 2)
plot(sub3[ , 1], subpops.list[[3]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 3 (Constant) | Bandwidth = 3", xlab = "Day", ylab = "Infecteds")
lines(sub3[ , 1], smooth.emp.b3[ , 3], col = "cornflowerblue", lwd = 2)
plot(sub4[ , 1], subpops.list[[4]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 4 (Constant) | Bandwidth = 3", xlab = "Day", ylab = "Infecteds")
lines(sub4[ , 1], smooth.emp.b3[ , 4], col = "cornflowerblue", lwd = 2)
plot(sub5[ , 1], subpops.list[[5]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 5 (Warming) | Bandwidth = 3", xlab = "Day", ylab = "Infecteds")
lines(sub5[ , 1], smooth.emp.b3[ , 5], col = "red3", lwd = 2)
plot(sub6[ , 1], subpops.list[[6]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 6 (Warming) | Bandwidth = 3", xlab = "Day", ylab = "Infecteds")
lines(sub6[ , 1], smooth.emp.b3[ , 6], col = "red3", lwd = 2)
plot(sub7[ , 1], subpops.list[[7]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 7 (Warming) | Bandwidth = 3", xlab = "Day", ylab = "Infecteds")
lines(sub7[ , 1], smooth.emp.b3[ , 7], col = "red3", lwd = 2)
plot(sub8[ , 1], subpops.list[[8]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 8 (Warming) | Bandwidth = 3", xlab = "Day", ylab = "Infecteds")
lines(sub8[ , 1], smooth.emp.b3[ , 8], col = "red3", lwd = 2)
dev.off()
```

# Smoothing: Bandwidth = 4

## Simulated Data

```{r}
all.days <- seq(1, 120, 1)
day.nums <- seq(0, length(all.days) - 1, 1)
for (i in 2:(length(all.days) - 1)){
  temp <- c(seq(-i + 1, -1, 1), 0, seq(1, length(all.days) - i, 1))
  day.nums <- cbind(day.nums, temp)
}
day.nums <- cbind(day.nums, rev(seq(0, length(all.days) - 1, 1))*(-1))

width <- 4
smooth.sim.const.b4 <- matrix(nrow = length(all.days), ncol = ncol(const.df)) 
smooth.sim.warm.b4 <- matrix(nrow = length(all.days), ncol = ncol(warm.df))
for (j in 1:ncol(const.df)){
  temp <- c(); smoothed.c <- c(); smoothed.w <- c()
  for (i in 1:length(all.days)){
    temp <- exp(-(day.nums[ , i]^2) / (2*(width^2)))
    temp <- temp/sum(temp)
    smoothed.c[i] <- sum(const.df[ , j]*temp)
    smoothed.w[i] <- sum(warm.df[ , j]*temp)
  }
  smooth.sim.const.b4[ , j] <- smoothed.c
  smooth.sim.warm.b4[ , j] <- smoothed.w
}

par(mfrow = c(1, 2))
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Constant Temperature", xlab = "Day", ylab = "Infecteds")
for (i in 1:1000){
  lines(1:120, const.df[ , i], col = "grey")
  lines(1:120, smooth.sim.const.b4[ , i], col = "cornflowerblue", lty = 2)
}
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Rising Temperature", xlab = "Day", ylab = "Infecteds")
for (i in 1:1000){
  lines(1:120, warm.df[ , i], col = "grey")
  lines(1:120, smooth.sim.warm.b4[ , i], col = "red3", lty = 2)
}

set.seed(123); sample <- sample(1:1000, 20)
par(mfrow = c(1, 2))
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Constant Temperature (Select Implementations)", xlab = "Day", ylab = "Infecteds")
for (i in sample){
  lines(1:120, const.df[ , i], col = "grey")
  lines(1:120, smooth.sim.const.b4[ , i], col = "cornflowerblue", lty = 2)
}
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Rising Temperature (Select Implementations)", xlab = "Day", ylab = "Infecteds")
for (i in sample){
  lines(1:120, warm.df[ , i], col = "grey")
  lines(1:120, smooth.sim.warm.b4[ , i], col = "red3", lty = 2)
}
```

## Empirical Data

```{r}
interval <- 3
day.nums <- seq(0, sub1[ , 1][length(sub1[ , 1])] - interval, 3)
suppressWarnings(
for (i in sub1[ , 1][1:39]){
  temp <- c(seq(-i, -interval, interval), 0, seq(interval, sub1[ , 1][length(sub1[ , 1])] - interval*2, interval))
  day.nums <- cbind(day.nums, temp)
})

width <- 4; subpops <- seq(1, 8, 1); smooth.emp.b4 <- matrix(nrow = nrow(sub1), ncol = length(widths))
for (j in subpops){
  temp <- c(); smoothed <- c()
  for (i in 1:nrow(sub1)){
    temp <- exp(-(day.nums[ , i]^2) / (2*(width^2)))
    temp <- temp/sum(temp)
    smoothed[i] <- sum(subpops.list[[j]][ , 2]*temp)
  }
  smooth.emp.b4[ , j] <- smoothed
}
cols <- c("navy", "cornflowerblue", "steelblue2", "lightblue2")
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 5),
     main = "Constant Temperature", xlab = "Day", ylab = "Infecteds")
for (i in 1:4){
  lines(sub1[ , 1], subpops.list[[i]][ , 2], col = "grey", lwd = 2)
  lines(sub1[ , 1], smooth.emp.b4[ , i], col = cols[i], lwd = 2)
}
cols <- c("darkred", "firebrick", "red2", "tomato")
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 5),
     main = "Rising Temperature", xlab = "Day", ylab = "Infecteds")
for (i in 5:8){
  lines(sub1[ , 1], subpops.list[[i]][ , 2], col = "grey", lwd = 2)
  lines(sub1[ , 1], smooth.emp.b4[ , i], col = cols[i - 4], lwd = 2)
}
```

### Plotting

```{r}
png("Plots/EmpDat_Detrend_B4.png", width = 2000, height = 800)
par(mfrow = c(2, 4))
plot(sub1[ , 1], subpops.list[[1]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 1 (Constant) | Bandwidth = 4", xlab = "Day", ylab = "Infecteds")
lines(sub1[ , 1], smooth.emp.b4[ , 1], col = "cornflowerblue", lwd = 2)
plot(sub2[ , 1], subpops.list[[2]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 2 (Constant) | Bandwidth = 4", xlab = "Day", ylab = "Infecteds")
lines(sub2[ , 1], smooth.emp.b4[ , 2], col = "cornflowerblue", lwd = 2)
plot(sub4[ , 1], subpops.list[[3]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 3 (Constant) | Bandwidth = 4", xlab = "Day", ylab = "Infecteds")
lines(sub4[ , 1], smooth.emp.b4[ , 3], col = "cornflowerblue", lwd = 2)
plot(sub4[ , 1], subpops.list[[4]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 4 (Constant) | Bandwidth = 4", xlab = "Day", ylab = "Infecteds")
lines(sub4[ , 1], smooth.emp.b4[ , 4], col = "cornflowerblue", lwd = 2)
plot(sub5[ , 1], subpops.list[[5]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 5 (Warming) | Bandwidth = 4", xlab = "Day", ylab = "Infecteds")
lines(sub5[ , 1], smooth.emp.b4[ , 5], col = "red3", lwd = 2)
plot(sub6[ , 1], subpops.list[[6]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 6 (Warming) | Bandwidth = 4", xlab = "Day", ylab = "Infecteds")
lines(sub6[ , 1], smooth.emp.b4[ , 6], col = "red3", lwd = 2)
plot(sub7[ , 1], subpops.list[[7]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 7 (Warming) | Bandwidth = 4", xlab = "Day", ylab = "Infecteds")
lines(sub7[ , 1], smooth.emp.b4[ , 7], col = "red3", lwd = 2)
plot(sub8[ , 1], subpops.list[[8]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 8 (Warming) | Bandwidth = 4", xlab = "Day", ylab = "Infecteds")
lines(sub8[ , 1], smooth.emp.b4[ , 8], col = "red3", lwd = 2)
dev.off()
```

# Plotting

## Smoothed Simulated Data (Varying Bandwidths)

```{r}
png("Plots/SimDat_Detrend_VaryBandwidth.png", width = 2000, height = 800)
set.seed(123); sample <- sample(1:1000, 20)
par(mfrow = c(1, 2))
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Constant Temperature (Select Implementations) | Varying Bandwidths", xlab = "Day", ylab = "Infecteds")
for (i in sample){
  lines(1:120, const.df[ , i], col = "grey")
  lines(1:120, smooth.sim.const.b2[ , i], col = "navy", lty = 2)
  lines(1:120, smooth.sim.const.b3[ , i], col = "cornflowerblue", lty = 3)
  lines(1:120, smooth.sim.const.b4[ , i], col = "lightblue", lty = 4)
}
legend("topleft", legend = c("Raw Data", "Bandwidth = 2", "Bandwidth = 3", "Bandwidth = 4"), 
       col = c("grey", "navy", "cornflowerblue", "lightblue"), lty = c(1, 2, 3, 4), lwd = 2)
plot(0, 0, type = "n", xlim = c(1, 120), ylim = c(0, 120),
     main = "Rising Temperature (Select Implementations) | Varying Bandwidths", xlab = "Day", ylab = "Infecteds")
for (i in sample){
  lines(1:120, warm.df[ , i], col = "grey")
  lines(1:120, smooth.sim.warm.b2[ , i], col = "darkred", lty = 2)
  lines(1:120, smooth.sim.warm.b3[ , i], col = "red3", lty = 3)
  lines(1:120, smooth.sim.warm.b4[ , i], col = "orchid", lty = 4)
}
legend("topleft", legend = c("Raw Data", "Bandwidth = 2", "Bandwidth = 3", "Bandwidth = 4"), 
       col = c("grey", "darkred", "red3", "orchid"), lty = c(1, 2, 3, 4), lwd = 2)
dev.off()
```

## Smoothed Empirical Data (Varying Bandwidths)

```{r}
png("Plots/EmpDat_Detrend_VaryBandwidth.png", width = 2000, height = 800)
par(mfrow = c(2, 4))
plot(sub1[ , 1], subpops.list[[1]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 1 (Constant)", xlab = "Day", ylab = "Infecteds")
lines(sub1[ , 1], smooth.emp.b2[ , 1], col = "navy", lwd = 2)
lines(sub1[ , 1], smooth.emp.b3[ , 1], col = "cornflowerblue", lwd = 2)
lines(sub1[ , 1], smooth.emp.b4[ , 1], col = "lightblue", lwd = 2)
legend("topleft", legend = c("Raw Data", "Bandwidth = 2", "Bandwidth = 3", "Bandwidth = 4"), 
       col = c("grey", "navy", "cornflowerblue", "lightblue"), lty = 1, lwd = 2)

plot(sub1[ , 1], subpops.list[[2]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 2 (Constant)", xlab = "Day", ylab = "Infecteds")
lines(sub1[ , 1], smooth.emp.b2[ , 2], col = "navy", lwd = 2)
lines(sub1[ , 1], smooth.emp.b3[ , 2], col = "cornflowerblue", lwd = 2)
lines(sub1[ , 1], smooth.emp.b4[ , 2], col = "lightblue", lwd = 2)
legend("topleft", legend = c("Raw Data", "Bandwidth = 2", "Bandwidth = 3", "Bandwidth = 4"), 
       col = c("grey", "navy", "cornflowerblue", "lightblue"), lty = 1, lwd = 2)

plot(sub1[ , 1], subpops.list[[3]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 3 (Constant)", xlab = "Day", ylab = "Infecteds")
lines(sub1[ , 1], smooth.emp.b2[ , 3], col = "navy", lwd = 2)
lines(sub1[ , 1], smooth.emp.b3[ , 3], col = "cornflowerblue", lwd = 2)
lines(sub1[ , 1], smooth.emp.b4[ , 3], col = "lightblue", lwd = 2)
legend("topleft", legend = c("Raw Data", "Bandwidth = 2", "Bandwidth = 3", "Bandwidth = 4"), 
       col = c("grey", "navy", "cornflowerblue", "lightblue"), lty = 1, lwd = 2)

plot(sub1[ , 1], subpops.list[[4]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 4 (Constant)", xlab = "Day", ylab = "Infecteds")
lines(sub1[ , 1], smooth.emp.b2[ , 4], col = "navy", lwd = 2)
lines(sub1[ , 1], smooth.emp.b3[ , 4], col = "cornflowerblue", lwd = 2)
lines(sub1[ , 1], smooth.emp.b4[ , 4], col = "lightblue", lwd = 2)
legend("topleft", legend = c("Raw Data", "Bandwidth = 2", "Bandwidth = 3", "Bandwidth = 4"), 
       col = c("grey", "navy", "cornflowerblue", "lightblue"), lty = 1, lwd = 2)

plot(sub1[ , 1], subpops.list[[5]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 5 (Warming)", xlab = "Day", ylab = "Infecteds")
lines(sub1[ , 1], smooth.emp.b2[ , 5], col = "darkred", lwd = 2)
lines(sub1[ , 1], smooth.emp.b3[ , 5], col = "red3", lwd = 2)
lines(sub1[ , 1], smooth.emp.b4[ , 5], col = "orchid", lwd = 2)
legend("topleft", legend = c("Raw Data", "Bandwidth = 2", "Bandwidth = 3", "Bandwidth = 4"), 
       col = c("grey", "darkred", "red3", "orchid"), lty = 1, lwd = 2)

plot(sub1[ , 1], subpops.list[[6]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 6 (Warming)", xlab = "Day", ylab = "Infecteds")
lines(sub1[ , 1], smooth.emp.b2[ , 6], col = "darkred", lwd = 2)
lines(sub1[ , 1], smooth.emp.b3[ , 6], col = "red3", lwd = 2)
lines(sub1[ , 1], smooth.emp.b4[ , 6], col = "orchid", lwd = 2)
legend("topleft", legend = c("Raw Data", "Bandwidth = 2", "Bandwidth = 3", "Bandwidth = 4"), 
       col = c("grey", "darkred", "red3", "orchid"), lty = 1, lwd = 2)

plot(sub1[ , 1], subpops.list[[7]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 7 (Warming)", xlab = "Day", ylab = "Infecteds")
lines(sub1[ , 1], smooth.emp.b2[ , 7], col = "darkred", lwd = 2)
lines(sub1[ , 1], smooth.emp.b3[ , 7], col = "red3", lwd = 2)
lines(sub1[ , 1], smooth.emp.b4[ , 7], col = "orchid", lwd = 2)
legend("topleft", legend = c("Raw Data", "Bandwidth = 2", "Bandwidth = 3", "Bandwidth = 4"), 
       col = c("grey", "darkred", "red3", "orchid"), lty = 1, lwd = 2)

plot(sub1[ , 1], subpops.list[[8]][ , 2], type = "l", ylim = c(0, 5), col = "grey", lwd = 2,
     main = "Subpopulation 8 (Warming)", xlab = "Day", ylab = "Infecteds")
lines(sub1[ , 1], smooth.emp.b2[ , 8], col = "darkred", lwd = 2)
lines(sub1[ , 1], smooth.emp.b3[ , 8], col = "red3", lwd = 2)
lines(sub1[ , 1], smooth.emp.b4[ , 8], col = "orchid", lwd = 2)
legend("topleft", legend = c("Raw Data", "Bandwidth = 2", "Bandwidth = 3", "Bandwidth = 4"), 
       col = c("grey", "darkred", "red3", "orchid"), lty = 1, lwd = 2)
dev.off()
```

# Saving Relevant Objects

```{r}
saveRDS(smooth.sim.const.b1, "Objects/smooth-sim-const-b1.rds")
saveRDS(smooth.sim.warm.b1, "Objects/smooth-sim-warm-b1.rds")
saveRDS(smooth.sim.const.b2, "Objects/smooth-sim-const-b2.rds")
saveRDS(smooth.sim.warm.b2, "Objects/smooth-sim-warm-b2.rds")
saveRDS(smooth.sim.const.b3, "Objects/smooth-sim-const-b3.rds")
saveRDS(smooth.sim.warm.b3, "Objects/smooth-sim-warm-b3.rds")
saveRDS(smooth.sim.const.b4, "Objects/smooth-sim-const-b4.rds")
saveRDS(smooth.sim.warm.b4, "Objects/smooth-sim-warm-b4.rds")
saveRDS(smooth.emp.b1, "Objects/smooth-emp-b1.rds")
saveRDS(smooth.emp.b2, "Objects/smooth-emp-b2.rds")
saveRDS(smooth.emp.b3, "Objects/smooth-emp-b3.rds")
saveRDS(smooth.emp.b4, "Objects/smooth-emp-b4.rds")
```
