==> 6_EWS-Detrend-SimData.Rmd <==
# Looking for EWS in Detrended Simulated Data

## Required Packages

```{r, include = FALSE}
require(scales)
require(evobiR)
require(gdata)
require(doBy)
require(ecp)
require(strucchange)
require(TTR)
require(moments)
```

# Simulated Data

Both of these data frames are 120 rows (days) by 1000 columns (populations).

```{r}
smooth.sim.const.b2 <- readRDS("Objects/smooth-sim-const-b2.rds")
smooth.sim.warm.b2 <- readRDS("Objects/smooth-sim-warm-b2.rds")
smooth.sim.const.b3 <- readRDS("Objects/smooth-sim-const-b3.rds")
smooth.sim.warm.b3 <- readRDS("Objects/smooth-sim-warm-b3.rds")
smooth.sim.const.b4 <- readRDS("Objects/smooth-sim-const-b4.rds")
smooth.sim.warm.b4 <- readRDS("Objects/smooth-sim-warm-b4.rds")
```

# BANDWIDTH = 3

# Calculating Statistical Metrics Along Variously-Sized Sliding Windows

For details on these metrics, see `3_EWS-Sim.Rmd`.

## 5 Day Sliding Window

### Standard Deviation, Mean, Skewness, Kurtosis

```{r}
num.of.sims <- 10^3

# Making empty matricies
constant.sims.sd.5 <- matrix(nrow = 115, ncol = num.of.sims)
constant.sims.mean.5 <- matrix(nrow = 115, ncol = num.of.sims)
constant.sims.sk.5 <- matrix(nrow = 115, ncol = num.of.sims)
constant.sims.kt.5 <- matrix(nrow = 115, ncol = num.of.sims)

warming.sims.sd.5 <- matrix(nrow = 115, ncol = num.of.sims)
warming.sims.mean.5 <- matrix(nrow = 115, ncol = num.of.sims)
warming.sims.sk.5 <- matrix(nrow = 115, ncol = num.of.sims)
warming.sims.kt.5 <- matrix(nrow = 115, ncol = num.of.sims)

# Use SlidingWindow to calculate sd, mean, skewness, and kurtosis
for(i in 1:num.of.sims){
  constant.sims.sd.5[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.const.b3[,i], window = 5,step = 1)
  constant.sims.mean.5[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.const.b3[,i], window = 5, step = 1)
  constant.sims.sk.5[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.const.b3[,i], window = 5, step = 1)
  constant.sims.kt.5[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.const.b3[,i], window = 5, step = 1)
}

for(i in 1:num.of.sims){
  warming.sims.sd.5[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.warm.b3[,i], window = 5,step = 1)
  warming.sims.mean.5[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.warm.b3[,i], window = 5, step = 1)
  warming.sims.sk.5[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.warm.b3[,i], window = 5, step = 1)
  warming.sims.kt.5[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.warm.b3[,i], window = 5, step = 1)
}
```

### Variation

```{r}
constant.sims.var.5 <- constant.sims.sd.5*constant.sims.sd.5
warming.sims.var.5 <- warming.sims.sd.5*warming.sims.sd.5
```

### Co-Efficient of Variation

```{r}
constant.sims.cv.5 <- constant.sims.sd.5/constant.sims.mean.5
warming.sims.cv.5 <- warming.sims.sd.5/warming.sims.mean.5
```

### Index of Dispersion

```{r}
constant.sims.id.5 <- constant.sims.var.5/constant.sims.mean.5
warming.sims.id.5 <- warming.sims.var.5/warming.sims.mean.5
```

### First Difference Variance

```{r}
constant.sims.fdv.5 <- matrix(ncol = num.of.sims, nrow = 115)
warming.sims.fdv.5 <- matrix(ncol = num.of.sims, nrow = 115)

for(i in 1:num.of.sims){
  for(j in 2:115){
    constant.sims.fdv.5[j - 1,i] <- constant.sims.var.5[j,i] - constant.sims.var.5[j - 1,i]
    warming.sims.fdv.5[j - 1,i] <- warming.sims.var.5[j,i] - warming.sims.var.5[j - 1,i]
  }
}

# Get rid of NA (last) row
constant.sims.fdv.5 <- constant.sims.fdv.5[1:114, ]
warming.sims.fdv.5 <- warming.sims.fdv.5[1:114, ]
```

### Autocorrelation and Autocovariance

```{r}
constant.sims.ac.5 <- matrix(ncol = num.of.sims, nrow = 120)
constant.sims.acv.5 <- matrix(ncol = num.of.sims, nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:115){
    window.start <- i
    window.end <- i + 4
    
    segment <- smooth.sim.const.b3[c(window.start:window.end),j]
    
    constant.sims.ac.5[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    constant.sims.acv.5[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}

warming.sims.ac.5 <- matrix(ncol=num.of.sims,nrow = 120)
warming.sims.acv.5 <- matrix(ncol=num.of.sims,nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:115){
    window.start <- i
    window.end <- i + 4
    
    segment <- smooth.sim.warm.b3[c(window.start:window.end),j]
    
    warming.sims.ac.5[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    warming.sims.acv.5[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}
```

### Decay Time

```{r}
constant.sims.dt.5 <- matrix(ncol = num.of.sims, nrow = 55)
warming.sims.dt.5 <- matrix(nco = num.of.sims, nrow = 55)

for(i in 1:num.of.sims){
  for(j in 1:55){
    suppressWarnings(constant.sims.dt.5[j, i] <- -j/log(constant.sims.ac.5[j, i]))
    suppressWarnings(warming.sims.dt.5[j, i] <- -j/log(warming.sims.ac.5[j, i]))
  }
}
```

## 15 Day Sliding Window

### Standard Deviation, Mean, Skewness, Kurtosis

```{r}
num.of.sims <- 10^3

# Making empty matricies
constant.sims.sd.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.mean.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.sk.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.kt.15 <- matrix(nrow = 105, ncol = num.of.sims)

warming.sims.sd.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.mean.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.sk.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.kt.15 <- matrix(nrow = 105, ncol = num.of.sims)

# Use SlidingWindow to calculate sd, mean, skewness, and kurtosis
for(i in 1:num.of.sims){
  constant.sims.sd.15[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.const.b3[,i], window = 15,step = 1)
  constant.sims.mean.15[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.const.b3[,i], window = 15, step = 1)
  constant.sims.sk.15[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.const.b3[,i], window = 15, step = 1)
  constant.sims.kt.15[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.const.b3[,i], window = 15, step = 1)
}

for(i in 1:num.of.sims){
  warming.sims.sd.15[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.warm.b3[,i], window = 15,step = 1)
  warming.sims.mean.15[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.warm.b3[,i], window = 15, step = 1)
  warming.sims.sk.15[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.warm.b3[,i], window = 15, step = 1)
  warming.sims.kt.15[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.warm.b3[,i], window = 15, step = 1)
}
```

### Variation

```{r}
constant.sims.var.15 <- constant.sims.sd.15*constant.sims.sd.15
warming.sims.var.15 <- warming.sims.sd.15*warming.sims.sd.15
```

### Co-Efficient of Variation

```{r}
constant.sims.cv.15 <- constant.sims.sd.15/constant.sims.mean.15
warming.sims.cv.15 <- warming.sims.sd.15/warming.sims.mean.15
```

### Index of Dispersion

```{r}
constant.sims.id.15 <- constant.sims.var.15/constant.sims.mean.15
warming.sims.id.15 <- warming.sims.var.15/warming.sims.mean.15
```

### First Difference Variance

```{r}
constant.sims.fdv.15 <- matrix(ncol = num.of.sims, nrow = 105)
warming.sims.fdv.15 <- matrix(ncol = num.of.sims, nrow = 105)

for(i in 1:num.of.sims){
  for(j in 2:105){
    constant.sims.fdv.15[j - 1,i] <- constant.sims.var.15[j,i] - constant.sims.var.15[j - 1,i]
    warming.sims.fdv.15[j - 1,i] <- warming.sims.var.15[j,i] - warming.sims.var.15[j - 1,i]
  }
}

# Get rid of NA (last) row
constant.sims.fdv.15 <- constant.sims.fdv.15[1:104, ]
warming.sims.fdv.15 <- warming.sims.fdv.15[1:104, ]
```

### Autocorrelation and Autocovariance

```{r}
constant.sims.ac.15 <- matrix(ncol = num.of.sims, nrow = 120)
constant.sims.acv.15 <- matrix(ncol = num.of.sims, nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:105){
    window.start <- i
    window.end <- i + 14
    
    segment <- smooth.sim.const.b3[c(window.start:window.end),j]
    
    constant.sims.ac.15[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    constant.sims.acv.15[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}

warming.sims.ac.15 <- matrix(ncol=num.of.sims,nrow = 120)
warming.sims.acv.15 <- matrix(ncol=num.of.sims,nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:105){
    window.start <- i
    window.end <- i + 14
    
    segment <- smooth.sim.warm.b3[c(window.start:window.end),j]
    
    warming.sims.ac.15[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    warming.sims.acv.15[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}
```

### Decay Time

```{r}
constant.sims.dt.15 <- matrix(ncol = num.of.sims, nrow = 45)
warming.sims.dt.15 <- matrix(nco = num.of.sims, nrow = 45)

for(i in 1:num.of.sims){
  for(j in 1:45){
    suppressWarnings(constant.sims.dt.15[j, i] <- -j/log(constant.sims.ac.15[j, i]))
    suppressWarnings(warming.sims.dt.15[j, i] <- -j/log(warming.sims.ac.15[j, i]))
  }
}
```

## 30 Day Sliding Window

### Standard Deviation, Mean, Skewness, Kurtosis

```{r}
num.of.sims <- 10^3

# Making empty matricies
constant.sims.sd.30 <- matrix(nrow = 90, ncol = num.of.sims)
constant.sims.mean.30 <- matrix(nrow = 90, ncol = num.of.sims)
constant.sims.sk.30 <- matrix(nrow = 90, ncol = num.of.sims)
constant.sims.kt.30 <- matrix(nrow = 90, ncol = num.of.sims)

warming.sims.sd.30 <- matrix(nrow = 90, ncol = num.of.sims)
warming.sims.mean.30 <- matrix(nrow = 90, ncol = num.of.sims)
warming.sims.sk.30 <- matrix(nrow = 90, ncol = num.of.sims)
warming.sims.kt.30 <- matrix(nrow = 90, ncol = num.of.sims)

# Use SlidingWindow to calculate sd, mean, skewness, and kurtosis
for(i in 1:num.of.sims){
  constant.sims.sd.30[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.const.b3[,i], window = 30,step = 1)
  constant.sims.mean.30[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.const.b3[,i], window = 30, step = 1)
  constant.sims.sk.30[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.const.b3[,i], window = 30, step = 1)
  constant.sims.kt.30[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.const.b3[,i], window = 30, step = 1)
}

for(i in 1:num.of.sims){
  warming.sims.sd.30[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.warm.b3[,i], window = 30,step = 1)
  warming.sims.mean.30[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.warm.b3[,i], window = 30, step = 1)
  warming.sims.sk.30[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.warm.b3[,i], window = 30, step = 1)
  warming.sims.kt.30[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.warm.b3[,i], window = 30, step = 1)
}
```

### Variation

```{r}
constant.sims.var.30 <- constant.sims.sd.30*constant.sims.sd.30
warming.sims.var.30 <- warming.sims.sd.30*warming.sims.sd.30
```

### Co-Efficient of Variation

```{r}
constant.sims.cv.30 <- constant.sims.sd.30/constant.sims.mean.30
warming.sims.cv.30 <- warming.sims.sd.30/warming.sims.mean.30
```

### Index of Dispersion

```{r}
constant.sims.id.30 <- constant.sims.var.30/constant.sims.mean.30
warming.sims.id.30 <- warming.sims.var.30/warming.sims.mean.30
```

### First Difference Variance

```{r}
constant.sims.fdv.30 <- matrix(ncol = num.of.sims, nrow = 90)
warming.sims.fdv.30 <- matrix(ncol = num.of.sims, nrow = 90)

for(i in 1:num.of.sims){
  for(j in 2:90){
    constant.sims.fdv.30[j - 1,i] <- constant.sims.var.30[j,i] - constant.sims.var.30[j - 1,i]
    warming.sims.fdv.30[j - 1,i] <- warming.sims.var.30[j,i] - warming.sims.var.30[j - 1,i]
  }
}

# Get rid of NA (last) row
constant.sims.fdv.30 <- constant.sims.fdv.30[1:89, ]
warming.sims.fdv.30 <- warming.sims.fdv.30[1:89, ]
```

### Autocorrelation and Autocovariance

```{r}
constant.sims.ac.30 <- matrix(ncol = num.of.sims, nrow = 120)
constant.sims.acv.30 <- matrix(ncol = num.of.sims, nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:90){
    window.start <- i
    window.end <- i + 29
    
    segment <- smooth.sim.const.b3[c(window.start:window.end),j]
    
    constant.sims.ac.30[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    constant.sims.acv.30[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}

warming.sims.ac.30 <- matrix(ncol=num.of.sims,nrow = 120)
warming.sims.acv.30 <- matrix(ncol=num.of.sims,nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:90){
    window.start <- i
    window.end <- i + 29
    
    segment <- smooth.sim.warm.b3[c(window.start:window.end),j]
    
    warming.sims.ac.30[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    warming.sims.acv.30[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}
```

### Decay Time

```{r}
constant.sims.dt.30 <- matrix(ncol = num.of.sims, nrow = 30)
warming.sims.dt.30 <- matrix(nco = num.of.sims, nrow = 30)

for(i in 1:num.of.sims){
  for(j in 1:30){
    suppressWarnings(constant.sims.dt.30[j, i] <- -j/log(constant.sims.ac.30[j, i]))
    suppressWarnings(warming.sims.dt.30[j, i] <- -j/log(warming.sims.ac.30[j, i]))
  }
}
```

# Area Under Curve (AUC) Tests

Function to calculate AUC:

```{r}
calc_auc <- function(test.taus, null.taus){
  
  combined.taus <- c(test.taus, null.taus)
  r <- rank(combined.taus, na.last = NA)
  
  test.taus.length <- sum(!is.na(test.taus))
  null.taus.length <- sum(!is.na(null.taus))
  
  r1 <- sum(r[1:test.taus.length])
  
  n1 <- test.taus.length
  n2 <- null.taus.length
  
  (r1 - n1 * (n1 + 1) / 2) / (n1 * n2)
}
```

Metrics to evaluate:

|                           |                  |
|---------------------------|------------------|
| Standard Deviation        | `sd`             |
| Mean                      | `mean`           |
| Skew                      | `skewness`       |
| Kurtosis                  | `kurtosis`       |
| Variance                  | `variance`       |
| Co-Efficient of Variance  | `cv`             |
| Index of Dispersion       | `id`             |
| First Difference Variance | `first.diff.var` |
| Autocorrelation           | `cor`            |
| Autocovariance            | `cov`            |
| Decay Time                | `decay.time`     |

## 5 Day Sliding Window

```{r}
constant.sims.var.5 <- constant.sims.var.5[1:55, ]
constant.sims.mean.5 <- constant.sims.mean.5[1:55, ]
constant.sims.cv.5 <- constant.sims.cv.5[1:55, ]
constant.sims.sk.5 <-  constant.sims.sk.5[1:55, ]
constant.sims.kt.5 <-  constant.sims.kt.5[1:55, ]
constant.sims.id.5 <-  constant.sims.id.5[1:55, ]
constant.sims.ac.5 <-  constant.sims.ac.5[1:55, ]
constant.sims.acv.5 <-  constant.sims.acv.5[1:55, ]
constant.sims.fdv.5 <-  constant.sims.fdv.5[1:55, ]
constant.sims.dt.5 <-  constant.sims.dt.5[1:55, ]

warming.sims.var.5 <- warming.sims.var.5[1:55, ]
warming.sims.mean.5 <- warming.sims.mean.5[1:55, ]
warming.sims.cv.5 <- warming.sims.cv.5[1:55, ]
warming.sims.sk.5 <-  warming.sims.sk.5[1:55, ]
warming.sims.kt.5 <-  warming.sims.kt.5[1:55, ]
warming.sims.id.5 <-  warming.sims.id.5[1:55, ]
warming.sims.ac.5 <-  warming.sims.ac.5[1:55, ]
warming.sims.acv.5 <-  warming.sims.acv.5[1:55, ]
warming.sims.fdv.5 <-  warming.sims.fdv.5[1:55, ]
warming.sims.dt.5 <-  warming.sims.dt.5[1:55, ]

constant.sims.var.5.tau <- c()
constant.sims.mean.5.tau <- c()
constant.sims.cv.5.tau <- c()
constant.sims.sk.5.tau <-  c()
constant.sims.kt.5.tau <- c()
constant.sims.id.5.tau <- c()
constant.sims.ac.5.tau <- c()
constant.sims.acv.5.tau <- c()
constant.sims.fdv.5.tau <-  c()
constant.sims.dt.5.tau <- c()

warming.sims.var.5.tau <- c()
warming.sims.mean.5.tau <- c()
warming.sims.cv.5.tau <- c()
warming.sims.sk.5.tau <-  c()
warming.sims.kt.5.tau <- c()
warming.sims.id.5.tau <-  c()
warming.sims.ac.5.tau <-  c()
warming.sims.acv.5.tau <-  c()
warming.sims.fdv.5.tau <-  c()
warming.sims.dt.5.tau <-  c()
```

Calculate Kendall's tau for each metric:

```{r}
time.5 <- c(1:55)

for(i in 1:num.of.sims){
  constant.sims.mean.5.tau[i] <- cor(constant.sims.mean.5[ , i], time.5, method = "kendall")
  constant.sims.sk.5.tau[i] <- cor(constant.sims.sk.5[ , i], time.5, method = "kendall")
  constant.sims.kt.5.tau[i] <- cor(constant.sims.kt.5[ , i], time.5, method = "kendall")
  constant.sims.var.5.tau[i] <- cor(constant.sims.var.5[ , i], time.5, method = "kendall")
  constant.sims.cv.5.tau[i] <- cor(constant.sims.cv.5[ , i], time.5, method = "kendall")
  constant.sims.id.5.tau[i] <- cor(constant.sims.id.5[ , i], time.5, method = "kendall")
  constant.sims.fdv.5.tau[i] <- cor(constant.sims.fdv.5[ , i], time.5, method = "kendall")
  constant.sims.ac.5.tau[i] <- cor(constant.sims.ac.5[ , i], time.5, method = "kendall")
  constant.sims.acv.5.tau[i] <- cor(constant.sims.acv.5[ , i], time.5, method = "kendall")
  constant.sims.dt.5.tau[i] <- cor(constant.sims.dt.5[ , i], time.5, method = "kendall")
  
  warming.sims.mean.5.tau[i] <- cor(warming.sims.mean.5[ , i], time.5, method = "kendall")
  warming.sims.sk.5.tau[i] <- cor(warming.sims.sk.5[ , i], time.5, method = "kendall")
  warming.sims.kt.5.tau[i] <- cor(warming.sims.kt.5[ , i], time.5, method = "kendall")
  warming.sims.var.5.tau[i] <- cor(warming.sims.var.5[ , i], time.5, method = "kendall")
  warming.sims.cv.5.tau[i] <- cor(warming.sims.cv.5[ , i], time.5, method = "kendall")
  warming.sims.id.5.tau[i] <- cor(warming.sims.id.5[ , i], time.5, method = "kendall")
  warming.sims.fdv.5.tau[i] <- cor(warming.sims.fdv.5[ , i], time.5, method = "kendall")
  warming.sims.ac.5.tau[i] <- cor(warming.sims.ac.5[ , i], time.5, method = "kendall")
  warming.sims.acv.5.tau[i] <- cor(warming.sims.acv.5[ , i], time.5, method = "kendall")
  warming.sims.dt.5.tau[i] <- cor(warming.sims.dt.5[ , i], time.5, method = "kendall")
}
```

Calculating AUC:

```{r}
AUC_var_5 <- calc_auc(warming.sims.var.5.tau, constant.sims.var.5.tau); AUC_var_5
AUC_mean_5 <- calc_auc(warming.sims.mean.5.tau, constant.sims.mean.5.tau); AUC_mean_5
AUC_kt_5 <- calc_auc(warming.sims.kt.5.tau, constant.sims.kt.5.tau); AUC_kt_5
AUC_sk_5 <- calc_auc(warming.sims.sk.5.tau, constant.sims.sk.5.tau); AUC_sk_5
AUC_id_5 <- calc_auc(warming.sims.id.5.tau, constant.sims.id.5.tau); AUC_id_5
AUC_fdv_5 <- calc_auc(warming.sims.fdv.5.tau, constant.sims.fdv.5.tau); AUC_fdv_5
AUC_ac_5 <- calc_auc(warming.sims.ac.5.tau, constant.sims.ac.5.tau); AUC_ac_5
AUC_acv_5 <- calc_auc(warming.sims.acv.5.tau, constant.sims.acv.5.tau); AUC_acv_5
AUC_cv_5 <- calc_auc(warming.sims.cv.5.tau, constant.sims.cv.5.tau); AUC_cv_5
AUC_dt_5 <- calc_auc(warming.sims.dt.5.tau, constant.sims.dt.5.tau); AUC_dt_5
```

## 15 Day Sliding Window

```{r}
constant.sims.var.15 <- constant.sims.var.15[1:45, ]
constant.sims.mean.15 <- constant.sims.mean.15[1:45, ]
constant.sims.cv.15 <- constant.sims.cv.15[1:45, ]
constant.sims.sk.15 <-  constant.sims.sk.15[1:45, ]
constant.sims.kt.15 <-  constant.sims.kt.15[1:45, ]
constant.sims.id.15 <-  constant.sims.id.15[1:45, ]
constant.sims.ac.15 <-  constant.sims.ac.15[1:45, ]
constant.sims.acv.15 <-  constant.sims.acv.15[1:45, ]
constant.sims.fdv.15 <-  constant.sims.fdv.15[1:45, ]
constant.sims.dt.15 <-  constant.sims.dt.15[1:45, ]

warming.sims.var.15 <- warming.sims.var.15[1:45, ]
warming.sims.mean.15 <- warming.sims.mean.15[1:45, ]
warming.sims.cv.15 <- warming.sims.cv.15[1:45, ]
warming.sims.sk.15 <-  warming.sims.sk.15[1:45, ]
warming.sims.kt.15 <-  warming.sims.kt.15[1:45, ]
warming.sims.id.15 <-  warming.sims.id.15[1:45, ]
warming.sims.ac.15 <-  warming.sims.ac.15[1:45, ]
warming.sims.acv.15 <-  warming.sims.acv.15[1:45, ]
warming.sims.fdv.15 <-  warming.sims.fdv.15[1:45, ]
warming.sims.dt.15 <-  warming.sims.dt.15[1:45, ]

constant.sims.var.15.tau <- c()
constant.sims.mean.15.tau <- c()
constant.sims.cv.15.tau <- c()
constant.sims.sk.15.tau <-  c()
constant.sims.kt.15.tau <- c()
constant.sims.id.15.tau <- c()
constant.sims.ac.15.tau <- c()
constant.sims.acv.15.tau <- c()
constant.sims.fdv.15.tau <-  c()
constant.sims.dt.15.tau <- c()

warming.sims.var.15.tau <- c()
warming.sims.mean.15.tau <- c()
warming.sims.cv.15.tau <- c()
warming.sims.sk.15.tau <-  c()
warming.sims.kt.15.tau <- c()
warming.sims.id.15.tau <-  c()
warming.sims.ac.15.tau <-  c()
warming.sims.acv.15.tau <-  c()
warming.sims.fdv.15.tau <-  c()
warming.sims.dt.15.tau <-  c()
```

Calculate Kendall's tau for each metric:

```{r}
time.15 <- c(1:45)

for(i in 1:num.of.sims){
  constant.sims.mean.15.tau[i] <- cor(constant.sims.mean.15[ , i], time.15, method = "kendall")
  constant.sims.sk.15.tau[i] <- cor(constant.sims.sk.15[ , i], time.15, method = "kendall")
  constant.sims.kt.15.tau[i] <- cor(constant.sims.kt.15[ , i], time.15, method = "kendall")
  constant.sims.var.15.tau[i] <- cor(constant.sims.var.15[ , i], time.15, method = "kendall")
  constant.sims.cv.15.tau[i] <- cor(constant.sims.cv.15[ , i], time.15, method = "kendall")
  constant.sims.id.15.tau[i] <- cor(constant.sims.id.15[ , i], time.15, method = "kendall")
  constant.sims.fdv.15.tau[i] <- cor(constant.sims.fdv.15[ , i], time.15, method = "kendall")
  constant.sims.ac.15.tau[i] <- cor(constant.sims.ac.15[ , i], time.15, method = "kendall")
  constant.sims.acv.15.tau[i] <- cor(constant.sims.acv.15[ , i], time.15, method = "kendall")
  constant.sims.dt.15.tau[i] <- cor(constant.sims.dt.15[ , i], time.15, method = "kendall")
  
  warming.sims.mean.15.tau[i] <- cor(warming.sims.mean.15[ , i], time.15, method = "kendall")
  warming.sims.sk.15.tau[i] <- cor(warming.sims.sk.15[ , i], time.15, method = "kendall")
  warming.sims.kt.15.tau[i] <- cor(warming.sims.kt.15[ , i], time.15, method = "kendall")
  warming.sims.var.15.tau[i] <- cor(warming.sims.var.15[ , i], time.15, method = "kendall")
  warming.sims.cv.15.tau[i] <- cor(warming.sims.cv.15[ , i], time.15, method = "kendall")
  warming.sims.id.15.tau[i] <- cor(warming.sims.id.15[ , i], time.15, method = "kendall")
  warming.sims.fdv.15.tau[i] <- cor(warming.sims.fdv.15[ , i], time.15, method = "kendall")
  warming.sims.ac.15.tau[i] <- cor(warming.sims.ac.15[ , i], time.15, method = "kendall")
  warming.sims.acv.15.tau[i] <- cor(warming.sims.acv.15[ , i], time.15, method = "kendall")
  warming.sims.dt.15.tau[i] <- cor(warming.sims.dt.15[ , i], time.15, method = "kendall")
}
```

Calculating AUC:

```{r}
AUC_var_15 <- calc_auc(warming.sims.var.15.tau, constant.sims.var.15.tau); AUC_var_15
AUC_mean_15 <- calc_auc(warming.sims.mean.15.tau, constant.sims.mean.15.tau); AUC_mean_15
AUC_kt_15 <- calc_auc(warming.sims.kt.15.tau, constant.sims.kt.15.tau); AUC_kt_15
AUC_sk_15 <- calc_auc(warming.sims.sk.15.tau, constant.sims.sk.15.tau); AUC_sk_15
AUC_id_15 <- calc_auc(warming.sims.id.15.tau, constant.sims.id.15.tau); AUC_id_15
AUC_fdv_15 <- calc_auc(warming.sims.fdv.15.tau, constant.sims.fdv.15.tau); AUC_fdv_15
AUC_ac_15 <- calc_auc(warming.sims.ac.15.tau, constant.sims.ac.15.tau); AUC_ac_15
AUC_acv_15 <- calc_auc(warming.sims.acv.15.tau, constant.sims.acv.15.tau); AUC_acv_15
AUC_cv_15 <- calc_auc(warming.sims.cv.15.tau, constant.sims.cv.15.tau); AUC_cv_15
AUC_dt_15 <- calc_auc(warming.sims.dt.15.tau, constant.sims.dt.15.tau); AUC_dt_15
```

## 30 Day Sliding Window

```{r}
constant.sims.var.30 <- constant.sims.var.30[1:30, ]
constant.sims.mean.30 <- constant.sims.mean.30[1:30, ]
constant.sims.cv.30 <- constant.sims.cv.30[1:30, ]
constant.sims.sk.30 <-  constant.sims.sk.30[1:30, ]
constant.sims.kt.30 <-  constant.sims.kt.30[1:30, ]
constant.sims.id.30 <-  constant.sims.id.30[1:30, ]
constant.sims.ac.30 <-  constant.sims.ac.30[1:30, ]
constant.sims.acv.30 <-  constant.sims.acv.30[1:30, ]
constant.sims.fdv.30 <-  constant.sims.fdv.30[1:30, ]
constant.sims.dt.30 <-  constant.sims.dt.30[1:30, ]

warming.sims.var.30 <- warming.sims.var.30[1:30, ]
warming.sims.mean.30 <- warming.sims.mean.30[1:30, ]
warming.sims.cv.30 <- warming.sims.cv.30[1:30, ]
warming.sims.sk.30 <-  warming.sims.sk.30[1:30, ]
warming.sims.kt.30 <-  warming.sims.kt.30[1:30, ]
warming.sims.id.30 <-  warming.sims.id.30[1:30, ]
warming.sims.ac.30 <-  warming.sims.ac.30[1:30, ]
warming.sims.acv.30 <-  warming.sims.acv.30[1:30, ]
warming.sims.fdv.30 <-  warming.sims.fdv.30[1:30, ]
warming.sims.dt.30 <-  warming.sims.dt.30[1:30, ]

constant.sims.var.30.tau <- c()
constant.sims.mean.30.tau <- c()
constant.sims.cv.30.tau <- c()
constant.sims.sk.30.tau <-  c()
constant.sims.kt.30.tau <- c()
constant.sims.id.30.tau <- c()
constant.sims.ac.30.tau <- c()
constant.sims.acv.30.tau <- c()
constant.sims.fdv.30.tau <-  c()
constant.sims.dt.30.tau <- c()

warming.sims.var.30.tau <- c()
warming.sims.mean.30.tau <- c()
warming.sims.cv.30.tau <- c()
warming.sims.sk.30.tau <-  c()
warming.sims.kt.30.tau <- c()
warming.sims.id.30.tau <-  c()
warming.sims.ac.30.tau <-  c()
warming.sims.acv.30.tau <-  c()
warming.sims.fdv.30.tau <-  c()
warming.sims.dt.30.tau <-  c()
```

Calculate Kendall's tau for each metric:

```{r}
time.30 <- c(1:30)

for(i in 1:num.of.sims){
  constant.sims.mean.30.tau[i] <- cor(constant.sims.mean.30[ , i], time.30, method = "kendall")
  constant.sims.sk.30.tau[i] <- cor(constant.sims.sk.30[ , i], time.30, method = "kendall")
  constant.sims.kt.30.tau[i] <- cor(constant.sims.kt.30[ , i], time.30, method = "kendall")
  constant.sims.var.30.tau[i] <- cor(constant.sims.var.30[ , i], time.30, method = "kendall")
  constant.sims.cv.30.tau[i] <- cor(constant.sims.cv.30[ , i], time.30, method = "kendall")
  constant.sims.id.30.tau[i] <- cor(constant.sims.id.30[ , i], time.30, method = "kendall")
  constant.sims.fdv.30.tau[i] <- cor(constant.sims.fdv.30[ , i], time.30, method = "kendall")
  constant.sims.ac.30.tau[i] <- cor(constant.sims.ac.30[ , i], time.30, method = "kendall")
  constant.sims.acv.30.tau[i] <- cor(constant.sims.acv.30[ , i], time.30, method = "kendall")
  constant.sims.dt.30.tau[i] <- cor(constant.sims.dt.30[ , i], time.30, method = "kendall")
  
  warming.sims.mean.30.tau[i] <- cor(warming.sims.mean.30[ , i], time.30, method = "kendall")
  warming.sims.sk.30.tau[i] <- cor(warming.sims.sk.30[ , i], time.30, method = "kendall")
  warming.sims.kt.30.tau[i] <- cor(warming.sims.kt.30[ , i], time.30, method = "kendall")
  warming.sims.var.30.tau[i] <- cor(warming.sims.var.30[ , i], time.30, method = "kendall")
  warming.sims.cv.30.tau[i] <- cor(warming.sims.cv.30[ , i], time.30, method = "kendall")
  warming.sims.id.30.tau[i] <- cor(warming.sims.id.30[ , i], time.30, method = "kendall")
  warming.sims.fdv.30.tau[i] <- cor(warming.sims.fdv.30[ , i], time.30, method = "kendall")
  warming.sims.ac.30.tau[i] <- cor(warming.sims.ac.30[ , i], time.30, method = "kendall")
  warming.sims.acv.30.tau[i] <- cor(warming.sims.acv.30[ , i], time.30, method = "kendall")
  warming.sims.dt.30.tau[i] <- cor(warming.sims.dt.30[ , i], time.30, method = "kendall")
}
```

Calculating AUC:

```{r}
AUC_var_30 <- calc_auc(warming.sims.var.30.tau, constant.sims.var.30.tau); AUC_var_30
AUC_mean_30 <- calc_auc(warming.sims.mean.30.tau, constant.sims.mean.30.tau); AUC_mean_30
AUC_kt_30 <- calc_auc(warming.sims.kt.30.tau, constant.sims.kt.30.tau); AUC_kt_30
AUC_sk_30 <- calc_auc(warming.sims.sk.30.tau, constant.sims.sk.30.tau); AUC_sk_30
AUC_id_30 <- calc_auc(warming.sims.id.30.tau, constant.sims.id.30.tau); AUC_id_30
AUC_fdv_30 <- calc_auc(warming.sims.fdv.30.tau, constant.sims.fdv.30.tau); AUC_fdv_30
AUC_ac_30 <- calc_auc(warming.sims.ac.30.tau, constant.sims.ac.30.tau); AUC_ac_30
AUC_acv_30 <- calc_auc(warming.sims.acv.30.tau, constant.sims.acv.30.tau); AUC_acv_30
AUC_cv_30 <- calc_auc(warming.sims.cv.30.tau, constant.sims.cv.30.tau); AUC_cv_30
AUC_dt_30 <- calc_auc(warming.sims.dt.30.tau, constant.sims.dt.30.tau); AUC_dt_30
```

## Summarize Outputs

```{r}
metrics <- c("mean", "skewness", "kurtosis", "variance", "cv", "id", "first.diff.var", "cor", "cov", "decay.time")

AUC_5 <- c(AUC_mean_5, AUC_sk_5, AUC_kt_5, AUC_var_5, AUC_cv_5, 
           AUC_id_5, AUC_fdv_5, AUC_ac_5 ,AUC_acv_5, AUC_dt_5)
AUC_15 <- c(AUC_mean_15, AUC_sk_15, AUC_kt_15, AUC_var_15, AUC_cv_15, 
           AUC_id_15, AUC_fdv_15, AUC_ac_15 ,AUC_acv_15, AUC_dt_15)
AUC_30 <- c(AUC_mean_30, AUC_sk_30, AUC_kt_30, AUC_var_30, AUC_cv_30, 
           AUC_id_30, AUC_fdv_30, AUC_ac_30 ,AUC_acv_30, AUC_dt_30)
AUC_df <- data.frame(metrics, AUC_5, AUC_15, AUC_30); AUC_df
saveRDS(AUC_df, "Objects/AUC_SmoothSimulatedData_B3.rds")
```

A quick plot!

```{r}
metrics_full <- c("Mean", "Skewness", "Kurtosis", "Variance", "Covariance", "Index of Dispersion", "First Difference Variance", "Autocorrelation", "Autocovariance", "Decay Time")

## Plot 
plot(1:10, AUC_df$AUC_5, type = "p", 
     main = "Smoothed Simulated Data | Bandwidth = 3", xaxt = "none", xlab = "", ylim = c(0.3, 0.7), ylab = "AUC",
     col = "forestgreen", pch = 19)
points(1:10, AUC_df$AUC_15, col = "orchid", pch = 19)
points(1:10, AUC_df$AUC_30, col = "#4B0082", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)
legend("bottomright", 
       legend = c("Sliding Window = 5 Days", "Sliding Window = 15 Days", "Sliding Window = 30 Days"),
       col = c("forestgreen", "orchid", "#4B0082"), pch = 19)

## Saving plot as .png
png(file = "Plots/AUC_SmoothSimulatedData_B3.png", width = 1500, height = 350)
plot(1:10, AUC_df$AUC_5, type = "p", 
     main = "Smoothed Simulated Data | Bandwidth = 3", xaxt = "none", xlab = "", ylim = c(0.3, 0.7), ylab = "AUC",
     col = "forestgreen", pch = 19)
points(1:10, AUC_df$AUC_15, col = "orchid", pch = 19)
points(1:10, AUC_df$AUC_30, col = "#4B0082", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full)
legend("bottomright", 
       legend = c("Sliding Window = 5 Days", "Sliding Window = 15 Days", "Sliding Window = 30 Days"),
       col = c("forestgreen", "orchid", "#4B0082"), pch = 19)
dev.off()
```

# BANDWIDTH = 2

# Calculating Statistical Metrics Along Variously-Sized Sliding Windows

For details on these metrics, see `3_EWS-Sim.Rmd`.

## 5 Day Sliding Window

### Standard Deviation, Mean, Skewness, Kurtosis

```{r}
num.of.sims <- 10^3

# Making empty matricies
constant.sims.sd.5 <- matrix(nrow = 115, ncol = num.of.sims)
constant.sims.mean.5 <- matrix(nrow = 115, ncol = num.of.sims)
constant.sims.sk.5 <- matrix(nrow = 115, ncol = num.of.sims)
constant.sims.kt.5 <- matrix(nrow = 115, ncol = num.of.sims)

warming.sims.sd.5 <- matrix(nrow = 115, ncol = num.of.sims)
warming.sims.mean.5 <- matrix(nrow = 115, ncol = num.of.sims)
warming.sims.sk.5 <- matrix(nrow = 115, ncol = num.of.sims)
warming.sims.kt.5 <- matrix(nrow = 115, ncol = num.of.sims)

# Use SlidingWindow to calculate sd, mean, skewness, and kurtosis
for(i in 1:num.of.sims){
  constant.sims.sd.5[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.const.b2[,i], window = 5,step = 1)
  constant.sims.mean.5[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.const.b2[,i], window = 5, step = 1)
  constant.sims.sk.5[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.const.b2[,i], window = 5, step = 1)
  constant.sims.kt.5[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.const.b2[,i], window = 5, step = 1)
}

for(i in 1:num.of.sims){
  warming.sims.sd.5[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.warm.b2[,i], window = 5,step = 1)
  warming.sims.mean.5[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.warm.b2[,i], window = 5, step = 1)
  warming.sims.sk.5[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.warm.b2[,i], window = 5, step = 1)
  warming.sims.kt.5[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.warm.b2[,i], window = 5, step = 1)
}
```

### Variation

```{r}
constant.sims.var.5 <- constant.sims.sd.5*constant.sims.sd.5
warming.sims.var.5 <- warming.sims.sd.5*warming.sims.sd.5
```

### Co-Efficient of Variation

```{r}
constant.sims.cv.5 <- constant.sims.sd.5/constant.sims.mean.5
warming.sims.cv.5 <- warming.sims.sd.5/warming.sims.mean.5
```

### Index of Dispersion

```{r}
constant.sims.id.5 <- constant.sims.var.5/constant.sims.mean.5
warming.sims.id.5 <- warming.sims.var.5/warming.sims.mean.5
```

### First Difference Variance

```{r}
constant.sims.fdv.5 <- matrix(ncol = num.of.sims, nrow = 115)
warming.sims.fdv.5 <- matrix(ncol = num.of.sims, nrow = 115)

for(i in 1:num.of.sims){
  for(j in 2:115){
    constant.sims.fdv.5[j - 1,i] <- constant.sims.var.5[j,i] - constant.sims.var.5[j - 1,i]
    warming.sims.fdv.5[j - 1,i] <- warming.sims.var.5[j,i] - warming.sims.var.5[j - 1,i]
  }
}

# Get rid of NA (last) row
constant.sims.fdv.5 <- constant.sims.fdv.5[1:114, ]
warming.sims.fdv.5 <- warming.sims.fdv.5[1:114, ]
```

### Autocorrelation and Autocovariance

```{r}
constant.sims.ac.5 <- matrix(ncol = num.of.sims, nrow = 120)
constant.sims.acv.5 <- matrix(ncol = num.of.sims, nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:115){
    window.start <- i
    window.end <- i + 4
    
    segment <- smooth.sim.const.b2[c(window.start:window.end),j]
    
    constant.sims.ac.5[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    constant.sims.acv.5[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}

warming.sims.ac.5 <- matrix(ncol=num.of.sims,nrow = 120)
warming.sims.acv.5 <- matrix(ncol=num.of.sims,nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:115){
    window.start <- i
    window.end <- i + 4
    
    segment <- smooth.sim.warm.b2[c(window.start:window.end),j]
    
    warming.sims.ac.5[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    warming.sims.acv.5[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}
```

### Decay Time

```{r}
constant.sims.dt.5 <- matrix(ncol = num.of.sims, nrow = 55)
warming.sims.dt.5 <- matrix(nco = num.of.sims, nrow = 55)

for(i in 1:num.of.sims){
  for(j in 1:55){
    suppressWarnings(constant.sims.dt.5[j, i] <- -j/log(constant.sims.ac.5[j, i]))
    suppressWarnings(warming.sims.dt.5[j, i] <- -j/log(warming.sims.ac.5[j, i]))
  }
}
```

## 15 Day Sliding Window

### Standard Deviation, Mean, Skewness, Kurtosis

```{r}
num.of.sims <- 10^3

# Making empty matricies
constant.sims.sd.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.mean.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.sk.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.kt.15 <- matrix(nrow = 105, ncol = num.of.sims)

warming.sims.sd.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.mean.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.sk.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.kt.15 <- matrix(nrow = 105, ncol = num.of.sims)

# Use SlidingWindow to calculate sd, mean, skewness, and kurtosis
for(i in 1:num.of.sims){
  constant.sims.sd.15[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.const.b2[,i], window = 15,step = 1)
  constant.sims.mean.15[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.const.b2[,i], window = 15, step = 1)
  constant.sims.sk.15[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.const.b2[,i], window = 15, step = 1)
  constant.sims.kt.15[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.const.b2[,i], window = 15, step = 1)
}

for(i in 1:num.of.sims){
  warming.sims.sd.15[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.warm.b2[,i], window = 15,step = 1)
  warming.sims.mean.15[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.warm.b2[,i], window = 15, step = 1)
  warming.sims.sk.15[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.warm.b2[,i], window = 15, step = 1)
  warming.sims.kt.15[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.warm.b2[,i], window = 15, step = 1)
}
```

### Variation

```{r}
constant.sims.var.15 <- constant.sims.sd.15*constant.sims.sd.15
warming.sims.var.15 <- warming.sims.sd.15*warming.sims.sd.15
```

### Co-Efficient of Variation

```{r}
constant.sims.cv.15 <- constant.sims.sd.15/constant.sims.mean.15
warming.sims.cv.15 <- warming.sims.sd.15/warming.sims.mean.15
```

### Index of Dispersion

```{r}
constant.sims.id.15 <- constant.sims.var.15/constant.sims.mean.15
warming.sims.id.15 <- warming.sims.var.15/warming.sims.mean.15
```

### First Difference Variance

```{r}
constant.sims.fdv.15 <- matrix(ncol = num.of.sims, nrow = 105)
warming.sims.fdv.15 <- matrix(ncol = num.of.sims, nrow = 105)

for(i in 1:num.of.sims){
  for(j in 2:105){
    constant.sims.fdv.15[j - 1,i] <- constant.sims.var.15[j,i] - constant.sims.var.15[j - 1,i]
    warming.sims.fdv.15[j - 1,i] <- warming.sims.var.15[j,i] - warming.sims.var.15[j - 1,i]
  }
}

# Get rid of NA (last) row
constant.sims.fdv.15 <- constant.sims.fdv.15[1:104, ]
warming.sims.fdv.15 <- warming.sims.fdv.15[1:104, ]
```

### Autocorrelation and Autocovariance

```{r}
constant.sims.ac.15 <- matrix(ncol = num.of.sims, nrow = 120)
constant.sims.acv.15 <- matrix(ncol = num.of.sims, nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:105){
    window.start <- i
    window.end <- i + 14
    
    segment <- smooth.sim.const.b2[c(window.start:window.end),j]
    
    constant.sims.ac.15[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    constant.sims.acv.15[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}

warming.sims.ac.15 <- matrix(ncol=num.of.sims,nrow = 120)
warming.sims.acv.15 <- matrix(ncol=num.of.sims,nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:105){
    window.start <- i
    window.end <- i + 14
    
    segment <- smooth.sim.warm.b2[c(window.start:window.end),j]
    
    warming.sims.ac.15[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    warming.sims.acv.15[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}
```

### Decay Time

```{r}
constant.sims.dt.15 <- matrix(ncol = num.of.sims, nrow = 45)
warming.sims.dt.15 <- matrix(nco = num.of.sims, nrow = 45)

for(i in 1:num.of.sims){
  for(j in 1:45){
    suppressWarnings(constant.sims.dt.15[j, i] <- -j/log(constant.sims.ac.15[j, i]))
    suppressWarnings(warming.sims.dt.15[j, i] <- -j/log(warming.sims.ac.15[j, i]))
  }
}
```

## 30 Day Sliding Window

### Standard Deviation, Mean, Skewness, Kurtosis

```{r}
num.of.sims <- 10^3

# Making empty matricies
constant.sims.sd.30 <- matrix(nrow = 90, ncol = num.of.sims)
constant.sims.mean.30 <- matrix(nrow = 90, ncol = num.of.sims)
constant.sims.sk.30 <- matrix(nrow = 90, ncol = num.of.sims)
constant.sims.kt.30 <- matrix(nrow = 90, ncol = num.of.sims)

warming.sims.sd.30 <- matrix(nrow = 90, ncol = num.of.sims)
warming.sims.mean.30 <- matrix(nrow = 90, ncol = num.of.sims)
warming.sims.sk.30 <- matrix(nrow = 90, ncol = num.of.sims)
warming.sims.kt.30 <- matrix(nrow = 90, ncol = num.of.sims)

# Use SlidingWindow to calculate sd, mean, skewness, and kurtosis
for(i in 1:num.of.sims){
  constant.sims.sd.30[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.const.b2[,i], window = 30,step = 1)
  constant.sims.mean.30[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.const.b2[,i], window = 30, step = 1)
  constant.sims.sk.30[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.const.b2[,i], window = 30, step = 1)
  constant.sims.kt.30[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.const.b2[,i], window = 30, step = 1)
}

for(i in 1:num.of.sims){
  warming.sims.sd.30[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.warm.b2[,i], window = 30,step = 1)
  warming.sims.mean.30[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.warm.b2[,i], window = 30, step = 1)
  warming.sims.sk.30[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.warm.b2[,i], window = 30, step = 1)
  warming.sims.kt.30[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.warm.b2[,i], window = 30, step = 1)
}
```

### Variation

```{r}
constant.sims.var.30 <- constant.sims.sd.30*constant.sims.sd.30
warming.sims.var.30 <- warming.sims.sd.30*warming.sims.sd.30
```

### Co-Efficient of Variation

```{r}
constant.sims.cv.30 <- constant.sims.sd.30/constant.sims.mean.30
warming.sims.cv.30 <- warming.sims.sd.30/warming.sims.mean.30
```

### Index of Dispersion

```{r}
constant.sims.id.30 <- constant.sims.var.30/constant.sims.mean.30
warming.sims.id.30 <- warming.sims.var.30/warming.sims.mean.30
```

### First Difference Variance

```{r}
constant.sims.fdv.30 <- matrix(ncol = num.of.sims, nrow = 90)
warming.sims.fdv.30 <- matrix(ncol = num.of.sims, nrow = 90)

for(i in 1:num.of.sims){
  for(j in 2:90){
    constant.sims.fdv.30[j - 1,i] <- constant.sims.var.30[j,i] - constant.sims.var.30[j - 1,i]
    warming.sims.fdv.30[j - 1,i] <- warming.sims.var.30[j,i] - warming.sims.var.30[j - 1,i]
  }
}

# Get rid of NA (last) row
constant.sims.fdv.30 <- constant.sims.fdv.30[1:89, ]
warming.sims.fdv.30 <- warming.sims.fdv.30[1:89, ]
```

### Autocorrelation and Autocovariance

```{r}
constant.sims.ac.30 <- matrix(ncol = num.of.sims, nrow = 120)
constant.sims.acv.30 <- matrix(ncol = num.of.sims, nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:90){
    window.start <- i
    window.end <- i + 29
    
    segment <- smooth.sim.const.b2[c(window.start:window.end),j]
    
    constant.sims.ac.30[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    constant.sims.acv.30[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}

warming.sims.ac.30 <- matrix(ncol=num.of.sims,nrow = 120)
warming.sims.acv.30 <- matrix(ncol=num.of.sims,nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:90){
    window.start <- i
    window.end <- i + 29
    
    segment <- smooth.sim.warm.b2[c(window.start:window.end),j]
    
    warming.sims.ac.30[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    warming.sims.acv.30[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}
```

### Decay Time

```{r}
constant.sims.dt.30 <- matrix(ncol = num.of.sims, nrow = 30)
warming.sims.dt.30 <- matrix(nco = num.of.sims, nrow = 30)

for(i in 1:num.of.sims){
  for(j in 1:30){
    suppressWarnings(constant.sims.dt.30[j, i] <- -j/log(constant.sims.ac.30[j, i]))
    suppressWarnings(warming.sims.dt.30[j, i] <- -j/log(warming.sims.ac.30[j, i]))
  }
}
```

# Area Under Curve (AUC) Tests

Function to calculate AUC:

```{r}
calc_auc <- function(test.taus, null.taus){
  
  combined.taus <- c(test.taus, null.taus)
  r <- rank(combined.taus, na.last = NA)
  
  test.taus.length <- sum(!is.na(test.taus))
  null.taus.length <- sum(!is.na(null.taus))
  
  r1 <- sum(r[1:test.taus.length])
  
  n1 <- test.taus.length
  n2 <- null.taus.length
  
  (r1 - n1 * (n1 + 1) / 2) / (n1 * n2)
}
```

Metrics to evaluate:

|                           |                  |
|---------------------------|------------------|
| Standard Deviation        | `sd`             |
| Mean                      | `mean`           |
| Skew                      | `skewness`       |
| Kurtosis                  | `kurtosis`       |
| Variance                  | `variance`       |
| Co-Efficient of Variance  | `cv`             |
| Index of Dispersion       | `id`             |
| First Difference Variance | `first.diff.var` |
| Autocorrelation           | `cor`            |
| Autocovariance            | `cov`            |
| Decay Time                | `decay.time`     |

## 5 Day Sliding Window

```{r}
constant.sims.var.5 <- constant.sims.var.5[1:55, ]
constant.sims.mean.5 <- constant.sims.mean.5[1:55, ]
constant.sims.cv.5 <- constant.sims.cv.5[1:55, ]
constant.sims.sk.5 <-  constant.sims.sk.5[1:55, ]
constant.sims.kt.5 <-  constant.sims.kt.5[1:55, ]
constant.sims.id.5 <-  constant.sims.id.5[1:55, ]
constant.sims.ac.5 <-  constant.sims.ac.5[1:55, ]
constant.sims.acv.5 <-  constant.sims.acv.5[1:55, ]
constant.sims.fdv.5 <-  constant.sims.fdv.5[1:55, ]
constant.sims.dt.5 <-  constant.sims.dt.5[1:55, ]

warming.sims.var.5 <- warming.sims.var.5[1:55, ]
warming.sims.mean.5 <- warming.sims.mean.5[1:55, ]
warming.sims.cv.5 <- warming.sims.cv.5[1:55, ]
warming.sims.sk.5 <-  warming.sims.sk.5[1:55, ]
warming.sims.kt.5 <-  warming.sims.kt.5[1:55, ]
warming.sims.id.5 <-  warming.sims.id.5[1:55, ]
warming.sims.ac.5 <-  warming.sims.ac.5[1:55, ]
warming.sims.acv.5 <-  warming.sims.acv.5[1:55, ]
warming.sims.fdv.5 <-  warming.sims.fdv.5[1:55, ]
warming.sims.dt.5 <-  warming.sims.dt.5[1:55, ]

constant.sims.var.5.tau <- c()
constant.sims.mean.5.tau <- c()
constant.sims.cv.5.tau <- c()
constant.sims.sk.5.tau <-  c()
constant.sims.kt.5.tau <- c()
constant.sims.id.5.tau <- c()
constant.sims.ac.5.tau <- c()
constant.sims.acv.5.tau <- c()
constant.sims.fdv.5.tau <-  c()
constant.sims.dt.5.tau <- c()

warming.sims.var.5.tau <- c()
warming.sims.mean.5.tau <- c()
warming.sims.cv.5.tau <- c()
warming.sims.sk.5.tau <-  c()
warming.sims.kt.5.tau <- c()
warming.sims.id.5.tau <-  c()
warming.sims.ac.5.tau <-  c()
warming.sims.acv.5.tau <-  c()
warming.sims.fdv.5.tau <-  c()
warming.sims.dt.5.tau <-  c()
```

Calculate Kendall's tau for each metric:

```{r}
time.5 <- c(1:55)

for(i in 1:num.of.sims){
  constant.sims.mean.5.tau[i] <- cor(constant.sims.mean.5[ , i], time.5, method = "kendall")
  constant.sims.sk.5.tau[i] <- cor(constant.sims.sk.5[ , i], time.5, method = "kendall")
  constant.sims.kt.5.tau[i] <- cor(constant.sims.kt.5[ , i], time.5, method = "kendall")
  constant.sims.var.5.tau[i] <- cor(constant.sims.var.5[ , i], time.5, method = "kendall")
  constant.sims.cv.5.tau[i] <- cor(constant.sims.cv.5[ , i], time.5, method = "kendall")
  constant.sims.id.5.tau[i] <- cor(constant.sims.id.5[ , i], time.5, method = "kendall")
  constant.sims.fdv.5.tau[i] <- cor(constant.sims.fdv.5[ , i], time.5, method = "kendall")
  constant.sims.ac.5.tau[i] <- cor(constant.sims.ac.5[ , i], time.5, method = "kendall")
  constant.sims.acv.5.tau[i] <- cor(constant.sims.acv.5[ , i], time.5, method = "kendall")
  constant.sims.dt.5.tau[i] <- cor(constant.sims.dt.5[ , i], time.5, method = "kendall")
  
  warming.sims.mean.5.tau[i] <- cor(warming.sims.mean.5[ , i], time.5, method = "kendall")
  warming.sims.sk.5.tau[i] <- cor(warming.sims.sk.5[ , i], time.5, method = "kendall")
  warming.sims.kt.5.tau[i] <- cor(warming.sims.kt.5[ , i], time.5, method = "kendall")
  warming.sims.var.5.tau[i] <- cor(warming.sims.var.5[ , i], time.5, method = "kendall")
  warming.sims.cv.5.tau[i] <- cor(warming.sims.cv.5[ , i], time.5, method = "kendall")
  warming.sims.id.5.tau[i] <- cor(warming.sims.id.5[ , i], time.5, method = "kendall")
  warming.sims.fdv.5.tau[i] <- cor(warming.sims.fdv.5[ , i], time.5, method = "kendall")
  warming.sims.ac.5.tau[i] <- cor(warming.sims.ac.5[ , i], time.5, method = "kendall")
  warming.sims.acv.5.tau[i] <- cor(warming.sims.acv.5[ , i], time.5, method = "kendall")
  warming.sims.dt.5.tau[i] <- cor(warming.sims.dt.5[ , i], time.5, method = "kendall")
}
```

Calculating AUC:

```{r}
AUC_var_5 <- calc_auc(warming.sims.var.5.tau, constant.sims.var.5.tau); AUC_var_5
AUC_mean_5 <- calc_auc(warming.sims.mean.5.tau, constant.sims.mean.5.tau); AUC_mean_5
AUC_kt_5 <- calc_auc(warming.sims.kt.5.tau, constant.sims.kt.5.tau); AUC_kt_5
AUC_sk_5 <- calc_auc(warming.sims.sk.5.tau, constant.sims.sk.5.tau); AUC_sk_5
AUC_id_5 <- calc_auc(warming.sims.id.5.tau, constant.sims.id.5.tau); AUC_id_5
AUC_fdv_5 <- calc_auc(warming.sims.fdv.5.tau, constant.sims.fdv.5.tau); AUC_fdv_5
AUC_ac_5 <- calc_auc(warming.sims.ac.5.tau, constant.sims.ac.5.tau); AUC_ac_5
AUC_acv_5 <- calc_auc(warming.sims.acv.5.tau, constant.sims.acv.5.tau); AUC_acv_5
AUC_cv_5 <- calc_auc(warming.sims.cv.5.tau, constant.sims.cv.5.tau); AUC_cv_5
AUC_dt_5 <- calc_auc(warming.sims.dt.5.tau, constant.sims.dt.5.tau); AUC_dt_5
```

## 15 Day Sliding Window

```{r}
constant.sims.var.15 <- constant.sims.var.15[1:45, ]
constant.sims.mean.15 <- constant.sims.mean.15[1:45, ]
constant.sims.cv.15 <- constant.sims.cv.15[1:45, ]
constant.sims.sk.15 <-  constant.sims.sk.15[1:45, ]
constant.sims.kt.15 <-  constant.sims.kt.15[1:45, ]
constant.sims.id.15 <-  constant.sims.id.15[1:45, ]
constant.sims.ac.15 <-  constant.sims.ac.15[1:45, ]
constant.sims.acv.15 <-  constant.sims.acv.15[1:45, ]
constant.sims.fdv.15 <-  constant.sims.fdv.15[1:45, ]
constant.sims.dt.15 <-  constant.sims.dt.15[1:45, ]

warming.sims.var.15 <- warming.sims.var.15[1:45, ]
warming.sims.mean.15 <- warming.sims.mean.15[1:45, ]
warming.sims.cv.15 <- warming.sims.cv.15[1:45, ]
warming.sims.sk.15 <-  warming.sims.sk.15[1:45, ]
warming.sims.kt.15 <-  warming.sims.kt.15[1:45, ]
warming.sims.id.15 <-  warming.sims.id.15[1:45, ]
warming.sims.ac.15 <-  warming.sims.ac.15[1:45, ]
warming.sims.acv.15 <-  warming.sims.acv.15[1:45, ]
warming.sims.fdv.15 <-  warming.sims.fdv.15[1:45, ]
warming.sims.dt.15 <-  warming.sims.dt.15[1:45, ]

constant.sims.var.15.tau <- c()
constant.sims.mean.15.tau <- c()
constant.sims.cv.15.tau <- c()
constant.sims.sk.15.tau <-  c()
constant.sims.kt.15.tau <- c()
constant.sims.id.15.tau <- c()
constant.sims.ac.15.tau <- c()
constant.sims.acv.15.tau <- c()
constant.sims.fdv.15.tau <-  c()
constant.sims.dt.15.tau <- c()

warming.sims.var.15.tau <- c()
warming.sims.mean.15.tau <- c()
warming.sims.cv.15.tau <- c()
warming.sims.sk.15.tau <-  c()
warming.sims.kt.15.tau <- c()
warming.sims.id.15.tau <-  c()
warming.sims.ac.15.tau <-  c()
warming.sims.acv.15.tau <-  c()
warming.sims.fdv.15.tau <-  c()
warming.sims.dt.15.tau <-  c()
```

Calculate Kendall's tau for each metric:

```{r}
time.15 <- c(1:45)

for(i in 1:num.of.sims){
  constant.sims.mean.15.tau[i] <- cor(constant.sims.mean.15[ , i], time.15, method = "kendall")
  constant.sims.sk.15.tau[i] <- cor(constant.sims.sk.15[ , i], time.15, method = "kendall")
  constant.sims.kt.15.tau[i] <- cor(constant.sims.kt.15[ , i], time.15, method = "kendall")
  constant.sims.var.15.tau[i] <- cor(constant.sims.var.15[ , i], time.15, method = "kendall")
  constant.sims.cv.15.tau[i] <- cor(constant.sims.cv.15[ , i], time.15, method = "kendall")
  constant.sims.id.15.tau[i] <- cor(constant.sims.id.15[ , i], time.15, method = "kendall")
  constant.sims.fdv.15.tau[i] <- cor(constant.sims.fdv.15[ , i], time.15, method = "kendall")
  constant.sims.ac.15.tau[i] <- cor(constant.sims.ac.15[ , i], time.15, method = "kendall")
  constant.sims.acv.15.tau[i] <- cor(constant.sims.acv.15[ , i], time.15, method = "kendall")
  constant.sims.dt.15.tau[i] <- cor(constant.sims.dt.15[ , i], time.15, method = "kendall")
  
  warming.sims.mean.15.tau[i] <- cor(warming.sims.mean.15[ , i], time.15, method = "kendall")
  warming.sims.sk.15.tau[i] <- cor(warming.sims.sk.15[ , i], time.15, method = "kendall")
  warming.sims.kt.15.tau[i] <- cor(warming.sims.kt.15[ , i], time.15, method = "kendall")
  warming.sims.var.15.tau[i] <- cor(warming.sims.var.15[ , i], time.15, method = "kendall")
  warming.sims.cv.15.tau[i] <- cor(warming.sims.cv.15[ , i], time.15, method = "kendall")
  warming.sims.id.15.tau[i] <- cor(warming.sims.id.15[ , i], time.15, method = "kendall")
  warming.sims.fdv.15.tau[i] <- cor(warming.sims.fdv.15[ , i], time.15, method = "kendall")
  warming.sims.ac.15.tau[i] <- cor(warming.sims.ac.15[ , i], time.15, method = "kendall")
  warming.sims.acv.15.tau[i] <- cor(warming.sims.acv.15[ , i], time.15, method = "kendall")
  warming.sims.dt.15.tau[i] <- cor(warming.sims.dt.15[ , i], time.15, method = "kendall")
}
```

Calculating AUC:

```{r}
AUC_var_15 <- calc_auc(warming.sims.var.15.tau, constant.sims.var.15.tau); AUC_var_15
AUC_mean_15 <- calc_auc(warming.sims.mean.15.tau, constant.sims.mean.15.tau); AUC_mean_15
AUC_kt_15 <- calc_auc(warming.sims.kt.15.tau, constant.sims.kt.15.tau); AUC_kt_15
AUC_sk_15 <- calc_auc(warming.sims.sk.15.tau, constant.sims.sk.15.tau); AUC_sk_15
AUC_id_15 <- calc_auc(warming.sims.id.15.tau, constant.sims.id.15.tau); AUC_id_15
AUC_fdv_15 <- calc_auc(warming.sims.fdv.15.tau, constant.sims.fdv.15.tau); AUC_fdv_15
AUC_ac_15 <- calc_auc(warming.sims.ac.15.tau, constant.sims.ac.15.tau); AUC_ac_15
AUC_acv_15 <- calc_auc(warming.sims.acv.15.tau, constant.sims.acv.15.tau); AUC_acv_15
AUC_cv_15 <- calc_auc(warming.sims.cv.15.tau, constant.sims.cv.15.tau); AUC_cv_15
AUC_dt_15 <- calc_auc(warming.sims.dt.15.tau, constant.sims.dt.15.tau); AUC_dt_15
```

## 30 Day Sliding Window

```{r}
constant.sims.var.30 <- constant.sims.var.30[1:30, ]
constant.sims.mean.30 <- constant.sims.mean.30[1:30, ]
constant.sims.cv.30 <- constant.sims.cv.30[1:30, ]
constant.sims.sk.30 <-  constant.sims.sk.30[1:30, ]
constant.sims.kt.30 <-  constant.sims.kt.30[1:30, ]
constant.sims.id.30 <-  constant.sims.id.30[1:30, ]
constant.sims.ac.30 <-  constant.sims.ac.30[1:30, ]
constant.sims.acv.30 <-  constant.sims.acv.30[1:30, ]
constant.sims.fdv.30 <-  constant.sims.fdv.30[1:30, ]
constant.sims.dt.30 <-  constant.sims.dt.30[1:30, ]

warming.sims.var.30 <- warming.sims.var.30[1:30, ]
warming.sims.mean.30 <- warming.sims.mean.30[1:30, ]
warming.sims.cv.30 <- warming.sims.cv.30[1:30, ]
warming.sims.sk.30 <-  warming.sims.sk.30[1:30, ]
warming.sims.kt.30 <-  warming.sims.kt.30[1:30, ]
warming.sims.id.30 <-  warming.sims.id.30[1:30, ]
warming.sims.ac.30 <-  warming.sims.ac.30[1:30, ]
warming.sims.acv.30 <-  warming.sims.acv.30[1:30, ]
warming.sims.fdv.30 <-  warming.sims.fdv.30[1:30, ]
warming.sims.dt.30 <-  warming.sims.dt.30[1:30, ]

constant.sims.var.30.tau <- c()
constant.sims.mean.30.tau <- c()
constant.sims.cv.30.tau <- c()
constant.sims.sk.30.tau <-  c()
constant.sims.kt.30.tau <- c()
constant.sims.id.30.tau <- c()
constant.sims.ac.30.tau <- c()
constant.sims.acv.30.tau <- c()
constant.sims.fdv.30.tau <-  c()
constant.sims.dt.30.tau <- c()

warming.sims.var.30.tau <- c()
warming.sims.mean.30.tau <- c()
warming.sims.cv.30.tau <- c()
warming.sims.sk.30.tau <-  c()
warming.sims.kt.30.tau <- c()
warming.sims.id.30.tau <-  c()
warming.sims.ac.30.tau <-  c()
warming.sims.acv.30.tau <-  c()
warming.sims.fdv.30.tau <-  c()
warming.sims.dt.30.tau <-  c()
```

Calculate Kendall's tau for each metric:

```{r}
time.30 <- c(1:30)

for(i in 1:num.of.sims){
  constant.sims.mean.30.tau[i] <- cor(constant.sims.mean.30[ , i], time.30, method = "kendall")
  constant.sims.sk.30.tau[i] <- cor(constant.sims.sk.30[ , i], time.30, method = "kendall")
  constant.sims.kt.30.tau[i] <- cor(constant.sims.kt.30[ , i], time.30, method = "kendall")
  constant.sims.var.30.tau[i] <- cor(constant.sims.var.30[ , i], time.30, method = "kendall")
  constant.sims.cv.30.tau[i] <- cor(constant.sims.cv.30[ , i], time.30, method = "kendall")
  constant.sims.id.30.tau[i] <- cor(constant.sims.id.30[ , i], time.30, method = "kendall")
  constant.sims.fdv.30.tau[i] <- cor(constant.sims.fdv.30[ , i], time.30, method = "kendall")
  constant.sims.ac.30.tau[i] <- cor(constant.sims.ac.30[ , i], time.30, method = "kendall")
  constant.sims.acv.30.tau[i] <- cor(constant.sims.acv.30[ , i], time.30, method = "kendall")
  constant.sims.dt.30.tau[i] <- cor(constant.sims.dt.30[ , i], time.30, method = "kendall")
  
  warming.sims.mean.30.tau[i] <- cor(warming.sims.mean.30[ , i], time.30, method = "kendall")
  warming.sims.sk.30.tau[i] <- cor(warming.sims.sk.30[ , i], time.30, method = "kendall")
  warming.sims.kt.30.tau[i] <- cor(warming.sims.kt.30[ , i], time.30, method = "kendall")
  warming.sims.var.30.tau[i] <- cor(warming.sims.var.30[ , i], time.30, method = "kendall")
  warming.sims.cv.30.tau[i] <- cor(warming.sims.cv.30[ , i], time.30, method = "kendall")
  warming.sims.id.30.tau[i] <- cor(warming.sims.id.30[ , i], time.30, method = "kendall")
  warming.sims.fdv.30.tau[i] <- cor(warming.sims.fdv.30[ , i], time.30, method = "kendall")
  warming.sims.ac.30.tau[i] <- cor(warming.sims.ac.30[ , i], time.30, method = "kendall")
  warming.sims.acv.30.tau[i] <- cor(warming.sims.acv.30[ , i], time.30, method = "kendall")
  warming.sims.dt.30.tau[i] <- cor(warming.sims.dt.30[ , i], time.30, method = "kendall")
}
```

Calculating AUC:

```{r}
AUC_var_30 <- calc_auc(warming.sims.var.30.tau, constant.sims.var.30.tau); AUC_var_30
AUC_mean_30 <- calc_auc(warming.sims.mean.30.tau, constant.sims.mean.30.tau); AUC_mean_30
AUC_kt_30 <- calc_auc(warming.sims.kt.30.tau, constant.sims.kt.30.tau); AUC_kt_30
AUC_sk_30 <- calc_auc(warming.sims.sk.30.tau, constant.sims.sk.30.tau); AUC_sk_30
AUC_id_30 <- calc_auc(warming.sims.id.30.tau, constant.sims.id.30.tau); AUC_id_30
AUC_fdv_30 <- calc_auc(warming.sims.fdv.30.tau, constant.sims.fdv.30.tau); AUC_fdv_30
AUC_ac_30 <- calc_auc(warming.sims.ac.30.tau, constant.sims.ac.30.tau); AUC_ac_30
AUC_acv_30 <- calc_auc(warming.sims.acv.30.tau, constant.sims.acv.30.tau); AUC_acv_30
AUC_cv_30 <- calc_auc(warming.sims.cv.30.tau, constant.sims.cv.30.tau); AUC_cv_30
AUC_dt_30 <- calc_auc(warming.sims.dt.30.tau, constant.sims.dt.30.tau); AUC_dt_30
```

## Summarize Outputs

```{r}
metrics <- c("mean", "skewness", "kurtosis", "variance", "cv", "id", "first.diff.var", "cor", "cov", "decay.time")

AUC_5 <- c(AUC_mean_5, AUC_sk_5, AUC_kt_5, AUC_var_5, AUC_cv_5, 
           AUC_id_5, AUC_fdv_5, AUC_ac_5 ,AUC_acv_5, AUC_dt_5)
AUC_15 <- c(AUC_mean_15, AUC_sk_15, AUC_kt_15, AUC_var_15, AUC_cv_15, 
           AUC_id_15, AUC_fdv_15, AUC_ac_15 ,AUC_acv_15, AUC_dt_15)
AUC_30 <- c(AUC_mean_30, AUC_sk_30, AUC_kt_30, AUC_var_30, AUC_cv_30, 
           AUC_id_30, AUC_fdv_30, AUC_ac_30 ,AUC_acv_30, AUC_dt_30)
AUC_df <- data.frame(metrics, AUC_5, AUC_15, AUC_30); AUC_df
saveRDS(AUC_df, "Objects/AUC_SmoothSimulatedData_B2.rds")
```

A quick plot!

```{r}
metrics_full <- c("Mean", "Skewness", "Kurtosis", "Variance", "Covariance", "Index of Dispersion", "First Difference Variance", "Autocorrelation", "Autocovariance", "Decay Time")

## Plot 
plot(1:10, AUC_df$AUC_5, type = "p", 
     main = "Smoothed Simulated Data | Bandwidth = 2", xaxt = "none", xlab = "", ylim = c(0.3, 0.7), ylab = "AUC",
     col = "forestgreen", pch = 19)
points(1:10, AUC_df$AUC_15, col = "orchid", pch = 19)
points(1:10, AUC_df$AUC_30, col = "#4B0082", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)
legend("bottomright", 
       legend = c("Sliding Window = 5 Days", "Sliding Window = 15 Days", "Sliding Window = 30 Days"),
       col = c("forestgreen", "orchid", "#4B0082"), pch = 19)

## Saving plot as .png
png(file = "Plots/AUC_SmoothSimulatedData_B2.png", width = 1500, height = 350)
plot(1:10, AUC_df$AUC_5, type = "p", 
     main = "Smoothed Simulated Data | Bandwidth = 2", xaxt = "none", xlab = "", ylim = c(0.3, 0.7), ylab = "AUC",
     col = "forestgreen", pch = 19)
points(1:10, AUC_df$AUC_15, col = "orchid", pch = 19)
points(1:10, AUC_df$AUC_30, col = "#4B0082", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full)
legend("bottomright", 
       legend = c("Sliding Window = 5 Days", "Sliding Window = 15 Days", "Sliding Window = 30 Days"),
       col = c("forestgreen", "orchid", "#4B0082"), pch = 19)
dev.off()
```

# BANDWIDTH = 4

# Calculating Statistical Metrics Along Variously-Sized Sliding Windows

For details on these metrics, see `3_EWS-Sim.Rmd`.

## 5 Day Sliding Window

### Standard Deviation, Mean, Skewness, Kurtosis

```{r}
num.of.sims <- 10^3

# Making empty matricies
constant.sims.sd.5 <- matrix(nrow = 115, ncol = num.of.sims)
constant.sims.mean.5 <- matrix(nrow = 115, ncol = num.of.sims)
constant.sims.sk.5 <- matrix(nrow = 115, ncol = num.of.sims)
constant.sims.kt.5 <- matrix(nrow = 115, ncol = num.of.sims)

warming.sims.sd.5 <- matrix(nrow = 115, ncol = num.of.sims)
warming.sims.mean.5 <- matrix(nrow = 115, ncol = num.of.sims)
warming.sims.sk.5 <- matrix(nrow = 115, ncol = num.of.sims)
warming.sims.kt.5 <- matrix(nrow = 115, ncol = num.of.sims)

# Use SlidingWindow to calculate sd, mean, skewness, and kurtosis
for(i in 1:num.of.sims){
  constant.sims.sd.5[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.const.b4[,i], window = 5,step = 1)
  constant.sims.mean.5[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.const.b4[,i], window = 5, step = 1)
  constant.sims.sk.5[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.const.b4[,i], window = 5, step = 1)
  constant.sims.kt.5[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.const.b4[,i], window = 5, step = 1)
}

for(i in 1:num.of.sims){
  warming.sims.sd.5[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.warm.b4[,i], window = 5,step = 1)
  warming.sims.mean.5[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.warm.b4[,i], window = 5, step = 1)
  warming.sims.sk.5[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.warm.b4[,i], window = 5, step = 1)
  warming.sims.kt.5[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.warm.b4[,i], window = 5, step = 1)
}
```

### Variation

```{r}
constant.sims.var.5 <- constant.sims.sd.5*constant.sims.sd.5
warming.sims.var.5 <- warming.sims.sd.5*warming.sims.sd.5
```

### Co-Efficient of Variation

```{r}
constant.sims.cv.5 <- constant.sims.sd.5/constant.sims.mean.5
warming.sims.cv.5 <- warming.sims.sd.5/warming.sims.mean.5
```

### Index of Dispersion

```{r}
constant.sims.id.5 <- constant.sims.var.5/constant.sims.mean.5
warming.sims.id.5 <- warming.sims.var.5/warming.sims.mean.5
```

### First Difference Variance

```{r}
constant.sims.fdv.5 <- matrix(ncol = num.of.sims, nrow = 115)
warming.sims.fdv.5 <- matrix(ncol = num.of.sims, nrow = 115)

for(i in 1:num.of.sims){
  for(j in 2:115){
    constant.sims.fdv.5[j - 1,i] <- constant.sims.var.5[j,i] - constant.sims.var.5[j - 1,i]
    warming.sims.fdv.5[j - 1,i] <- warming.sims.var.5[j,i] - warming.sims.var.5[j - 1,i]
  }
}

# Get rid of NA (last) row
constant.sims.fdv.5 <- constant.sims.fdv.5[1:114, ]
warming.sims.fdv.5 <- warming.sims.fdv.5[1:114, ]
```

### Autocorrelation and Autocovariance

```{r}
constant.sims.ac.5 <- matrix(ncol = num.of.sims, nrow = 120)
constant.sims.acv.5 <- matrix(ncol = num.of.sims, nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:115){
    window.start <- i
    window.end <- i + 4
    
    segment <- smooth.sim.const.b4[c(window.start:window.end),j]
    
    constant.sims.ac.5[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    constant.sims.acv.5[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}

warming.sims.ac.5 <- matrix(ncol=num.of.sims,nrow = 120)
warming.sims.acv.5 <- matrix(ncol=num.of.sims,nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:115){
    window.start <- i
    window.end <- i + 4
    
    segment <- smooth.sim.warm.b4[c(window.start:window.end),j]
    
    warming.sims.ac.5[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    warming.sims.acv.5[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}
```

### Decay Time

```{r}
constant.sims.dt.5 <- matrix(ncol = num.of.sims, nrow = 55)
warming.sims.dt.5 <- matrix(nco = num.of.sims, nrow = 55)

for(i in 1:num.of.sims){
  for(j in 1:55){
    suppressWarnings(constant.sims.dt.5[j, i] <- -j/log(constant.sims.ac.5[j, i]))
    suppressWarnings(warming.sims.dt.5[j, i] <- -j/log(warming.sims.ac.5[j, i]))
  }
}
```

## 15 Day Sliding Window

### Standard Deviation, Mean, Skewness, Kurtosis

```{r}
num.of.sims <- 10^3

# Making empty matricies
constant.sims.sd.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.mean.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.sk.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.kt.15 <- matrix(nrow = 105, ncol = num.of.sims)

warming.sims.sd.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.mean.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.sk.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.kt.15 <- matrix(nrow = 105, ncol = num.of.sims)

# Use SlidingWindow to calculate sd, mean, skewness, and kurtosis
for(i in 1:num.of.sims){
  constant.sims.sd.15[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.const.b4[,i], window = 15,step = 1)
  constant.sims.mean.15[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.const.b4[,i], window = 15, step = 1)
  constant.sims.sk.15[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.const.b4[,i], window = 15, step = 1)
  constant.sims.kt.15[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.const.b4[,i], window = 15, step = 1)
}

for(i in 1:num.of.sims){
  warming.sims.sd.15[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.warm.b4[,i], window = 15,step = 1)
  warming.sims.mean.15[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.warm.b4[,i], window = 15, step = 1)
  warming.sims.sk.15[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.warm.b4[,i], window = 15, step = 1)
  warming.sims.kt.15[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.warm.b4[,i], window = 15, step = 1)
}
```

### Variation

```{r}
constant.sims.var.15 <- constant.sims.sd.15*constant.sims.sd.15
warming.sims.var.15 <- warming.sims.sd.15*warming.sims.sd.15
```

### Co-Efficient of Variation

```{r}
constant.sims.cv.15 <- constant.sims.sd.15/constant.sims.mean.15
warming.sims.cv.15 <- warming.sims.sd.15/warming.sims.mean.15
```

### Index of Dispersion

```{r}
constant.sims.id.15 <- constant.sims.var.15/constant.sims.mean.15
warming.sims.id.15 <- warming.sims.var.15/warming.sims.mean.15
```

### First Difference Variance

```{r}
constant.sims.fdv.15 <- matrix(ncol = num.of.sims, nrow = 105)
warming.sims.fdv.15 <- matrix(ncol = num.of.sims, nrow = 105)

for(i in 1:num.of.sims){
  for(j in 2:105){
    constant.sims.fdv.15[j - 1,i] <- constant.sims.var.15[j,i] - constant.sims.var.15[j - 1,i]
    warming.sims.fdv.15[j - 1,i] <- warming.sims.var.15[j,i] - warming.sims.var.15[j - 1,i]
  }
}

# Get rid of NA (last) row
constant.sims.fdv.15 <- constant.sims.fdv.15[1:104, ]
warming.sims.fdv.15 <- warming.sims.fdv.15[1:104, ]
```

### Autocorrelation and Autocovariance

```{r}
constant.sims.ac.15 <- matrix(ncol = num.of.sims, nrow = 120)
constant.sims.acv.15 <- matrix(ncol = num.of.sims, nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:105){
    window.start <- i
    window.end <- i + 14
    
    segment <- smooth.sim.const.b4[c(window.start:window.end),j]
    
    constant.sims.ac.15[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    constant.sims.acv.15[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}

warming.sims.ac.15 <- matrix(ncol=num.of.sims,nrow = 120)
warming.sims.acv.15 <- matrix(ncol=num.of.sims,nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:105){
    window.start <- i
    window.end <- i + 14
    
    segment <- smooth.sim.warm.b4[c(window.start:window.end),j]
    
    warming.sims.ac.15[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    warming.sims.acv.15[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}
```

### Decay Time

```{r}
constant.sims.dt.15 <- matrix(ncol = num.of.sims, nrow = 45)
warming.sims.dt.15 <- matrix(nco = num.of.sims, nrow = 45)

for(i in 1:num.of.sims){
  for(j in 1:45){
    suppressWarnings(constant.sims.dt.15[j, i] <- -j/log(constant.sims.ac.15[j, i]))
    suppressWarnings(warming.sims.dt.15[j, i] <- -j/log(warming.sims.ac.15[j, i]))
  }
}
```

## 30 Day Sliding Window

### Standard Deviation, Mean, Skewness, Kurtosis

```{r}
num.of.sims <- 10^3

# Making empty matricies
constant.sims.sd.30 <- matrix(nrow = 90, ncol = num.of.sims)
constant.sims.mean.30 <- matrix(nrow = 90, ncol = num.of.sims)
constant.sims.sk.30 <- matrix(nrow = 90, ncol = num.of.sims)
constant.sims.kt.30 <- matrix(nrow = 90, ncol = num.of.sims)

warming.sims.sd.30 <- matrix(nrow = 90, ncol = num.of.sims)
warming.sims.mean.30 <- matrix(nrow = 90, ncol = num.of.sims)
warming.sims.sk.30 <- matrix(nrow = 90, ncol = num.of.sims)
warming.sims.kt.30 <- matrix(nrow = 90, ncol = num.of.sims)

# Use SlidingWindow to calculate sd, mean, skewness, and kurtosis
for(i in 1:num.of.sims){
  constant.sims.sd.30[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.const.b4[,i], window = 30,step = 1)
  constant.sims.mean.30[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.const.b4[,i], window = 30, step = 1)
  constant.sims.sk.30[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.const.b4[,i], window = 30, step = 1)
  constant.sims.kt.30[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.const.b4[,i], window = 30, step = 1)
}

for(i in 1:num.of.sims){
  warming.sims.sd.30[ ,i] <- SlidingWindow(FUN = "sd", data = smooth.sim.warm.b4[,i], window = 30,step = 1)
  warming.sims.mean.30[ ,i] <- SlidingWindow(FUN = "mean", data = smooth.sim.warm.b4[,i], window = 30, step = 1)
  warming.sims.sk.30[ ,i] <- SlidingWindow(FUN = "skewness", data = smooth.sim.warm.b4[,i], window = 30, step = 1)
  warming.sims.kt.30[ ,i] <- SlidingWindow(FUN = "kurtosis", data = smooth.sim.warm.b4[,i], window = 30, step = 1)
}
```

### Variation

```{r}
constant.sims.var.30 <- constant.sims.sd.30*constant.sims.sd.30
warming.sims.var.30 <- warming.sims.sd.30*warming.sims.sd.30
```

### Co-Efficient of Variation

```{r}
constant.sims.cv.30 <- constant.sims.sd.30/constant.sims.mean.30
warming.sims.cv.30 <- warming.sims.sd.30/warming.sims.mean.30
```

### Index of Dispersion

```{r}
constant.sims.id.30 <- constant.sims.var.30/constant.sims.mean.30
warming.sims.id.30 <- warming.sims.var.30/warming.sims.mean.30
```

### First Difference Variance

```{r}
constant.sims.fdv.30 <- matrix(ncol = num.of.sims, nrow = 90)
warming.sims.fdv.30 <- matrix(ncol = num.of.sims, nrow = 90)

for(i in 1:num.of.sims){
  for(j in 2:90){
    constant.sims.fdv.30[j - 1,i] <- constant.sims.var.30[j,i] - constant.sims.var.30[j - 1,i]
    warming.sims.fdv.30[j - 1,i] <- warming.sims.var.30[j,i] - warming.sims.var.30[j - 1,i]
  }
}

# Get rid of NA (last) row
constant.sims.fdv.30 <- constant.sims.fdv.30[1:89, ]
warming.sims.fdv.30 <- warming.sims.fdv.30[1:89, ]
```

### Autocorrelation and Autocovariance

```{r}
constant.sims.ac.30 <- matrix(ncol = num.of.sims, nrow = 120)
constant.sims.acv.30 <- matrix(ncol = num.of.sims, nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:90){
    window.start <- i
    window.end <- i + 29
    
    segment <- smooth.sim.const.b4[c(window.start:window.end),j]
    
    constant.sims.ac.30[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    constant.sims.acv.30[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}

warming.sims.ac.30 <- matrix(ncol=num.of.sims,nrow = 120)
warming.sims.acv.30 <- matrix(ncol=num.of.sims,nrow = 120)

for(j in 1:num.of.sims){
  for(i in 1:90){
    window.start <- i
    window.end <- i + 29
    
    segment <- smooth.sim.warm.b4[c(window.start:window.end),j]
    
    warming.sims.ac.30[i,j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    warming.sims.acv.30[i,j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}
```

### Decay Time

```{r}
constant.sims.dt.30 <- matrix(ncol = num.of.sims, nrow = 30)
warming.sims.dt.30 <- matrix(nco = num.of.sims, nrow = 30)

for(i in 1:num.of.sims){
  for(j in 1:30){
    suppressWarnings(constant.sims.dt.30[j, i] <- -j/log(constant.sims.ac.30[j, i]))
    suppressWarnings(warming.sims.dt.30[j, i] <- -j/log(warming.sims.ac.30[j, i]))
  }
}
```

# Area Under Curve (AUC) Tests

Function to calculate AUC:

```{r}
calc_auc <- function(test.taus, null.taus){
  
  combined.taus <- c(test.taus, null.taus)
  r <- rank(combined.taus, na.last = NA)
  
  test.taus.length <- sum(!is.na(test.taus))
  null.taus.length <- sum(!is.na(null.taus))
  
  r1 <- sum(r[1:test.taus.length])
  
  n1 <- test.taus.length
  n2 <- null.taus.length
  
  (r1 - n1 * (n1 + 1) / 2) / (n1 * n2)
}
```

Metrics to evaluate:

|                           |                  |
|---------------------------|------------------|
| Standard Deviation        | `sd`             |
| Mean                      | `mean`           |
| Skew                      | `skewness`       |
| Kurtosis                  | `kurtosis`       |
| Variance                  | `variance`       |
| Co-Efficient of Variance  | `cv`             |
| Index of Dispersion       | `id`             |
| First Difference Variance | `first.diff.var` |
| Autocorrelation           | `cor`            |
| Autocovariance            | `cov`            |
| Decay Time                | `decay.time`     |

## 5 Day Sliding Window

```{r}
constant.sims.var.5 <- constant.sims.var.5[1:55, ]
constant.sims.mean.5 <- constant.sims.mean.5[1:55, ]
constant.sims.cv.5 <- constant.sims.cv.5[1:55, ]
constant.sims.sk.5 <-  constant.sims.sk.5[1:55, ]
constant.sims.kt.5 <-  constant.sims.kt.5[1:55, ]
constant.sims.id.5 <-  constant.sims.id.5[1:55, ]
constant.sims.ac.5 <-  constant.sims.ac.5[1:55, ]
constant.sims.acv.5 <-  constant.sims.acv.5[1:55, ]
constant.sims.fdv.5 <-  constant.sims.fdv.5[1:55, ]
constant.sims.dt.5 <-  constant.sims.dt.5[1:55, ]

warming.sims.var.5 <- warming.sims.var.5[1:55, ]
warming.sims.mean.5 <- warming.sims.mean.5[1:55, ]
warming.sims.cv.5 <- warming.sims.cv.5[1:55, ]
warming.sims.sk.5 <-  warming.sims.sk.5[1:55, ]
warming.sims.kt.5 <-  warming.sims.kt.5[1:55, ]
warming.sims.id.5 <-  warming.sims.id.5[1:55, ]
warming.sims.ac.5 <-  warming.sims.ac.5[1:55, ]
warming.sims.acv.5 <-  warming.sims.acv.5[1:55, ]
warming.sims.fdv.5 <-  warming.sims.fdv.5[1:55, ]
warming.sims.dt.5 <-  warming.sims.dt.5[1:55, ]

constant.sims.var.5.tau <- c()
constant.sims.mean.5.tau <- c()
constant.sims.cv.5.tau <- c()
constant.sims.sk.5.tau <-  c()
constant.sims.kt.5.tau <- c()
constant.sims.id.5.tau <- c()
constant.sims.ac.5.tau <- c()
constant.sims.acv.5.tau <- c()
constant.sims.fdv.5.tau <-  c()
constant.sims.dt.5.tau <- c()

warming.sims.var.5.tau <- c()
warming.sims.mean.5.tau <- c()
warming.sims.cv.5.tau <- c()
warming.sims.sk.5.tau <-  c()
warming.sims.kt.5.tau <- c()
warming.sims.id.5.tau <-  c()
warming.sims.ac.5.tau <-  c()
warming.sims.acv.5.tau <-  c()
warming.sims.fdv.5.tau <-  c()
warming.sims.dt.5.tau <-  c()
```

Calculate Kendall's tau for each metric:

```{r}
time.5 <- c(1:55)

for(i in 1:num.of.sims){
  constant.sims.mean.5.tau[i] <- cor(constant.sims.mean.5[ , i], time.5, method = "kendall")
  constant.sims.sk.5.tau[i] <- cor(constant.sims.sk.5[ , i], time.5, method = "kendall")
  constant.sims.kt.5.tau[i] <- cor(constant.sims.kt.5[ , i], time.5, method = "kendall")
  constant.sims.var.5.tau[i] <- cor(constant.sims.var.5[ , i], time.5, method = "kendall")
  constant.sims.cv.5.tau[i] <- cor(constant.sims.cv.5[ , i], time.5, method = "kendall")
  constant.sims.id.5.tau[i] <- cor(constant.sims.id.5[ , i], time.5, method = "kendall")
  constant.sims.fdv.5.tau[i] <- cor(constant.sims.fdv.5[ , i], time.5, method = "kendall")
  constant.sims.ac.5.tau[i] <- cor(constant.sims.ac.5[ , i], time.5, method = "kendall")
  constant.sims.acv.5.tau[i] <- cor(constant.sims.acv.5[ , i], time.5, method = "kendall")
  constant.sims.dt.5.tau[i] <- cor(constant.sims.dt.5[ , i], time.5, method = "kendall")
  
  warming.sims.mean.5.tau[i] <- cor(warming.sims.mean.5[ , i], time.5, method = "kendall")
  warming.sims.sk.5.tau[i] <- cor(warming.sims.sk.5[ , i], time.5, method = "kendall")
  warming.sims.kt.5.tau[i] <- cor(warming.sims.kt.5[ , i], time.5, method = "kendall")
  warming.sims.var.5.tau[i] <- cor(warming.sims.var.5[ , i], time.5, method = "kendall")
  warming.sims.cv.5.tau[i] <- cor(warming.sims.cv.5[ , i], time.5, method = "kendall")
  warming.sims.id.5.tau[i] <- cor(warming.sims.id.5[ , i], time.5, method = "kendall")
  warming.sims.fdv.5.tau[i] <- cor(warming.sims.fdv.5[ , i], time.5, method = "kendall")
  warming.sims.ac.5.tau[i] <- cor(warming.sims.ac.5[ , i], time.5, method = "kendall")
  warming.sims.acv.5.tau[i] <- cor(warming.sims.acv.5[ , i], time.5, method = "kendall")
  warming.sims.dt.5.tau[i] <- cor(warming.sims.dt.5[ , i], time.5, method = "kendall")
}
```

Calculating AUC:

```{r}
AUC_var_5 <- calc_auc(warming.sims.var.5.tau, constant.sims.var.5.tau); AUC_var_5
AUC_mean_5 <- calc_auc(warming.sims.mean.5.tau, constant.sims.mean.5.tau); AUC_mean_5
AUC_kt_5 <- calc_auc(warming.sims.kt.5.tau, constant.sims.kt.5.tau); AUC_kt_5
AUC_sk_5 <- calc_auc(warming.sims.sk.5.tau, constant.sims.sk.5.tau); AUC_sk_5
AUC_id_5 <- calc_auc(warming.sims.id.5.tau, constant.sims.id.5.tau); AUC_id_5
AUC_fdv_5 <- calc_auc(warming.sims.fdv.5.tau, constant.sims.fdv.5.tau); AUC_fdv_5
AUC_ac_5 <- calc_auc(warming.sims.ac.5.tau, constant.sims.ac.5.tau); AUC_ac_5
AUC_acv_5 <- calc_auc(warming.sims.acv.5.tau, constant.sims.acv.5.tau); AUC_acv_5
AUC_cv_5 <- calc_auc(warming.sims.cv.5.tau, constant.sims.cv.5.tau); AUC_cv_5
AUC_dt_5 <- calc_auc(warming.sims.dt.5.tau, constant.sims.dt.5.tau); AUC_dt_5
```

## 15 Day Sliding Window

```{r}
constant.sims.var.15 <- constant.sims.var.15[1:45, ]
constant.sims.mean.15 <- constant.sims.mean.15[1:45, ]
constant.sims.cv.15 <- constant.sims.cv.15[1:45, ]
constant.sims.sk.15 <-  constant.sims.sk.15[1:45, ]
constant.sims.kt.15 <-  constant.sims.kt.15[1:45, ]
constant.sims.id.15 <-  constant.sims.id.15[1:45, ]
constant.sims.ac.15 <-  constant.sims.ac.15[1:45, ]
constant.sims.acv.15 <-  constant.sims.acv.15[1:45, ]
constant.sims.fdv.15 <-  constant.sims.fdv.15[1:45, ]
constant.sims.dt.15 <-  constant.sims.dt.15[1:45, ]

warming.sims.var.15 <- warming.sims.var.15[1:45, ]
warming.sims.mean.15 <- warming.sims.mean.15[1:45, ]
warming.sims.cv.15 <- warming.sims.cv.15[1:45, ]
warming.sims.sk.15 <-  warming.sims.sk.15[1:45, ]
warming.sims.kt.15 <-  warming.sims.kt.15[1:45, ]
warming.sims.id.15 <-  warming.sims.id.15[1:45, ]
warming.sims.ac.15 <-  warming.sims.ac.15[1:45, ]
warming.sims.acv.15 <-  warming.sims.acv.15[1:45, ]
warming.sims.fdv.15 <-  warming.sims.fdv.15[1:45, ]
warming.sims.dt.15 <-  warming.sims.dt.15[1:45, ]

constant.sims.var.15.tau <- c()
constant.sims.mean.15.tau <- c()
constant.sims.cv.15.tau <- c()
constant.sims.sk.15.tau <-  c()
constant.sims.kt.15.tau <- c()
constant.sims.id.15.tau <- c()
constant.sims.ac.15.tau <- c()
constant.sims.acv.15.tau <- c()
constant.sims.fdv.15.tau <-  c()
constant.sims.dt.15.tau <- c()

warming.sims.var.15.tau <- c()
warming.sims.mean.15.tau <- c()
warming.sims.cv.15.tau <- c()
warming.sims.sk.15.tau <-  c()
warming.sims.kt.15.tau <- c()
warming.sims.id.15.tau <-  c()
warming.sims.ac.15.tau <-  c()
warming.sims.acv.15.tau <-  c()
warming.sims.fdv.15.tau <-  c()
warming.sims.dt.15.tau <-  c()
```

Calculate Kendall's tau for each metric:

```{r}
time.15 <- c(1:45)

for(i in 1:num.of.sims){
  constant.sims.mean.15.tau[i] <- cor(constant.sims.mean.15[ , i], time.15, method = "kendall")
  constant.sims.sk.15.tau[i] <- cor(constant.sims.sk.15[ , i], time.15, method = "kendall")
  constant.sims.kt.15.tau[i] <- cor(constant.sims.kt.15[ , i], time.15, method = "kendall")
  constant.sims.var.15.tau[i] <- cor(constant.sims.var.15[ , i], time.15, method = "kendall")
  constant.sims.cv.15.tau[i] <- cor(constant.sims.cv.15[ , i], time.15, method = "kendall")
  constant.sims.id.15.tau[i] <- cor(constant.sims.id.15[ , i], time.15, method = "kendall")
  constant.sims.fdv.15.tau[i] <- cor(constant.sims.fdv.15[ , i], time.15, method = "kendall")
  constant.sims.ac.15.tau[i] <- cor(constant.sims.ac.15[ , i], time.15, method = "kendall")
  constant.sims.acv.15.tau[i] <- cor(constant.sims.acv.15[ , i], time.15, method = "kendall")
  constant.sims.dt.15.tau[i] <- cor(constant.sims.dt.15[ , i], time.15, method = "kendall")
  
  warming.sims.mean.15.tau[i] <- cor(warming.sims.mean.15[ , i], time.15, method = "kendall")
  warming.sims.sk.15.tau[i] <- cor(warming.sims.sk.15[ , i], time.15, method = "kendall")
  warming.sims.kt.15.tau[i] <- cor(warming.sims.kt.15[ , i], time.15, method = "kendall")
  warming.sims.var.15.tau[i] <- cor(warming.sims.var.15[ , i], time.15, method = "kendall")
  warming.sims.cv.15.tau[i] <- cor(warming.sims.cv.15[ , i], time.15, method = "kendall")
  warming.sims.id.15.tau[i] <- cor(warming.sims.id.15[ , i], time.15, method = "kendall")
  warming.sims.fdv.15.tau[i] <- cor(warming.sims.fdv.15[ , i], time.15, method = "kendall")
  warming.sims.ac.15.tau[i] <- cor(warming.sims.ac.15[ , i], time.15, method = "kendall")
  warming.sims.acv.15.tau[i] <- cor(warming.sims.acv.15[ , i], time.15, method = "kendall")
  warming.sims.dt.15.tau[i] <- cor(warming.sims.dt.15[ , i], time.15, method = "kendall")
}
```

Calculating AUC:

```{r}
AUC_var_15 <- calc_auc(warming.sims.var.15.tau, constant.sims.var.15.tau); AUC_var_15
AUC_mean_15 <- calc_auc(warming.sims.mean.15.tau, constant.sims.mean.15.tau); AUC_mean_15
AUC_kt_15 <- calc_auc(warming.sims.kt.15.tau, constant.sims.kt.15.tau); AUC_kt_15
AUC_sk_15 <- calc_auc(warming.sims.sk.15.tau, constant.sims.sk.15.tau); AUC_sk_15
AUC_id_15 <- calc_auc(warming.sims.id.15.tau, constant.sims.id.15.tau); AUC_id_15
AUC_fdv_15 <- calc_auc(warming.sims.fdv.15.tau, constant.sims.fdv.15.tau); AUC_fdv_15
AUC_ac_15 <- calc_auc(warming.sims.ac.15.tau, constant.sims.ac.15.tau); AUC_ac_15
AUC_acv_15 <- calc_auc(warming.sims.acv.15.tau, constant.sims.acv.15.tau); AUC_acv_15
AUC_cv_15 <- calc_auc(warming.sims.cv.15.tau, constant.sims.cv.15.tau); AUC_cv_15
AUC_dt_15 <- calc_auc(warming.sims.dt.15.tau, constant.sims.dt.15.tau); AUC_dt_15
```

## 30 Day Sliding Window

```{r}
constant.sims.var.30 <- constant.sims.var.30[1:30, ]
constant.sims.mean.30 <- constant.sims.mean.30[1:30, ]
constant.sims.cv.30 <- constant.sims.cv.30[1:30, ]
constant.sims.sk.30 <-  constant.sims.sk.30[1:30, ]
constant.sims.kt.30 <-  constant.sims.kt.30[1:30, ]
constant.sims.id.30 <-  constant.sims.id.30[1:30, ]
constant.sims.ac.30 <-  constant.sims.ac.30[1:30, ]
constant.sims.acv.30 <-  constant.sims.acv.30[1:30, ]
constant.sims.fdv.30 <-  constant.sims.fdv.30[1:30, ]
constant.sims.dt.30 <-  constant.sims.dt.30[1:30, ]

warming.sims.var.30 <- warming.sims.var.30[1:30, ]
warming.sims.mean.30 <- warming.sims.mean.30[1:30, ]
warming.sims.cv.30 <- warming.sims.cv.30[1:30, ]
warming.sims.sk.30 <-  warming.sims.sk.30[1:30, ]
warming.sims.kt.30 <-  warming.sims.kt.30[1:30, ]
warming.sims.id.30 <-  warming.sims.id.30[1:30, ]
warming.sims.ac.30 <-  warming.sims.ac.30[1:30, ]
warming.sims.acv.30 <-  warming.sims.acv.30[1:30, ]
warming.sims.fdv.30 <-  warming.sims.fdv.30[1:30, ]
warming.sims.dt.30 <-  warming.sims.dt.30[1:30, ]

constant.sims.var.30.tau <- c()
constant.sims.mean.30.tau <- c()
constant.sims.cv.30.tau <- c()
constant.sims.sk.30.tau <-  c()
constant.sims.kt.30.tau <- c()
constant.sims.id.30.tau <- c()
constant.sims.ac.30.tau <- c()
constant.sims.acv.30.tau <- c()
constant.sims.fdv.30.tau <-  c()
constant.sims.dt.30.tau <- c()

warming.sims.var.30.tau <- c()
warming.sims.mean.30.tau <- c()
warming.sims.cv.30.tau <- c()
warming.sims.sk.30.tau <-  c()
warming.sims.kt.30.tau <- c()
warming.sims.id.30.tau <-  c()
warming.sims.ac.30.tau <-  c()
warming.sims.acv.30.tau <-  c()
warming.sims.fdv.30.tau <-  c()
warming.sims.dt.30.tau <-  c()
```

Calculate Kendall's tau for each metric:

```{r}
time.30 <- c(1:30)

for(i in 1:num.of.sims){
  constant.sims.mean.30.tau[i] <- cor(constant.sims.mean.30[ , i], time.30, method = "kendall")
  constant.sims.sk.30.tau[i] <- cor(constant.sims.sk.30[ , i], time.30, method = "kendall")
  constant.sims.kt.30.tau[i] <- cor(constant.sims.kt.30[ , i], time.30, method = "kendall")
  constant.sims.var.30.tau[i] <- cor(constant.sims.var.30[ , i], time.30, method = "kendall")
  constant.sims.cv.30.tau[i] <- cor(constant.sims.cv.30[ , i], time.30, method = "kendall")
  constant.sims.id.30.tau[i] <- cor(constant.sims.id.30[ , i], time.30, method = "kendall")
  constant.sims.fdv.30.tau[i] <- cor(constant.sims.fdv.30[ , i], time.30, method = "kendall")
  constant.sims.ac.30.tau[i] <- cor(constant.sims.ac.30[ , i], time.30, method = "kendall")
  constant.sims.acv.30.tau[i] <- cor(constant.sims.acv.30[ , i], time.30, method = "kendall")
  constant.sims.dt.30.tau[i] <- cor(constant.sims.dt.30[ , i], time.30, method = "kendall")
  
  warming.sims.mean.30.tau[i] <- cor(warming.sims.mean.30[ , i], time.30, method = "kendall")
  warming.sims.sk.30.tau[i] <- cor(warming.sims.sk.30[ , i], time.30, method = "kendall")
  warming.sims.kt.30.tau[i] <- cor(warming.sims.kt.30[ , i], time.30, method = "kendall")
  warming.sims.var.30.tau[i] <- cor(warming.sims.var.30[ , i], time.30, method = "kendall")
  warming.sims.cv.30.tau[i] <- cor(warming.sims.cv.30[ , i], time.30, method = "kendall")
  warming.sims.id.30.tau[i] <- cor(warming.sims.id.30[ , i], time.30, method = "kendall")
  warming.sims.fdv.30.tau[i] <- cor(warming.sims.fdv.30[ , i], time.30, method = "kendall")
  warming.sims.ac.30.tau[i] <- cor(warming.sims.ac.30[ , i], time.30, method = "kendall")
  warming.sims.acv.30.tau[i] <- cor(warming.sims.acv.30[ , i], time.30, method = "kendall")
  warming.sims.dt.30.tau[i] <- cor(warming.sims.dt.30[ , i], time.30, method = "kendall")
}
```

Calculating AUC:

```{r}
AUC_var_30 <- calc_auc(warming.sims.var.30.tau, constant.sims.var.30.tau); AUC_var_30
AUC_mean_30 <- calc_auc(warming.sims.mean.30.tau, constant.sims.mean.30.tau); AUC_mean_30
AUC_kt_30 <- calc_auc(warming.sims.kt.30.tau, constant.sims.kt.30.tau); AUC_kt_30
AUC_sk_30 <- calc_auc(warming.sims.sk.30.tau, constant.sims.sk.30.tau); AUC_sk_30
AUC_id_30 <- calc_auc(warming.sims.id.30.tau, constant.sims.id.30.tau); AUC_id_30
AUC_fdv_30 <- calc_auc(warming.sims.fdv.30.tau, constant.sims.fdv.30.tau); AUC_fdv_30
AUC_ac_30 <- calc_auc(warming.sims.ac.30.tau, constant.sims.ac.30.tau); AUC_ac_30
AUC_acv_30 <- calc_auc(warming.sims.acv.30.tau, constant.sims.acv.30.tau); AUC_acv_30
AUC_cv_30 <- calc_auc(warming.sims.cv.30.tau, constant.sims.cv.30.tau); AUC_cv_30
AUC_dt_30 <- calc_auc(warming.sims.dt.30.tau, constant.sims.dt.30.tau); AUC_dt_30
```

## Summarize Outputs

```{r}
metrics <- c("mean", "skewness", "kurtosis", "variance", "cv", "id", "first.diff.var", "cor", "cov", "decay.time")

AUC_5 <- c(AUC_mean_5, AUC_sk_5, AUC_kt_5, AUC_var_5, AUC_cv_5, 
           AUC_id_5, AUC_fdv_5, AUC_ac_5 ,AUC_acv_5, AUC_dt_5)
AUC_15 <- c(AUC_mean_15, AUC_sk_15, AUC_kt_15, AUC_var_15, AUC_cv_15, 
           AUC_id_15, AUC_fdv_15, AUC_ac_15 ,AUC_acv_15, AUC_dt_15)
AUC_30 <- c(AUC_mean_30, AUC_sk_30, AUC_kt_30, AUC_var_30, AUC_cv_30, 
           AUC_id_30, AUC_fdv_30, AUC_ac_30 ,AUC_acv_30, AUC_dt_30)
AUC_df <- data.frame(metrics, AUC_5, AUC_15, AUC_30); AUC_df
saveRDS(AUC_df, "Objects/AUC_SmoothSimulatedData_B4.rds")
```

A quick plot!

```{r}
metrics_full <- c("Mean", "Skewness", "Kurtosis", "Variance", "Covariance", "Index of Dispersion", "First Difference Variance", "Autocorrelation", "Autocovariance", "Decay Time")

## Plot 
plot(1:10, AUC_df$AUC_5, type = "p", 
     main = "Smoothed Simulated Data | Bandwidth = 4", xaxt = "none", xlab = "", ylim = c(0.3, 0.7), ylab = "AUC",
     col = "forestgreen", pch = 19)
points(1:10, AUC_df$AUC_15, col = "orchid", pch = 19)
points(1:10, AUC_df$AUC_30, col = "#4B0082", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)
legend("bottomright", 
       legend = c("Sliding Window = 5 Days", "Sliding Window = 15 Days", "Sliding Window = 30 Days"),
       col = c("forestgreen", "orchid", "#4B0082"), pch = 19)

## Saving plot as .png
png(file = "Plots/AUC_SmoothSimulatedData_B4.png", width = 1500, height = 350)
plot(1:10, AUC_df$AUC_5, type = "p", 
     main = "Smoothed Simulated Data | Bandwidth = 4", xaxt = "none", xlab = "", ylim = c(0.3, 0.7), ylab = "AUC",
     col = "forestgreen", pch = 19)
points(1:10, AUC_df$AUC_15, col = "orchid", pch = 19)
points(1:10, AUC_df$AUC_30, col = "#4B0082", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full)
legend("bottomright", 
       legend = c("Sliding Window = 5 Days", "Sliding Window = 15 Days", "Sliding Window = 30 Days"),
       col = c("forestgreen", "orchid", "#4B0082"), pch = 19)
dev.off()
```

# Compare Bandwidths

```{r}
AUC.Smooth.SimDat.B2 <- readRDS("Objects/AUC_SmoothSimulatedData_B2.rds")
AUC.Smooth.SimDat.B3 <- readRDS("Objects/AUC_SmoothSimulatedData_B3.rds")
AUC.Smooth.SimDat.B4 <- readRDS("Objects/AUC_SmoothSimulatedData_B4.rds")

plot(1:10, AUC.Smooth.SimDat.B2$AUC_5, pch = 19, col = alpha("violetred", 0.5),
     main = "Smoothed Simulated Data | Varying Bandwidths", xlab = "", ylab = "AUC", ylim = c(0, 1))
points(1:10, AUC.Smooth.SimDat.B3$AUC_5, pch = 19, col = alpha("coral", 0.5))
points(1:10, AUC.Smooth.SimDat.B4$AUC_5, pch = 19, col = alpha("seagreen", 0.5))
points(1:10, AUC.Smooth.SimDat.B2$AUC_15, pch = 15, col = alpha("violetred", 0.5))
points(1:10, AUC.Smooth.SimDat.B3$AUC_15, pch = 15, col = alpha("coral", 0.5))
points(1:10, AUC.Smooth.SimDat.B4$AUC_15, pch = 15, col = alpha("seagreen", 0.5))
points(1:10, AUC.Smooth.SimDat.B2$AUC_30, pch = 17, col = alpha("violetred", 0.5))
points(1:10, AUC.Smooth.SimDat.B3$AUC_30, pch = 17, col = alpha("coral", 0.5))
points(1:10, AUC.Smooth.SimDat.B4$AUC_30, pch = 17, col = alpha("seagreen", 0.5))
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)
legend("topright", legend = c("Bandwidth = 2", "Bandwidth = 3", "Bandwidth = 4"),
       col = c("violetred", "coral", "seagreen"), pch = 19)
legend("bottomright", 
       legend = c("Sliding Window = 5", "Sliding Window = 15", "Sliding Window = 30"),
       col = c("grey"), pch = c(19, 15, 17))

png(file = "Plots/AUC_SmoothSimulatedData_VaryBandwidth.png", width = 1500, height = 350)
plot(1:10, AUC.Smooth.SimDat.B2$AUC_15, pch = 19, col = alpha("violetred", 0.5),
     main = "Smoothed Simulated Data | Sliding Window = 15", xlab = "", ylab = "AUC", ylim = c(0, 1))
points(1:10, AUC.Smooth.SimDat.B3$AUC_15, pch = 19, col = alpha("coral", 0.5))
points(1:10, AUC.Smooth.SimDat.B4$AUC_15, pch = 19, col = alpha("seagreen", 0.5))
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)
legend("topright", legend = c("Bandwidth = 2", "Bandwidth = 3", "Bandwidth = 4"),
       col = c("violetred", "coral", "seagreen"), pch = 19)
dev.off()
```
