==> 7_EWS-Detrend-EmpData.Rmd <==
# Looking for EWS in Detrended Empirical Data

## Required Packages

```{r, include = FALSE}
require(scales)
require(evobiR)
require(gdata)
require(doBy)
require(ecp)
require(strucchange)
require(TTR)
require(moments)
```

# Empirical Data

The object containing the smoothed empirical data is inaptly named `df`, so go to `5_Detrending.Rmd` to regenerate it, or just use the code below.

```{r}
smooth.emp.b2 <- readRDS("Objects/smooth-emp-b2.rds")
smooth.emp.b3 <- readRDS("Objects/smooth-emp-b3.rds")
smooth.emp.b4 <- readRDS("Objects/smooth-emp-b4.rds")

smooth.emp.b2 <- data.frame(smooth.emp.b2) 
colnames(smooth.emp.b2) <- c("sub1", "sub2", "sub3", "sub4",
                             "sub5", "sub6", "sub7", "sub8")
smooth.emp.b3 <- data.frame(smooth.emp.b3)
colnames(smooth.emp.b3) <- c("sub1", "sub2", "sub3", "sub4",
                             "sub5", "sub6", "sub7", "sub8")
smooth.emp.b4 <- data.frame(smooth.emp.b4)
colnames(smooth.emp.b4) <- c("sub1", "sub2", "sub3", "sub4",
                             "sub5", "sub6", "sub7", "sub8")
```

# BANDWIDTH = 3

# Calculating Statistical Metrics Along Variously-Sized Sliding Windows

## 15 Day Sliding Window

### Standard Deviation, Mean, Skewness, Kurtosis

```{r}
## Standard deviation
pop1.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b3$sub1, window = 5, step = 1)
pop2.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b3$sub2, window = 5, step = 1)
pop3.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b3$sub3, window = 5, step = 1)
pop4.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b3$sub4, window = 5, step = 1)
pop5.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b3$sub5, window = 5, step = 1)
pop6.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b3$sub6, window = 5, step = 1)
pop7.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b3$sub7, window = 5, step = 1)
pop8.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b3$sub8, window = 5, step = 1)

## Mean
pop1.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b3$sub1, window = 5, step = 1)
pop2.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b3$sub2, window = 5, step = 1)
pop3.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b3$sub3, window = 5, step = 1)
pop4.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b3$sub4, window = 5, step = 1)
pop5.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b3$sub5, window = 5, step = 1)
pop6.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b3$sub6, window = 5, step = 1)
pop7.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b3$sub7, window = 5, step = 1)
pop8.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b3$sub8, window = 5, step = 1)

## Skewness
pop1.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b3$sub1, window = 5, step = 1)
pop2.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b3$sub2, window = 5, step = 1)
pop3.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b3$sub3, window = 5, step = 1)
pop4.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b3$sub4, window = 5, step = 1)
pop5.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b3$sub5, window = 5, step = 1)
pop6.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b3$sub6, window = 5, step = 1)
pop7.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b3$sub7, window = 5, step = 1)
pop8.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b3$sub8, window = 5, step = 1)

## Kurtosis
pop1.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b3$sub1, window = 5, step = 1)
pop2.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b3$sub2, window = 5, step = 1)
pop3.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b3$sub3, window = 5, step = 1)
pop4.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b3$sub4, window = 5, step = 1)
pop5.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b3$sub5, window = 5, step = 1)
pop6.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b3$sub6, window = 5, step = 1)
pop7.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b3$sub7, window = 5, step = 1)
pop8.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b3$sub8, window = 5, step = 1)
```

### Variation

```{r}
pop1.var <- pop1.sd^2
pop2.var <- pop2.sd^2
pop3.var <- pop3.sd^2
pop4.var <- pop4.sd^2
pop5.var <- pop5.sd^2
pop6.var <- pop6.sd^2
pop7.var <- pop7.sd^2
pop8.var <- pop8.sd^2
```

### Co-Efficient of Variation

```{r}
pop1.cv <- pop1.sd/(pop1.mean + 1e-10)
pop2.cv <- pop2.sd/(pop2.mean + 1e-10)
pop3.cv <- pop3.sd/(pop3.mean + 1e-10)
pop4.cv <- pop4.sd/(pop4.mean + 1e-10)
pop5.cv <- pop5.sd/(pop5.mean + 1e-10)
pop6.cv <- pop6.sd/(pop6.mean + 1e-10)
pop7.cv <- pop7.sd/(pop7.mean + 1e-10)
pop8.cv <- pop8.sd/(pop8.mean + 1e-10)
```

### Index of Dispersion

```{r}
pop1.id <- pop1.var/(pop1.mean + 1e-10)
pop2.id <- pop2.var/(pop2.mean + 1e-10)
pop3.id <- pop3.var/(pop3.mean + 1e-10)
pop4.id <- pop4.var/(pop4.mean + 1e-10)
pop5.id <- pop5.var/(pop5.mean + 1e-10)
pop6.id <- pop6.var/(pop6.mean + 1e-10)
pop7.id <- pop7.var/(pop7.mean + 1e-10)
pop8.id <- pop8.var/(pop8.mean + 1e-10)
```

### First Difference Variance

```{r}
pop1.fdv <- c()
pop2.fdv <- c()
pop3.fdv <- c()
pop4.fdv <- c()
pop5.fdv <- c()
pop6.fdv <- c()
pop7.fdv <- c()
pop8.fdv <- c()

for(i in 2:35){
pop1.fdv[i - 1] <- pop1.var[i] - pop1.var[i - 1]
pop2.fdv[i - 1] <- pop2.var[i] - pop2.var[i - 1]
pop3.fdv[i - 1] <- pop3.var[i] - pop3.var[i - 1]
pop4.fdv[i - 1] <- pop4.var[i] - pop4.var[i - 1]
pop5.fdv[i - 1] <- pop5.var[i] - pop5.var[i - 1]
pop6.fdv[i - 1] <- pop6.var[i] - pop6.var[i - 1]
pop7.fdv[i - 1] <- pop7.var[i] - pop7.var[i - 1]
pop8.fdv[i - 1] <- pop8.var[i] - pop8.var[i - 1]
}
```

### Autocorrelation and Autocovariance

```{r}
pop1.ac <- c()
pop1.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b3$sub1[c(window.start:window.end)]
  
  pop1.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop1.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}

pop2.ac <- c()
pop2.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b3$sub2[c(window.start:window.end)]
  
  pop2.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop2.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}

pop3.ac <- c()
pop3.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b3$sub3[c(window.start:window.end)]
  
  pop3.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop3.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}


pop4.ac <- c()
pop4.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b3$sub4[c(window.start:window.end)]
  
  pop4.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop4.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}

pop5.ac <- c()
pop5.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b3$sub5[c(window.start:window.end)]
  
  pop5.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop5.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}

pop6.ac <- c()
pop6.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b3$sub6[c(window.start:window.end)]
  
  pop6.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop6.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}


pop7.ac <- c()
pop7.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b3$sub7[c(window.start:window.end)]
  
  pop7.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop7.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}

pop8.ac <- c()
pop8.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b3$sub8[c(window.start:window.end)]
  
  pop8.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop8.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}
```

### Decay Time

```{r}
pop1.dt <- c()
pop2.dt <- c()
pop3.dt <- c()
pop4.dt <- c()
pop5.dt <- c()
pop6.dt <- c()
pop7.dt <- c()
pop8.dt <- c()

for(i in 1:35){
  suppressWarnings(pop1.dt[i] <- -i/log(pop1.ac[i]))
  suppressWarnings(pop2.dt[i] <- -i/log(pop2.ac[i]))
  suppressWarnings(pop3.dt[i] <- -i/log(pop3.ac[i]))
  suppressWarnings(pop4.dt[i] <- -i/log(pop4.ac[i]))
  suppressWarnings(pop5.dt[i] <- -i/log(pop5.ac[i]))
  suppressWarnings(pop6.dt[i] <- -i/log(pop6.ac[i]))
  suppressWarnings(pop7.dt[i] <- -i/log(pop7.ac[i]))
  suppressWarnings(pop8.dt[i] <- -i/log(pop8.ac[i]))
}
```

# Area Under Curve (AUC) Tests

## 15 Day Sliding Window

Calculate Kendall's tau for each metric:

```{r}
time.15 <- seq(1, 15, 1)

constant.mean.tau<- c(cor(pop1.mean[1:15], time.15, method = "kendall"), cor(pop2.mean[1:15], time.15, method = "kendall"),
                      cor(pop3.mean[1:15], time.15, method = "kendall"), cor(pop4.mean[1:15], time.15, method = "kendall"))
warming.mean.tau <- c(cor(pop5.mean[1:15], time.15, method = "kendall"), cor(pop6.mean[1:15], time.15, method = "kendall"),
                      cor(pop7.mean[1:15], time.15, method = "kendall"), cor(pop8.mean[1:15], time.15, method = "kendall"))

constant.sk.tau<- c(cor(pop1.sk[1:15], time.15, method = "kendall"), cor(pop2.sk[1:15], time.15, method = "kendall"),
                          cor(pop3.sk[1:15], time.15, method = "kendall"), cor(pop4.sk[1:15], time.15, method = "kendall"))
warming.sk.tau <- c(cor(pop5.sk[1:15], time.15, method = "kendall"), cor(pop6.sk[1:15], time.15, method = "kendall"),
                          cor(pop7.sk[1:15], time.15, method = "kendall"), cor(pop8.sk[1:15], time.15, method = "kendall"))

constant.kt.tau<- c(cor(pop1.kt[1:15], time.15, method = "kendall"), cor(pop2.kt[1:15], time.15, method = "kendall"),
                          cor(pop3.kt[1:15], time.15, method = "kendall"), cor(pop4.kt[1:15], time.15, method = "kendall"))
warming.kt.tau <- c(cor(pop5.kt[1:15], time.15, method = "kendall"), cor(pop6.kt[1:15], time.15, method = "kendall"),
                          cor(pop7.kt[1:15], time.15, method = "kendall"), cor(pop8.kt[1:15], time.15, method = "kendall"))

constant.var.tau<- c(cor(pop1.var[1:15], time.15, method = "kendall"), cor(pop2.var[1:15], time.15, method = "kendall"),
                          cor(pop3.var[1:15], time.15, method = "kendall"), cor(pop4.var[1:15], time.15, method = "kendall"))
warming.var.tau<- c(cor(pop5.var[1:15], time.15, method = "kendall"), cor(pop6.var[1:15], time.15, method = "kendall"),
                         cor(pop7.var[1:15], time.15, method = "kendall"), cor(pop8.var[1:15], time.15, method = "kendall"))

constant.cv.tau<- c(cor(pop1.cv[1:15], time.15, method = "kendall"), cor(pop2.cv[1:15], time.15, method = "kendall"),
                    cor(pop3.cv[1:15], time.15, method = "kendall"), cor(pop4.cv[1:15], time.15, method = "kendall"))
warming.cv.tau <- c(cor(pop5.cv[1:15], time.15, method = "kendall"), cor(pop6.cv[1:15], time.15, method = "kendall"),
                    cor(pop7.cv[1:15], time.15, method = "kendall"), cor(pop8.cv[1:15], time.15, method = "kendall"))

constant.id.tau<- c(cor(pop1.id[1:15], time.15, method = "kendall"), cor(pop2.id[1:15], time.15, method = "kendall"),
                    cor(pop3.id[1:15], time.15, method = "kendall"), cor(pop4.id[1:15], time.15, method = "kendall"))
warming.id.tau <- c(cor(pop5.id[1:15], time.15, method = "kendall"), cor(pop6.id[1:15], time.15, method = "kendall"),
                    cor(pop7.id[1:15], time.15, method = "kendall"), cor(pop8.id[1:15], time.15, method = "kendall"))

constant.fdv.tau<- c(cor(pop1.fdv[1:15], time.15, method = "kendall"), cor(pop2.fdv[1:15], time.15, method = "kendall"),
                                     cor(pop3.fdv[1:15], time.15, method = "kendall"), cor(pop4.fdv[1:15], time.15, method = "kendall"))
warming.fdv.tau <- c(cor(pop5.fdv[1:15], time.15, method = "kendall"), cor(pop6.fdv[1:15], time.15, method = "kendall"),
                                     cor(pop7.fdv[1:15], time.15, method = "kendall"), cor(pop8.fdv[1:15], time.15, method = "kendall"))

constant.ac.tau<- c(cor(pop1.ac[1:15], time.15, method = "kendall"), cor(pop2.ac[1:15], time.15, method = "kendall"),
                    cor(pop3.ac[1:15], time.15, method = "kendall"), cor(pop4.ac[1:15], time.15, method = "kendall"))
warming.ac.tau <- c(cor(pop5.ac[1:15], time.15, method = "kendall"), cor(pop6.ac[1:15], time.15, method = "kendall"),
                    cor(pop7.ac[1:15], time.15, method = "kendall"), cor(pop8.ac[1:15], time.15, method = "kendall"))

constant.acv.tau<- c(cor(pop1.acv[1:15], time.15, method = "kendall"), cor(pop2.acv[1:15], time.15, method = "kendall"),
                     cor(pop3.acv[1:15], time.15, method = "kendall"), cor(pop4.acv[1:15], time.15, method = "kendall"))
warming.acv.tau <- c(cor(pop5.acv[1:15], time.15, method = "kendall"), cor(pop6.acv[1:15], time.15, method = "kendall"),
                     cor(pop7.acv[1:15], time.15, method = "kendall"), cor(pop8.acv[1:15], time.15, method = "kendall"))

constant.dt.tau<- c(cor(pop1.dt[1:15], time.15, method = "kendall"), cor(pop2.dt[1:15], time.15, method = "kendall"),
                            cor(pop3.dt[1:15], time.15, method = "kendall"), cor(pop4.dt[1:15], time.15, method = "kendall"))
warming.dt.tau <- c(cor(pop5.dt[1:15], time.15, method = "kendall"), cor(pop6.dt[1:15], time.15, method = "kendall"),
                            cor(pop7.dt[1:15], time.15, method = "kendall"), cor(pop8.dt[1:15], time.15, method = "kendall"))
```

Calculating AUC:

```{r}
AUC_mean_data <- calc_auc(warming.mean.tau,constant.mean.tau) ; AUC_mean_data
AUC_sk_data <- calc_auc(warming.sk.tau,constant.sk.tau) ; AUC_sk_data
AUC_kt_data <- calc_auc(warming.kt.tau,constant.kt.tau) ; AUC_kt_data
AUC_var_data <- calc_auc(warming.var.tau,constant.var.tau) ; AUC_var_data
AUC_cv_data <- calc_auc(warming.cv.tau,constant.cv.tau) ; AUC_cv_data
AUC_id_data <- calc_auc(warming.id.tau,constant.id.tau) ; AUC_id_data
AUC_fdv_data <- calc_auc(warming.fdv.tau,constant.fdv.tau) ; AUC_fdv_data
AUC_ac_data <- calc_auc(warming.ac.tau,constant.ac.tau) ; AUC_ac_data
AUC_acv_data <- calc_auc(warming.acv.tau,constant.acv.tau) ; AUC_acv_data
AUC_dt_data <- calc_auc(warming.dt.tau,constant.dt.tau) ; AUC_dt_data
```

## Summarize Outputs

```{r}
metrics <- c("mean", "skewness", "kurtosis", "variance", "cv", "id", "fdv", "ac", "acv", "dt")

AUC_data <- c(AUC_mean_data, 
              AUC_sk_data, 
              AUC_kt_data, 
              AUC_var_data,
              AUC_cv_data,
              AUC_id_data, 
              AUC_fdv_data, 
              AUC_ac_data, 
              AUC_acv_data, 
              AUC_dt_data)

AUC_df <- data.frame(metrics, AUC_data); AUC_df
saveRDS(AUC_df, "Objects/AUC_SmoothEmpiricalData_B3.rds")
```

A plot!

```{r}
metrics_full <- c("Mean", "Skewness", "Kurtosis", "Variance", "Covariance", "Index of Dispersion", "First Difference Variance", "Autocorrelation", "Autocovariance", "Decay Time")

## Plot
plot(1:10, AUC_df$AUC_data, type = "p", 
     main = "Smoothed Empirical Data, All Populations", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "orchid", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)
```

# Excluding Population 5

```{r}
time.15 <- seq(1,15,1)

warming.mean.tau_without5 <- c(cor(pop6.mean[1:15], time.15, method = "kendall"),
                               cor(pop7.mean[1:15], time.15, method = "kendall"), 
                               cor(pop8.mean[1:15], time.15, method = "kendall"))

warming.sk.tau_without5 <- c(cor(pop6.sk[1:15], time.15, method = "kendall"),
                                   cor(pop7.sk[1:15], time.15, method = "kendall"), 
                                   cor(pop8.sk[1:15], time.15, method = "kendall"))


warming.kt.tau_without5 <- c(cor(pop6.kt[1:15], time.15, method = "kendall"),
                                   cor(pop7.kt[1:15], time.15, method = "kendall"), 
                                   cor(pop8.kt[1:15], time.15, method = "kendall"))

warming.var.tau_without5 <- c(cor(pop6.var[1:15], time.15, method = "kendall"),
                                   cor(pop7.var[1:15], time.15, method = "kendall"), 
                                   cor(pop8.var[1:15], time.15, method = "kendall"))

warming.cv.tau_without5 <- c(cor(pop6.cv[1:15], time.15, method = "kendall"),
                             cor(pop7.cv[1:15], time.15, method = "kendall"), 
                             cor(pop8.cv[1:15], time.15, method = "kendall"))

warming.id.tau_without5 <- c(cor(pop6.id[1:15], time.15, method = "kendall"),
                             cor(pop7.id[1:15], time.15, method = "kendall"), 
                             cor(pop8.id[1:15], time.15, method = "kendall"))

warming.ac.tau_without5 <- c(cor(pop6.ac[1:15], time.15, method = "kendall"),
                             cor(pop7.ac[1:15], time.15, method = "kendall"), 
                             cor(pop8.ac[1:15], time.15, method = "kendall"))


warming.acv.tau_without5 <- c(cor(pop6.acv[1:15], time.15, method = "kendall"),
                              cor(pop7.acv[1:15], time.15, method = "kendall"), 
                              cor(pop8.acv[1:15], time.15, method = "kendall"))

warming.fdv.tau_without5 <- c(cor(pop6.fdv[1:15], time.15, method = "kendall"),
                                              cor(pop7.fdv[1:15], time.15, method = "kendall"),
                                              cor(pop8.fdv[1:15], time.15, method = "kendall"))

warming.dt.tau_without5 <- c(cor(pop6.dt[1:15], time.15, method = "kendall"),
                                     cor(pop7.dt[1:15], time.15, method = "kendall"), 
                                     cor(pop8.dt[1:15], time.15, method = "kendall"))
```

## Summarize Outputs

```{r}
AUC_mean_data_without5 <- calc_auc(warming.mean.tau_without5,constant.mean.tau); AUC_mean_data_without5 
AUC_sk_data_without5 <- calc_auc(warming.sk.tau_without5,constant.sk.tau); AUC_sk_data_without5 
AUC_kt_data_without5 <- calc_auc(warming.kt.tau_without5, constant.kt.tau); AUC_kt_data_without5 
AUC_var_data_without5 <- calc_auc(warming.var.tau_without5,constant.var.tau); AUC_var_data_without5 
AUC_cv_data_without5 <- calc_auc(warming.cv.tau_without5,constant.cv.tau); AUC_cv_data_without5
AUC_id_data_without5 <- calc_auc(warming.id.tau_without5,constant.id.tau); AUC_id_data_without5 
AUC_fdv_data_without5 <- calc_auc(warming.fdv.tau_without5,constant.fdv.tau); AUC_fdv_data_without5 
AUC_ac_data_without5 <- calc_auc(warming.ac.tau_without5,constant.ac.tau); AUC_ac_data_without5 
AUC_acv_data_without5 <- calc_auc(warming.acv.tau_without5,constant.acv.tau); AUC_acv_data_without5 
AUC_dt_data_without5 <- calc_auc(warming.dt.tau_without5,constant.dt.tau); AUC_dt_data_without5 
```

```{r}
AUC_data <- c(AUC_mean_data_without5,
              AUC_sk_data_without5,
              AUC_kt_data_without5,
              AUC_var_data_without5,
              AUC_cv_data_without5,
              AUC_id_data_without5,
              AUC_fdv_data_without5,
              AUC_ac_data_without5,
              AUC_acv_data_without5,
              AUC_dt_data_without5)

AUC_df_without5 <- data.frame(metrics, AUC_data); AUC_df_without5
saveRDS(AUC_df_without5, "Objects/AUC_SmoothEmpiricalData_ExPop5_B3.rds")
```

A plot!

```{r}
## Plots
plot(1:10, AUC_df_without5$AUC_data, type = "p", 
     main = "Smoothed Empirical Data, Excluding Population 5", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "red3", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)

plot(1:10, AUC_df$AUC_data, type = "p", 
     main = "Smoothed Empirical Data", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "orchid", pch = 19)
points(1:10, AUC_df_without5$AUC_data, col = "red3", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)
legend("topright", legend = c("All Populations", "Excluding Population 5"),
       col = c("orchid", "red3"), pch = 19)

## Saving plot as .png
png(file = "Plots/AUC_SmoothEmpiricalData_B3.png", width = 1500, height = 350)
plot(1:10, AUC_df$AUC_data, type = "p", 
     main = "Smoothed Empirical Data", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "orchid", pch = 19)
points(1:10, AUC_df_without5$AUC_data, col = "red3", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full)
legend("topright", legend = c("All Populations", "Excluding Population 5"),
       col = c("orchid", "red3"), pch = 19)
dev.off()
```

# BANDWIDTH = 2

# Calculating Statistical Metrics Along Variously-Sized Sliding Windows

## 15 Day Sliding Window

### Standard Deviation, Mean, Skewness, Kurtosis

```{r}
## Standard deviation
pop1.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b2$sub1, window = 5, step = 1)
pop2.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b2$sub2, window = 5, step = 1)
pop3.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b2$sub3, window = 5, step = 1)
pop4.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b2$sub4, window = 5, step = 1)
pop5.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b2$sub5, window = 5, step = 1)
pop6.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b2$sub6, window = 5, step = 1)
pop7.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b2$sub7, window = 5, step = 1)
pop8.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b2$sub8, window = 5, step = 1)

## Mean
pop1.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b2$sub1, window = 5, step = 1)
pop2.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b2$sub2, window = 5, step = 1)
pop3.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b2$sub3, window = 5, step = 1)
pop4.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b2$sub4, window = 5, step = 1)
pop5.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b2$sub5, window = 5, step = 1)
pop6.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b2$sub6, window = 5, step = 1)
pop7.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b2$sub7, window = 5, step = 1)
pop8.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b2$sub8, window = 5, step = 1)

## Skewness
pop1.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b2$sub1, window = 5, step = 1)
pop2.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b2$sub2, window = 5, step = 1)
pop3.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b2$sub3, window = 5, step = 1)
pop4.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b2$sub4, window = 5, step = 1)
pop5.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b2$sub5, window = 5, step = 1)
pop6.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b2$sub6, window = 5, step = 1)
pop7.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b2$sub7, window = 5, step = 1)
pop8.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b2$sub8, window = 5, step = 1)

## Kurtosis
pop1.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b2$sub1, window = 5, step = 1)
pop2.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b2$sub2, window = 5, step = 1)
pop3.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b2$sub3, window = 5, step = 1)
pop4.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b2$sub4, window = 5, step = 1)
pop5.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b2$sub5, window = 5, step = 1)
pop6.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b2$sub6, window = 5, step = 1)
pop7.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b2$sub7, window = 5, step = 1)
pop8.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b2$sub8, window = 5, step = 1)
```

### Variation

```{r}
pop1.var <- pop1.sd^2
pop2.var <- pop2.sd^2
pop3.var <- pop3.sd^2
pop4.var <- pop4.sd^2
pop5.var <- pop5.sd^2
pop6.var <- pop6.sd^2
pop7.var <- pop7.sd^2
pop8.var <- pop8.sd^2
```

### Co-Efficient of Variation

```{r}
pop1.cv <- pop1.sd/(pop1.mean + 1e-10)
pop2.cv <- pop2.sd/(pop2.mean + 1e-10)
pop3.cv <- pop3.sd/(pop3.mean + 1e-10)
pop4.cv <- pop4.sd/(pop4.mean + 1e-10)
pop5.cv <- pop5.sd/(pop5.mean + 1e-10)
pop6.cv <- pop6.sd/(pop6.mean + 1e-10)
pop7.cv <- pop7.sd/(pop7.mean + 1e-10)
pop8.cv <- pop8.sd/(pop8.mean + 1e-10)
```

### Index of Dispersion

```{r}
pop1.id <- pop1.var/(pop1.mean + 1e-10)
pop2.id <- pop2.var/(pop2.mean + 1e-10)
pop3.id <- pop3.var/(pop3.mean + 1e-10)
pop4.id <- pop4.var/(pop4.mean + 1e-10)
pop5.id <- pop5.var/(pop5.mean + 1e-10)
pop6.id <- pop6.var/(pop6.mean + 1e-10)
pop7.id <- pop7.var/(pop7.mean + 1e-10)
pop8.id <- pop8.var/(pop8.mean + 1e-10)
```

### First Difference Variance

```{r}
pop1.fdv <- c()
pop2.fdv <- c()
pop3.fdv <- c()
pop4.fdv <- c()
pop5.fdv <- c()
pop6.fdv <- c()
pop7.fdv <- c()
pop8.fdv <- c()

for(i in 2:35){
pop1.fdv[i - 1] <- pop1.var[i] - pop1.var[i - 1]
pop2.fdv[i - 1] <- pop2.var[i] - pop2.var[i - 1]
pop3.fdv[i - 1] <- pop3.var[i] - pop3.var[i - 1]
pop4.fdv[i - 1] <- pop4.var[i] - pop4.var[i - 1]
pop5.fdv[i - 1] <- pop5.var[i] - pop5.var[i - 1]
pop6.fdv[i - 1] <- pop6.var[i] - pop6.var[i - 1]
pop7.fdv[i - 1] <- pop7.var[i] - pop7.var[i - 1]
pop8.fdv[i - 1] <- pop8.var[i] - pop8.var[i - 1]
}
```

### Autocorrelation and Autocovariance

```{r}
pop1.ac <- c()
pop1.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b2$sub1[c(window.start:window.end)]
  
  pop1.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop1.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}

pop2.ac <- c()
pop2.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b2$sub2[c(window.start:window.end)]
  
  pop2.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop2.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}

pop3.ac <- c()
pop3.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b2$sub3[c(window.start:window.end)]
  
  pop3.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop3.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}


pop4.ac <- c()
pop4.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b2$sub4[c(window.start:window.end)]
  
  pop4.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop4.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}

pop5.ac <- c()
pop5.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b2$sub5[c(window.start:window.end)]
  
  pop5.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop5.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}

pop6.ac <- c()
pop6.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b2$sub6[c(window.start:window.end)]
  
  pop6.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop6.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}


pop7.ac <- c()
pop7.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b2$sub7[c(window.start:window.end)]
  
  pop7.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop7.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}

pop8.ac <- c()
pop8.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b2$sub8[c(window.start:window.end)]
  
  pop8.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop8.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}
```

### Decay Time

```{r}
pop1.dt <- c()
pop2.dt <- c()
pop3.dt <- c()
pop4.dt <- c()
pop5.dt <- c()
pop6.dt <- c()
pop7.dt <- c()
pop8.dt <- c()

for(i in 1:35){
  suppressWarnings(pop1.dt[i] <- -i/log(pop1.ac[i]))
  suppressWarnings(pop2.dt[i] <- -i/log(pop2.ac[i]))
  suppressWarnings(pop3.dt[i] <- -i/log(pop3.ac[i]))
  suppressWarnings(pop4.dt[i] <- -i/log(pop4.ac[i]))
  suppressWarnings(pop5.dt[i] <- -i/log(pop5.ac[i]))
  suppressWarnings(pop6.dt[i] <- -i/log(pop6.ac[i]))
  suppressWarnings(pop7.dt[i] <- -i/log(pop7.ac[i]))
  suppressWarnings(pop8.dt[i] <- -i/log(pop8.ac[i]))
}
```

# Area Under Curve (AUC) Tests

## 15 Day Sliding Window

Calculate Kendall's tau for each metric:

```{r}
time.15 <- seq(1, 15, 1)

constant.mean.tau<- c(cor(pop1.mean[1:15], time.15, method = "kendall"), cor(pop2.mean[1:15], time.15, method = "kendall"),
                      cor(pop3.mean[1:15], time.15, method = "kendall"), cor(pop4.mean[1:15], time.15, method = "kendall"))
warming.mean.tau <- c(cor(pop5.mean[1:15], time.15, method = "kendall"), cor(pop6.mean[1:15], time.15, method = "kendall"),
                      cor(pop7.mean[1:15], time.15, method = "kendall"), cor(pop8.mean[1:15], time.15, method = "kendall"))

constant.sk.tau<- c(cor(pop1.sk[1:15], time.15, method = "kendall"), cor(pop2.sk[1:15], time.15, method = "kendall"),
                          cor(pop3.sk[1:15], time.15, method = "kendall"), cor(pop4.sk[1:15], time.15, method = "kendall"))
warming.sk.tau <- c(cor(pop5.sk[1:15], time.15, method = "kendall"), cor(pop6.sk[1:15], time.15, method = "kendall"),
                          cor(pop7.sk[1:15], time.15, method = "kendall"), cor(pop8.sk[1:15], time.15, method = "kendall"))

constant.kt.tau<- c(cor(pop1.kt[1:15], time.15, method = "kendall"), cor(pop2.kt[1:15], time.15, method = "kendall"),
                          cor(pop3.kt[1:15], time.15, method = "kendall"), cor(pop4.kt[1:15], time.15, method = "kendall"))
warming.kt.tau <- c(cor(pop5.kt[1:15], time.15, method = "kendall"), cor(pop6.kt[1:15], time.15, method = "kendall"),
                          cor(pop7.kt[1:15], time.15, method = "kendall"), cor(pop8.kt[1:15], time.15, method = "kendall"))

constant.var.tau<- c(cor(pop1.var[1:15], time.15, method = "kendall"), cor(pop2.var[1:15], time.15, method = "kendall"),
                          cor(pop3.var[1:15], time.15, method = "kendall"), cor(pop4.var[1:15], time.15, method = "kendall"))
warming.var.tau<- c(cor(pop5.var[1:15], time.15, method = "kendall"), cor(pop6.var[1:15], time.15, method = "kendall"),
                         cor(pop7.var[1:15], time.15, method = "kendall"), cor(pop8.var[1:15], time.15, method = "kendall"))

constant.cv.tau<- c(cor(pop1.cv[1:15], time.15, method = "kendall"), cor(pop2.cv[1:15], time.15, method = "kendall"),
                    cor(pop3.cv[1:15], time.15, method = "kendall"), cor(pop4.cv[1:15], time.15, method = "kendall"))
warming.cv.tau <- c(cor(pop5.cv[1:15], time.15, method = "kendall"), cor(pop6.cv[1:15], time.15, method = "kendall"),
                    cor(pop7.cv[1:15], time.15, method = "kendall"), cor(pop8.cv[1:15], time.15, method = "kendall"))

constant.id.tau<- c(cor(pop1.id[1:15], time.15, method = "kendall"), cor(pop2.id[1:15], time.15, method = "kendall"),
                    cor(pop3.id[1:15], time.15, method = "kendall"), cor(pop4.id[1:15], time.15, method = "kendall"))
warming.id.tau <- c(cor(pop5.id[1:15], time.15, method = "kendall"), cor(pop6.id[1:15], time.15, method = "kendall"),
                    cor(pop7.id[1:15], time.15, method = "kendall"), cor(pop8.id[1:15], time.15, method = "kendall"))

constant.fdv.tau<- c(cor(pop1.fdv[1:15], time.15, method = "kendall"), cor(pop2.fdv[1:15], time.15, method = "kendall"),
                                     cor(pop3.fdv[1:15], time.15, method = "kendall"), cor(pop4.fdv[1:15], time.15, method = "kendall"))
warming.fdv.tau <- c(cor(pop5.fdv[1:15], time.15, method = "kendall"), cor(pop6.fdv[1:15], time.15, method = "kendall"),
                                     cor(pop7.fdv[1:15], time.15, method = "kendall"), cor(pop8.fdv[1:15], time.15, method = "kendall"))

constant.ac.tau<- c(cor(pop1.ac[1:15], time.15, method = "kendall"), cor(pop2.ac[1:15], time.15, method = "kendall"),
                    cor(pop3.ac[1:15], time.15, method = "kendall"), cor(pop4.ac[1:15], time.15, method = "kendall"))
warming.ac.tau <- c(cor(pop5.ac[1:15], time.15, method = "kendall"), cor(pop6.ac[1:15], time.15, method = "kendall"),
                    cor(pop7.ac[1:15], time.15, method = "kendall"), cor(pop8.ac[1:15], time.15, method = "kendall"))

constant.acv.tau<- c(cor(pop1.acv[1:15], time.15, method = "kendall"), cor(pop2.acv[1:15], time.15, method = "kendall"),
                     cor(pop3.acv[1:15], time.15, method = "kendall"), cor(pop4.acv[1:15], time.15, method = "kendall"))
warming.acv.tau <- c(cor(pop5.acv[1:15], time.15, method = "kendall"), cor(pop6.acv[1:15], time.15, method = "kendall"),
                     cor(pop7.acv[1:15], time.15, method = "kendall"), cor(pop8.acv[1:15], time.15, method = "kendall"))

constant.dt.tau<- c(cor(pop1.dt[1:15], time.15, method = "kendall"), cor(pop2.dt[1:15], time.15, method = "kendall"),
                            cor(pop3.dt[1:15], time.15, method = "kendall"), cor(pop4.dt[1:15], time.15, method = "kendall"))
warming.dt.tau <- c(cor(pop5.dt[1:15], time.15, method = "kendall"), cor(pop6.dt[1:15], time.15, method = "kendall"),
                            cor(pop7.dt[1:15], time.15, method = "kendall"), cor(pop8.dt[1:15], time.15, method = "kendall"))
```

Calculating AUC:

```{r}
AUC_mean_data <- calc_auc(warming.mean.tau,constant.mean.tau) ; AUC_mean_data
AUC_sk_data <- calc_auc(warming.sk.tau,constant.sk.tau) ; AUC_sk_data
AUC_kt_data <- calc_auc(warming.kt.tau,constant.kt.tau) ; AUC_kt_data
AUC_var_data <- calc_auc(warming.var.tau,constant.var.tau) ; AUC_var_data
AUC_cv_data <- calc_auc(warming.cv.tau,constant.cv.tau) ; AUC_cv_data
AUC_id_data <- calc_auc(warming.id.tau,constant.id.tau) ; AUC_id_data
AUC_fdv_data <- calc_auc(warming.fdv.tau,constant.fdv.tau) ; AUC_fdv_data
AUC_ac_data <- calc_auc(warming.ac.tau,constant.ac.tau) ; AUC_ac_data
AUC_acv_data <- calc_auc(warming.acv.tau,constant.acv.tau) ; AUC_acv_data
AUC_dt_data <- calc_auc(warming.dt.tau,constant.dt.tau) ; AUC_dt_data
```

## Summarize Outputs

```{r}
metrics <- c("mean", "skewness", "kurtosis", "variance", "cv", "id", "fdv", "ac", "acv", "dt")

AUC_data <- c(AUC_mean_data, 
              AUC_sk_data, 
              AUC_kt_data, 
              AUC_var_data,
              AUC_cv_data,
              AUC_id_data, 
              AUC_fdv_data, 
              AUC_ac_data, 
              AUC_acv_data, 
              AUC_dt_data)

AUC_df <- data.frame(metrics, AUC_data); AUC_df
saveRDS(AUC_df, "Objects/AUC_SmoothEmpiricalData_B2.rds")
```

A plot!

```{r}
metrics_full <- c("Mean", "Skewness", "Kurtosis", "Variance", "Covariance", "Index of Dispersion", "First Difference Variance", "Autocorrelation", "Autocovariance", "Decay Time")

## Plot
plot(1:10, AUC_df$AUC_data, type = "p", 
     main = "Smoothed Empirical Data, All Populations", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "orchid", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)
```

# Excluding Population 5

```{r}
time.15 <- seq(1,15,1)

warming.mean.tau_without5 <- c(cor(pop6.mean[1:15], time.15, method = "kendall"),
                               cor(pop7.mean[1:15], time.15, method = "kendall"), 
                               cor(pop8.mean[1:15], time.15, method = "kendall"))

warming.sk.tau_without5 <- c(cor(pop6.sk[1:15], time.15, method = "kendall"),
                                   cor(pop7.sk[1:15], time.15, method = "kendall"), 
                                   cor(pop8.sk[1:15], time.15, method = "kendall"))


warming.kt.tau_without5 <- c(cor(pop6.kt[1:15], time.15, method = "kendall"),
                                   cor(pop7.kt[1:15], time.15, method = "kendall"), 
                                   cor(pop8.kt[1:15], time.15, method = "kendall"))

warming.var.tau_without5 <- c(cor(pop6.var[1:15], time.15, method = "kendall"),
                                   cor(pop7.var[1:15], time.15, method = "kendall"), 
                                   cor(pop8.var[1:15], time.15, method = "kendall"))

warming.cv.tau_without5 <- c(cor(pop6.cv[1:15], time.15, method = "kendall"),
                             cor(pop7.cv[1:15], time.15, method = "kendall"), 
                             cor(pop8.cv[1:15], time.15, method = "kendall"))

warming.id.tau_without5 <- c(cor(pop6.id[1:15], time.15, method = "kendall"),
                             cor(pop7.id[1:15], time.15, method = "kendall"), 
                             cor(pop8.id[1:15], time.15, method = "kendall"))

warming.ac.tau_without5 <- c(cor(pop6.ac[1:15], time.15, method = "kendall"),
                             cor(pop7.ac[1:15], time.15, method = "kendall"), 
                             cor(pop8.ac[1:15], time.15, method = "kendall"))


warming.acv.tau_without5 <- c(cor(pop6.acv[1:15], time.15, method = "kendall"),
                              cor(pop7.acv[1:15], time.15, method = "kendall"), 
                              cor(pop8.acv[1:15], time.15, method = "kendall"))

warming.fdv.tau_without5 <- c(cor(pop6.fdv[1:15], time.15, method = "kendall"),
                                              cor(pop7.fdv[1:15], time.15, method = "kendall"),
                                              cor(pop8.fdv[1:15], time.15, method = "kendall"))

warming.dt.tau_without5 <- c(cor(pop6.dt[1:15], time.15, method = "kendall"),
                                     cor(pop7.dt[1:15], time.15, method = "kendall"), 
                                     cor(pop8.dt[1:15], time.15, method = "kendall"))
```

## Summarize Outputs

```{r}
AUC_mean_data_without5 <- calc_auc(warming.mean.tau_without5,constant.mean.tau); AUC_mean_data_without5 
AUC_sk_data_without5 <- calc_auc(warming.sk.tau_without5,constant.sk.tau); AUC_sk_data_without5 
AUC_kt_data_without5 <- calc_auc(warming.kt.tau_without5, constant.kt.tau); AUC_kt_data_without5 
AUC_var_data_without5 <- calc_auc(warming.var.tau_without5,constant.var.tau); AUC_var_data_without5 
AUC_cv_data_without5 <- calc_auc(warming.cv.tau_without5,constant.cv.tau); AUC_cv_data_without5
AUC_id_data_without5 <- calc_auc(warming.id.tau_without5,constant.id.tau); AUC_id_data_without5 
AUC_fdv_data_without5 <- calc_auc(warming.fdv.tau_without5,constant.fdv.tau); AUC_fdv_data_without5 
AUC_ac_data_without5 <- calc_auc(warming.ac.tau_without5,constant.ac.tau); AUC_ac_data_without5 
AUC_acv_data_without5 <- calc_auc(warming.acv.tau_without5,constant.acv.tau); AUC_acv_data_without5 
AUC_dt_data_without5 <- calc_auc(warming.dt.tau_without5,constant.dt.tau); AUC_dt_data_without5 
```

```{r}
AUC_data <- c(AUC_mean_data_without5,
              AUC_sk_data_without5,
              AUC_kt_data_without5,
              AUC_var_data_without5,
              AUC_cv_data_without5,
              AUC_id_data_without5,
              AUC_fdv_data_without5,
              AUC_ac_data_without5,
              AUC_acv_data_without5,
              AUC_dt_data_without5)

AUC_df_without5 <- data.frame(metrics, AUC_data); AUC_df_without5
saveRDS(AUC_df_without5, "Objects/AUC_SmoothEmpiricalData_ExPop5_B2.rds")
```

A plot!

```{r}
## Plots
plot(1:10, AUC_df_without5$AUC_data, type = "p", 
     main = "Smoothed Empirical Data, Excluding Population 5", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "red3", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)

plot(1:10, AUC_df$AUC_data, type = "p", 
     main = "Smoothed Empirical Data", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "orchid", pch = 19)
points(1:10, AUC_df_without5$AUC_data, col = "red3", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)
legend("topright", legend = c("All Populations", "Excluding Population 5"),
       col = c("orchid", "red3"), pch = 19)

## Saving plot as .png
png(file = "Plots/AUC_SmoothEmpiricalData_B2.png", width = 1500, height = 350)
plot(1:10, AUC_df$AUC_data, type = "p", 
     main = "Smoothed Empirical Data", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "orchid", pch = 19)
points(1:10, AUC_df_without5$AUC_data, col = "red3", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full)
legend("topright", legend = c("All Populations", "Excluding Population 5"),
       col = c("orchid", "red3"), pch = 19)
dev.off()
```

# BANDWIDTH = 4

# Calculating Statistical Metrics Along Variously-Sized Sliding Windows

## 15 Day Sliding Window

### Standard Deviation, Mean, Skewness, Kurtosis

```{r}
## Standard deviation
pop1.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b4$sub1, window = 5, step = 1)
pop2.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b4$sub2, window = 5, step = 1)
pop3.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b4$sub3, window = 5, step = 1)
pop4.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b4$sub4, window = 5, step = 1)
pop5.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b4$sub5, window = 5, step = 1)
pop6.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b4$sub6, window = 5, step = 1)
pop7.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b4$sub7, window = 5, step = 1)
pop8.sd <- SlidingWindow(FUN = "sd", data = smooth.emp.b4$sub8, window = 5, step = 1)

## Mean
pop1.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b4$sub1, window = 5, step = 1)
pop2.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b4$sub2, window = 5, step = 1)
pop3.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b4$sub3, window = 5, step = 1)
pop4.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b4$sub4, window = 5, step = 1)
pop5.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b4$sub5, window = 5, step = 1)
pop6.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b4$sub6, window = 5, step = 1)
pop7.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b4$sub7, window = 5, step = 1)
pop8.mean <- SlidingWindow(FUN = "mean", data = smooth.emp.b4$sub8, window = 5, step = 1)

## Skewness
pop1.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b4$sub1, window = 5, step = 1)
pop2.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b4$sub2, window = 5, step = 1)
pop3.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b4$sub3, window = 5, step = 1)
pop4.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b4$sub4, window = 5, step = 1)
pop5.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b4$sub5, window = 5, step = 1)
pop6.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b4$sub6, window = 5, step = 1)
pop7.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b4$sub7, window = 5, step = 1)
pop8.sk <- SlidingWindow(FUN = "skewness", data = smooth.emp.b4$sub8, window = 5, step = 1)

## Kurtosis
pop1.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b4$sub1, window = 5, step = 1)
pop2.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b4$sub2, window = 5, step = 1)
pop3.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b4$sub3, window = 5, step = 1)
pop4.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b4$sub4, window = 5, step = 1)
pop5.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b4$sub5, window = 5, step = 1)
pop6.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b4$sub6, window = 5, step = 1)
pop7.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b4$sub7, window = 5, step = 1)
pop8.kt <- SlidingWindow(FUN = "kurtosis", data = smooth.emp.b4$sub8, window = 5, step = 1)
```

### Variation

```{r}
pop1.var <- pop1.sd^2
pop2.var <- pop2.sd^2
pop3.var <- pop3.sd^2
pop4.var <- pop4.sd^2
pop5.var <- pop5.sd^2
pop6.var <- pop6.sd^2
pop7.var <- pop7.sd^2
pop8.var <- pop8.sd^2
```

### Co-Efficient of Variation

```{r}
pop1.cv <- pop1.sd/(pop1.mean + 1e-10)
pop2.cv <- pop2.sd/(pop2.mean + 1e-10)
pop3.cv <- pop3.sd/(pop3.mean + 1e-10)
pop4.cv <- pop4.sd/(pop4.mean + 1e-10)
pop5.cv <- pop5.sd/(pop5.mean + 1e-10)
pop6.cv <- pop6.sd/(pop6.mean + 1e-10)
pop7.cv <- pop7.sd/(pop7.mean + 1e-10)
pop8.cv <- pop8.sd/(pop8.mean + 1e-10)
```

### Index of Dispersion

```{r}
pop1.id <- pop1.var/(pop1.mean + 1e-10)
pop2.id <- pop2.var/(pop2.mean + 1e-10)
pop3.id <- pop3.var/(pop3.mean + 1e-10)
pop4.id <- pop4.var/(pop4.mean + 1e-10)
pop5.id <- pop5.var/(pop5.mean + 1e-10)
pop6.id <- pop6.var/(pop6.mean + 1e-10)
pop7.id <- pop7.var/(pop7.mean + 1e-10)
pop8.id <- pop8.var/(pop8.mean + 1e-10)
```

### First Difference Variance

```{r}
pop1.fdv <- c()
pop2.fdv <- c()
pop3.fdv <- c()
pop4.fdv <- c()
pop5.fdv <- c()
pop6.fdv <- c()
pop7.fdv <- c()
pop8.fdv <- c()

for(i in 2:35){
pop1.fdv[i - 1] <- pop1.var[i] - pop1.var[i - 1]
pop2.fdv[i - 1] <- pop2.var[i] - pop2.var[i - 1]
pop3.fdv[i - 1] <- pop3.var[i] - pop3.var[i - 1]
pop4.fdv[i - 1] <- pop4.var[i] - pop4.var[i - 1]
pop5.fdv[i - 1] <- pop5.var[i] - pop5.var[i - 1]
pop6.fdv[i - 1] <- pop6.var[i] - pop6.var[i - 1]
pop7.fdv[i - 1] <- pop7.var[i] - pop7.var[i - 1]
pop8.fdv[i - 1] <- pop8.var[i] - pop8.var[i - 1]
}
```

### Autocorrelation and Autocovariance

```{r}
pop1.ac <- c()
pop1.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b4$sub1[c(window.start:window.end)]
  
  pop1.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop1.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}

pop2.ac <- c()
pop2.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b4$sub2[c(window.start:window.end)]
  
  pop2.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop2.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}

pop3.ac <- c()
pop3.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b4$sub3[c(window.start:window.end)]
  
  pop3.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop3.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}


pop4.ac <- c()
pop4.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b4$sub4[c(window.start:window.end)]
  
  pop4.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop4.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}

pop5.ac <- c()
pop5.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b4$sub5[c(window.start:window.end)]
  
  pop5.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop5.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}

pop6.ac <- c()
pop6.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b4$sub6[c(window.start:window.end)]
  
  pop6.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop6.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}


pop7.ac <- c()
pop7.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b4$sub7[c(window.start:window.end)]
  
  pop7.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop7.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}

pop8.ac <- c()
pop8.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  
  segment <- smooth.emp.b4$sub8[c(window.start:window.end)]
  
  pop8.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop8.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}
```

### Decay Time

```{r}
pop1.dt <- c()
pop2.dt <- c()
pop3.dt <- c()
pop4.dt <- c()
pop5.dt <- c()
pop6.dt <- c()
pop7.dt <- c()
pop8.dt <- c()

for(i in 1:35){
  suppressWarnings(pop1.dt[i] <- -i/log(pop1.ac[i]))
  suppressWarnings(pop2.dt[i] <- -i/log(pop2.ac[i]))
  suppressWarnings(pop3.dt[i] <- -i/log(pop3.ac[i]))
  suppressWarnings(pop4.dt[i] <- -i/log(pop4.ac[i]))
  suppressWarnings(pop5.dt[i] <- -i/log(pop5.ac[i]))
  suppressWarnings(pop6.dt[i] <- -i/log(pop6.ac[i]))
  suppressWarnings(pop7.dt[i] <- -i/log(pop7.ac[i]))
  suppressWarnings(pop8.dt[i] <- -i/log(pop8.ac[i]))
}
```

# Area Under Curve (AUC) Tests

## 15 Day Sliding Window

Calculate Kendall's tau for each metric:

```{r}
time.15 <- seq(1, 15, 1)

constant.mean.tau<- c(cor(pop1.mean[1:15], time.15, method = "kendall"), cor(pop2.mean[1:15], time.15, method = "kendall"),
                      cor(pop3.mean[1:15], time.15, method = "kendall"), cor(pop4.mean[1:15], time.15, method = "kendall"))
warming.mean.tau <- c(cor(pop5.mean[1:15], time.15, method = "kendall"), cor(pop6.mean[1:15], time.15, method = "kendall"),
                      cor(pop7.mean[1:15], time.15, method = "kendall"), cor(pop8.mean[1:15], time.15, method = "kendall"))

constant.sk.tau<- c(cor(pop1.sk[1:15], time.15, method = "kendall"), cor(pop2.sk[1:15], time.15, method = "kendall"),
                          cor(pop3.sk[1:15], time.15, method = "kendall"), cor(pop4.sk[1:15], time.15, method = "kendall"))
warming.sk.tau <- c(cor(pop5.sk[1:15], time.15, method = "kendall"), cor(pop6.sk[1:15], time.15, method = "kendall"),
                          cor(pop7.sk[1:15], time.15, method = "kendall"), cor(pop8.sk[1:15], time.15, method = "kendall"))

constant.kt.tau<- c(cor(pop1.kt[1:15], time.15, method = "kendall"), cor(pop2.kt[1:15], time.15, method = "kendall"),
                          cor(pop3.kt[1:15], time.15, method = "kendall"), cor(pop4.kt[1:15], time.15, method = "kendall"))
warming.kt.tau <- c(cor(pop5.kt[1:15], time.15, method = "kendall"), cor(pop6.kt[1:15], time.15, method = "kendall"),
                          cor(pop7.kt[1:15], time.15, method = "kendall"), cor(pop8.kt[1:15], time.15, method = "kendall"))

constant.var.tau<- c(cor(pop1.var[1:15], time.15, method = "kendall"), cor(pop2.var[1:15], time.15, method = "kendall"),
                          cor(pop3.var[1:15], time.15, method = "kendall"), cor(pop4.var[1:15], time.15, method = "kendall"))
warming.var.tau<- c(cor(pop5.var[1:15], time.15, method = "kendall"), cor(pop6.var[1:15], time.15, method = "kendall"),
                         cor(pop7.var[1:15], time.15, method = "kendall"), cor(pop8.var[1:15], time.15, method = "kendall"))

constant.cv.tau<- c(cor(pop1.cv[1:15], time.15, method = "kendall"), cor(pop2.cv[1:15], time.15, method = "kendall"),
                    cor(pop3.cv[1:15], time.15, method = "kendall"), cor(pop4.cv[1:15], time.15, method = "kendall"))
warming.cv.tau <- c(cor(pop5.cv[1:15], time.15, method = "kendall"), cor(pop6.cv[1:15], time.15, method = "kendall"),
                    cor(pop7.cv[1:15], time.15, method = "kendall"), cor(pop8.cv[1:15], time.15, method = "kendall"))

constant.id.tau<- c(cor(pop1.id[1:15], time.15, method = "kendall"), cor(pop2.id[1:15], time.15, method = "kendall"),
                    cor(pop3.id[1:15], time.15, method = "kendall"), cor(pop4.id[1:15], time.15, method = "kendall"))
warming.id.tau <- c(cor(pop5.id[1:15], time.15, method = "kendall"), cor(pop6.id[1:15], time.15, method = "kendall"),
                    cor(pop7.id[1:15], time.15, method = "kendall"), cor(pop8.id[1:15], time.15, method = "kendall"))

constant.fdv.tau<- c(cor(pop1.fdv[1:15], time.15, method = "kendall"), cor(pop2.fdv[1:15], time.15, method = "kendall"),
                                     cor(pop3.fdv[1:15], time.15, method = "kendall"), cor(pop4.fdv[1:15], time.15, method = "kendall"))
warming.fdv.tau <- c(cor(pop5.fdv[1:15], time.15, method = "kendall"), cor(pop6.fdv[1:15], time.15, method = "kendall"),
                                     cor(pop7.fdv[1:15], time.15, method = "kendall"), cor(pop8.fdv[1:15], time.15, method = "kendall"))

constant.ac.tau<- c(cor(pop1.ac[1:15], time.15, method = "kendall"), cor(pop2.ac[1:15], time.15, method = "kendall"),
                    cor(pop3.ac[1:15], time.15, method = "kendall"), cor(pop4.ac[1:15], time.15, method = "kendall"))
warming.ac.tau <- c(cor(pop5.ac[1:15], time.15, method = "kendall"), cor(pop6.ac[1:15], time.15, method = "kendall"),
                    cor(pop7.ac[1:15], time.15, method = "kendall"), cor(pop8.ac[1:15], time.15, method = "kendall"))

constant.acv.tau<- c(cor(pop1.acv[1:15], time.15, method = "kendall"), cor(pop2.acv[1:15], time.15, method = "kendall"),
                     cor(pop3.acv[1:15], time.15, method = "kendall"), cor(pop4.acv[1:15], time.15, method = "kendall"))
warming.acv.tau <- c(cor(pop5.acv[1:15], time.15, method = "kendall"), cor(pop6.acv[1:15], time.15, method = "kendall"),
                     cor(pop7.acv[1:15], time.15, method = "kendall"), cor(pop8.acv[1:15], time.15, method = "kendall"))

constant.dt.tau<- c(cor(pop1.dt[1:15], time.15, method = "kendall"), cor(pop2.dt[1:15], time.15, method = "kendall"),
                            cor(pop3.dt[1:15], time.15, method = "kendall"), cor(pop4.dt[1:15], time.15, method = "kendall"))
warming.dt.tau <- c(cor(pop5.dt[1:15], time.15, method = "kendall"), cor(pop6.dt[1:15], time.15, method = "kendall"),
                            cor(pop7.dt[1:15], time.15, method = "kendall"), cor(pop8.dt[1:15], time.15, method = "kendall"))
```

Calculating AUC:

```{r}
AUC_mean_data <- calc_auc(warming.mean.tau,constant.mean.tau) ; AUC_mean_data
AUC_sk_data <- calc_auc(warming.sk.tau,constant.sk.tau) ; AUC_sk_data
AUC_kt_data <- calc_auc(warming.kt.tau,constant.kt.tau) ; AUC_kt_data
AUC_var_data <- calc_auc(warming.var.tau,constant.var.tau) ; AUC_var_data
AUC_cv_data <- calc_auc(warming.cv.tau,constant.cv.tau) ; AUC_cv_data
AUC_id_data <- calc_auc(warming.id.tau,constant.id.tau) ; AUC_id_data
AUC_fdv_data <- calc_auc(warming.fdv.tau,constant.fdv.tau) ; AUC_fdv_data
AUC_ac_data <- calc_auc(warming.ac.tau,constant.ac.tau) ; AUC_ac_data
AUC_acv_data <- calc_auc(warming.acv.tau,constant.acv.tau) ; AUC_acv_data
AUC_dt_data <- calc_auc(warming.dt.tau,constant.dt.tau) ; AUC_dt_data
```

## Summarize Outputs

```{r}
metrics <- c("mean", "skewness", "kurtosis", "variance", "cv", "id", "fdv", "ac", "acv", "dt")

AUC_data <- c(AUC_mean_data, 
              AUC_sk_data, 
              AUC_kt_data, 
              AUC_var_data,
              AUC_cv_data,
              AUC_id_data, 
              AUC_fdv_data, 
              AUC_ac_data, 
              AUC_acv_data, 
              AUC_dt_data)

AUC_df <- data.frame(metrics, AUC_data); AUC_df
saveRDS(AUC_df, "Objects/AUC_SmoothEmpiricalData_B4.rds")
```

A plot!

```{r}
metrics_full <- c("Mean", "Skewness", "Kurtosis", "Variance", "Covariance", "Index of Dispersion", "First Difference Variance", "Autocorrelation", "Autocovariance", "Decay Time")

## Plot
plot(1:10, AUC_df$AUC_data, type = "p", 
     main = "Smoothed Empirical Data, All Populations", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "orchid", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)
```

# Excluding Population 5

```{r}
time.15 <- seq(1,15,1)

warming.mean.tau_without5 <- c(cor(pop6.mean[1:15], time.15, method = "kendall"),
                               cor(pop7.mean[1:15], time.15, method = "kendall"), 
                               cor(pop8.mean[1:15], time.15, method = "kendall"))

warming.sk.tau_without5 <- c(cor(pop6.sk[1:15], time.15, method = "kendall"),
                                   cor(pop7.sk[1:15], time.15, method = "kendall"), 
                                   cor(pop8.sk[1:15], time.15, method = "kendall"))


warming.kt.tau_without5 <- c(cor(pop6.kt[1:15], time.15, method = "kendall"),
                                   cor(pop7.kt[1:15], time.15, method = "kendall"), 
                                   cor(pop8.kt[1:15], time.15, method = "kendall"))

warming.var.tau_without5 <- c(cor(pop6.var[1:15], time.15, method = "kendall"),
                                   cor(pop7.var[1:15], time.15, method = "kendall"), 
                                   cor(pop8.var[1:15], time.15, method = "kendall"))

warming.cv.tau_without5 <- c(cor(pop6.cv[1:15], time.15, method = "kendall"),
                             cor(pop7.cv[1:15], time.15, method = "kendall"), 
                             cor(pop8.cv[1:15], time.15, method = "kendall"))

warming.id.tau_without5 <- c(cor(pop6.id[1:15], time.15, method = "kendall"),
                             cor(pop7.id[1:15], time.15, method = "kendall"), 
                             cor(pop8.id[1:15], time.15, method = "kendall"))

warming.ac.tau_without5 <- c(cor(pop6.ac[1:15], time.15, method = "kendall"),
                             cor(pop7.ac[1:15], time.15, method = "kendall"), 
                             cor(pop8.ac[1:15], time.15, method = "kendall"))


warming.acv.tau_without5 <- c(cor(pop6.acv[1:15], time.15, method = "kendall"),
                              cor(pop7.acv[1:15], time.15, method = "kendall"), 
                              cor(pop8.acv[1:15], time.15, method = "kendall"))

warming.fdv.tau_without5 <- c(cor(pop6.fdv[1:15], time.15, method = "kendall"),
                                              cor(pop7.fdv[1:15], time.15, method = "kendall"),
                                              cor(pop8.fdv[1:15], time.15, method = "kendall"))

warming.dt.tau_without5 <- c(cor(pop6.dt[1:15], time.15, method = "kendall"),
                                     cor(pop7.dt[1:15], time.15, method = "kendall"), 
                                     cor(pop8.dt[1:15], time.15, method = "kendall"))
```

## Summarize Outputs

```{r}
AUC_mean_data_without5 <- calc_auc(warming.mean.tau_without5,constant.mean.tau); AUC_mean_data_without5 
AUC_sk_data_without5 <- calc_auc(warming.sk.tau_without5,constant.sk.tau); AUC_sk_data_without5 
AUC_kt_data_without5 <- calc_auc(warming.kt.tau_without5, constant.kt.tau); AUC_kt_data_without5 
AUC_var_data_without5 <- calc_auc(warming.var.tau_without5,constant.var.tau); AUC_var_data_without5 
AUC_cv_data_without5 <- calc_auc(warming.cv.tau_without5,constant.cv.tau); AUC_cv_data_without5
AUC_id_data_without5 <- calc_auc(warming.id.tau_without5,constant.id.tau); AUC_id_data_without5 
AUC_fdv_data_without5 <- calc_auc(warming.fdv.tau_without5,constant.fdv.tau); AUC_fdv_data_without5 
AUC_ac_data_without5 <- calc_auc(warming.ac.tau_without5,constant.ac.tau); AUC_ac_data_without5 
AUC_acv_data_without5 <- calc_auc(warming.acv.tau_without5,constant.acv.tau); AUC_acv_data_without5 
AUC_dt_data_without5 <- calc_auc(warming.dt.tau_without5,constant.dt.tau); AUC_dt_data_without5 
```

```{r}
AUC_data <- c(AUC_mean_data_without5,
              AUC_sk_data_without5,
              AUC_kt_data_without5,
              AUC_var_data_without5,
              AUC_cv_data_without5,
              AUC_id_data_without5,
              AUC_fdv_data_without5,
              AUC_ac_data_without5,
              AUC_acv_data_without5,
              AUC_dt_data_without5)

AUC_df_without5 <- data.frame(metrics, AUC_data); AUC_df_without5
saveRDS(AUC_df_without5, "Objects/AUC_SmoothEmpiricalData_ExPop5_B4.rds")
```

A plot!

```{r}
## Plots
plot(1:10, AUC_df_without5$AUC_data, type = "p", 
     main = "Smoothed Empirical Data, Excluding Population 5", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "red3", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)

plot(1:10, AUC_df$AUC_data, type = "p", 
     main = "Smoothed Empirical Data", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "orchid", pch = 19)
points(1:10, AUC_df_without5$AUC_data, col = "red3", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)
legend("topright", legend = c("All Populations", "Excluding Population 5"),
       col = c("orchid", "red3"), pch = 19)

## Saving plots as .pngs
png(file = "Plots/AUC_SmoothEmpiricalData_B4.png", width = 1500, height = 350)
plot(1:10, AUC_df$AUC_data, type = "p", 
     main = "Smoothed Empirical Data", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "orchid", pch = 19)
points(1:10, AUC_df_without5$AUC_data, col = "red3", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full)
legend("topright", legend = c("All Populations", "Excluding Population 5"),
       col = c("orchid", "red3"), pch = 19)
dev.off()
```

# Compare Bandwidths

```{r}
AUC.Smooth.EmpDat.AllPopn.B2 <- readRDS("Objects/AUC_SmoothEmpiricalData_B2.rds")
AUC.Smooth.EmpDat.ExPopn5.B2 <- readRDS("Objects/AUC_SmoothEmpiricalData_ExPop5_B2.rds")
AUC.Smooth.EmpDat.AllPopn.B3 <- readRDS("Objects/AUC_SmoothEmpiricalData_B3.rds")
AUC.Smooth.EmpDat.ExPopn5.B3 <- readRDS("Objects/AUC_SmoothEmpiricalData_ExPop5_B3.rds")
AUC.Smooth.EmpDat.AllPopn.B4 <- readRDS("Objects/AUC_SmoothEmpiricalData_B4.rds")
AUC.Smooth.EmpDat.ExPopn5.B4 <- readRDS("Objects/AUC_SmoothEmpiricalData_ExPop5_B4.rds")

png(file = "Plots/AUC_SmoothEmpiricalData_VaryBandwidth.png", width = 1500, height = 350)
plot(1:10, AUC.Smooth.EmpDat.AllPopn.B2$AUC_data, pch = 19, col = alpha("violetred", 0.5),
     main = "Smoothed Empirical Data (All Pop'ns)", xlab = "", ylab = "AUC", ylim = c(0, 1))
points(1:10, AUC.Smooth.EmpDat.AllPopn.B3$AUC_data, pch = 19, col = alpha("coral", 0.5))
points(1:10, AUC.Smooth.EmpDat.AllPopn.B4$AUC_data, pch = 19, col = alpha("seagreen", 0.5))
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)
legend("bottomright", legend = c("Bandwidth = 2", "Bandwidth = 3", "Bandwidth = 4"),
       col = c("violetred", "coral", "seagreen"), pch = 19)
dev.off()

png(file = "Plots/AUC_SmoothEmpiricalData_ExPop5_VaryBandwidth.png", width = 1500, height = 350)
plot(1:10, AUC.Smooth.EmpDat.ExPopn5.B2$AUC_data, pch = 19, col = alpha("violetred", 0.5),
     main = "Smoothed Empirical Data (Excluding Pop'n 5)", xlab = "", ylab = "AUC", ylim = c(0, 1))
points(1:10, AUC.Smooth.EmpDat.ExPopn5.B3$AUC_data, pch = 19, col = alpha("coral", 0.5))
points(1:10, AUC.Smooth.EmpDat.ExPopn5.B4$AUC_data, pch = 19, col = alpha("seagreen", 0.5))
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)
legend("bottomright", legend = c("Bandwidth = 2", "Bandwidth = 3", "Bandwidth = 4"),
       col = c("violetred", "coral", "seagreen"), pch = 19)
dev.off()
```
