==> 11_ReAnalysis-RestrictedPeriod.Rmd <==
# **Looking for EWS in Days 30 to 60: Simulated and Empirical Data**

## **Required Packages**

```{r, include = FALSE}
require(scales)
require(evobiR)
require(gdata)
require(doBy)
require(ecp)
require(strucchange)
require(TTR)
require(moments)
require(tidyverse)
require(earlywarnings)
library(ggridges)
library(cowplot)
```

## Simulated Data

```{r}
## Constant temperature
constant1 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.1.csv")
constant2 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.2.csv")
constant3 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.3.csv")
constant4 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.4.csv")
constant5 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.5.csv")
constant6 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.6.csv")
constant7 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.7.csv")
constant8 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.8.csv")
constant9 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.9.csv")
constant10 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/constant.incidence.last.120.10.csv")

## Variable temperature
warming1 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.1.csv")
warming2 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.2.csv")
warming3 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.3.csv")
warming4 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.4.csv")
warming5 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.5.csv")
warming6 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.6.csv")
warming7 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.7.csv")
warming8 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.8.csv")
warming9 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.9.csv")
warming10 <- read.csv("Kirk_Data/Sim_Pops_EWS_Analyses/warming.incidence.last.120.10.csv")

constant.incidence.daily <- cbind(constant1[ ,2:101], constant2[ ,2:101], constant3[ ,2:101],
                                  constant4[ ,2:101], constant5[ ,2:101], constant6[ ,2:101],
                                  constant7[ ,2:101], constant8[ ,2:101], constant9[ ,2:101],
                                  constant10[ ,2:101])
warming.incidence.daily <- cbind(warming1[ ,2:101], warming2[ ,2:101], warming3[ ,2:101],
                                  warming4[ ,2:101], warming5[ ,2:101], warming6[ ,2:101],
                                  warming7[ ,2:101], warming8[ ,2:101], warming9[ ,2:101],
                                  warming10[ ,2:101])
```

### Calculating Metrics (5 Day Sliding Window)

```{r}
## SD, Mean, Skewness, Kurtosis
num.of.sims <- 10^3
constant.sims.sd.5 <- matrix(nrow = 115, ncol = num.of.sims)
constant.sims.mean.5 <- matrix(nrow = 115, ncol = num.of.sims)
constant.sims.sk.5 <- matrix(nrow = 115, ncol = num.of.sims)
constant.sims.kt.5 <- matrix(nrow = 115, ncol = num.of.sims)
warming.sims.sd.5 <- matrix(nrow = 115, ncol = num.of.sims)
warming.sims.mean.5 <- matrix(nrow = 115, ncol = num.of.sims)
warming.sims.sk.5 <- matrix(nrow = 115, ncol = num.of.sims)
warming.sims.kt.5 <- matrix(nrow = 115, ncol = num.of.sims)
for(i in 1:num.of.sims){
  constant.sims.sd.5[ ,i] <- SlidingWindow(FUN = "sd", data = constant.incidence.daily[ ,i], window = 5, step = 1)
  constant.sims.mean.5[ ,i] <- SlidingWindow(FUN = "mean", data = constant.incidence.daily[ ,i], window = 5, step = 1)
  constant.sims.sk.5[ ,i] <- SlidingWindow(FUN = "skewness", data = constant.incidence.daily[ ,i], window = 5, step = 1)
  constant.sims.kt.5[ ,i] <- SlidingWindow(FUN = "kurtosis", data = constant.incidence.daily[ ,i], window = 5, step = 1)
}
for(i in 1:num.of.sims){
  warming.sims.sd.5[ ,i] <- SlidingWindow(FUN = "sd", data = warming.incidence.daily[ ,i], window = 5, step = 1)
  warming.sims.mean.5[ ,i] <- SlidingWindow(FUN = "mean", data = warming.incidence.daily[ ,i], window = 5, step = 1)
  warming.sims.sk.5[ ,i] <- SlidingWindow(FUN = "skewness", data = warming.incidence.daily[ ,i], window = 5, step = 1)
  warming.sims.kt.5[ ,i] <- SlidingWindow(FUN = "kurtosis", data = warming.incidence.daily[ ,i], window = 5, step = 1)
}

## Variance
constant.sims.var.5 <- constant.sims.sd.5*constant.sims.sd.5
warming.sims.var.5 <- warming.sims.sd.5*warming.sims.sd.5

## Co-Efficient of Variation
constant.sims.cv.5 <- constant.sims.sd.5/constant.sims.mean.5
warming.sims.cv.5 <- warming.sims.sd.5/warming.sims.mean.5

## Index of Dispersion
constant.sims.id.5 <- constant.sims.var.5/constant.sims.mean.5
warming.sims.id.5 <- warming.sims.var.5/warming.sims.mean.5

## First Difference Variance
constant.sims.fdv.5 <- matrix(ncol = num.of.sims, nrow = 115)
warming.sims.fdv.5 <- matrix(ncol = num.of.sims, nrow = 115)
for(i in 1:num.of.sims){
  for(j in 2:115){
    constant.sims.fdv.5[j - 1,i] <- constant.sims.var.5[j,i] - constant.sims.var.5[j - 1,i]
    warming.sims.fdv.5[j - 1,i] <- warming.sims.var.5[j,i] - warming.sims.var.5[j - 1,i]
  }
}
constant.sims.fdv.5 <- constant.sims.fdv.5[1:114, ]
warming.sims.fdv.5 <- warming.sims.fdv.5[1:114, ]

## Autocorrelation and Autocovariance
constant.sims.ac.5 <- matrix(ncol = num.of.sims, nrow = 120)
constant.sims.acv.5 <- matrix(ncol = num.of.sims, nrow = 120)
for(j in 1:num.of.sims){
  for(i in 1:115){
    window.start <- i
    window.end <- i + 4
    segment <- constant.incidence.daily[c(window.start:window.end),j]
    constant.sims.ac.5[i, j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    constant.sims.acv.5[i, j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}
warming.sims.ac.5 <- matrix(ncol = num.of.sims,nrow = 120)
warming.sims.acv.5 <- matrix(ncol = num.of.sims,nrow = 120)
for(j in 1:num.of.sims){
  for(i in 1:115){
    window.start <- i
    window.end <- i + 4
    segment <- warming.incidence.daily[c(window.start:window.end),j]
    warming.sims.ac.5[i, j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    warming.sims.acv.5[i, j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}

## Decay Time
constant.sims.dt.5 <- matrix(ncol = num.of.sims, nrow = 55)
warming.sims.dt.5 <- matrix(nco = num.of.sims, nrow = 55)
for(i in 1:num.of.sims){
  for(j in 1:55){
    suppressWarnings(constant.sims.dt.5[j, i] <- -j/log(constant.sims.ac.5[j, i]))
    suppressWarnings(warming.sims.dt.5[j, i] <- -j/log(warming.sims.ac.5[j, i]))
  }
}
```

### Calculating Metrics (15 Day Sliding Window)

```{r}
## SD, Mean, Skewness, Kurtosis
num.of.sims <- 10^3
constant.sims.sd.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.mean.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.sk.15 <- matrix(nrow = 105, ncol = num.of.sims)
constant.sims.kt.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.sd.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.mean.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.sk.15 <- matrix(nrow = 105, ncol = num.of.sims)
warming.sims.kt.15 <- matrix(nrow = 105, ncol = num.of.sims)
for(i in 1:num.of.sims){
  constant.sims.sd.15[ ,i] <- SlidingWindow(FUN = "sd", data = constant.incidence.daily[ ,i], window = 15, step = 1)
  constant.sims.mean.15[ ,i] <- SlidingWindow(FUN = "mean", data = constant.incidence.daily[ ,i], window = 15, step = 1)
  constant.sims.sk.15[ ,i] <- SlidingWindow(FUN = "skewness", data = constant.incidence.daily[ ,i], window = 15, step = 1)
  constant.sims.kt.15[ ,i] <- SlidingWindow(FUN = "kurtosis", data = constant.incidence.daily[ ,i], window = 15, step = 1)
}
for(i in 1:num.of.sims){
  warming.sims.sd.15[ ,i] <- SlidingWindow(FUN = "sd", data = warming.incidence.daily[ ,i], window = 15, step = 1)
  warming.sims.mean.15[ ,i] <- SlidingWindow(FUN = "mean", data = warming.incidence.daily[ ,i], window = 15, step = 1)
  warming.sims.sk.15[ ,i] <- SlidingWindow(FUN = "skewness", data = warming.incidence.daily[ ,i], window = 15, step = 1)
  warming.sims.kt.15[ ,i] <- SlidingWindow(FUN = "kurtosis", data = warming.incidence.daily[ ,i], window = 15, step = 1)
}

## Variance
constant.sims.var.15 <- constant.sims.sd.15*constant.sims.sd.15
warming.sims.var.15 <- warming.sims.sd.15*warming.sims.sd.15

## Co-Efficient of Variation
constant.sims.cv.15 <- constant.sims.sd.15/constant.sims.mean.15
warming.sims.cv.15 <- warming.sims.sd.15/warming.sims.mean.15

## Index of Dispersion
constant.sims.id.15 <- constant.sims.var.15/constant.sims.mean.15
warming.sims.id.15 <- warming.sims.var.15/warming.sims.mean.15

## First Difference Variance
constant.sims.fdv.15 <- matrix(ncol = num.of.sims, nrow = 105)
warming.sims.fdv.15 <- matrix(ncol = num.of.sims, nrow = 105)
for(i in 1:num.of.sims){
  for(j in 2:105){
    constant.sims.fdv.15[j - 1,i] <- constant.sims.var.15[j,i] - constant.sims.var.15[j - 1,i]
    warming.sims.fdv.15[j - 1,i] <- warming.sims.var.15[j,i] - warming.sims.var.15[j - 1,i]
  }
}
constant.sims.fdv.15 <- constant.sims.fdv.15[1:104, ]
warming.sims.fdv.15 <- warming.sims.fdv.15[1:104, ]

## Autocorrelation and Autocovariance
constant.sims.ac.15 <- matrix(ncol = num.of.sims, nrow = 120)
constant.sims.acv.15 <- matrix(ncol = num.of.sims, nrow = 120)
for(j in 1:num.of.sims){
  for(i in 1:105){
    window.start <- i
    window.end <- i + 14
    segment <- constant.incidence.daily[c(window.start:window.end),j]
    constant.sims.ac.15[i, j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    constant.sims.acv.15[i, j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}
warming.sims.ac.15 <- matrix(ncol = num.of.sims,nrow = 120)
warming.sims.acv.15 <- matrix(ncol = num.of.sims,nrow = 120)
for(j in 1:num.of.sims){
  for(i in 1:105){
    window.start <- i
    window.end <- i + 14
    segment <- warming.incidence.daily[c(window.start:window.end),j]
    warming.sims.ac.15[i, j] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
    warming.sims.acv.15[i, j] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
  }
}

## Decay Time
constant.sims.dt.15 <- matrix(ncol = num.of.sims, nrow = 45)
warming.sims.dt.15 <- matrix(nco = num.of.sims, nrow = 45)
for(i in 1:num.of.sims){
  for(j in 1:45){
    suppressWarnings(constant.sims.dt.15[j, i] <- -j/log(constant.sims.ac.15[j, i]))
    suppressWarnings(warming.sims.dt.15[j, i] <- -j/log(warming.sims.ac.15[j, i]))
  }
}
```

### Kendall's Tau and AUC (5 Day Sliding Window) \| Days 20 to 55

```{r}
constant.sims.var.5.day20 <- constant.sims.var.5[20:55, ]
constant.sims.mean.5.day20 <- constant.sims.mean.5[20:55, ]
constant.sims.cv.5.day20 <- constant.sims.cv.5[20:55, ]
constant.sims.sk.5.day20 <- constant.sims.sk.5[20:55, ]
constant.sims.kt.5.day20 <- constant.sims.kt.5[20:55, ]
constant.sims.id.5.day20 <- constant.sims.id.5[20:55, ]
constant.sims.ac.5.day20 <- constant.sims.ac.5[20:55, ]
constant.sims.acv.5.day20 <- constant.sims.acv.5[20:55, ]
constant.sims.fdv.5.day20 <- constant.sims.fdv.5[20:55, ]
constant.sims.dt.5.day20 <- constant.sims.dt.5[20:55, ]

warming.sims.var.5.day20 <- warming.sims.var.5[20:55, ]
warming.sims.mean.5.day20 <- warming.sims.mean.5[20:55, ]
warming.sims.cv.5.day20 <- warming.sims.cv.5[20:55, ]
warming.sims.sk.5.day20 <- warming.sims.sk.5[20:55, ]
warming.sims.kt.5.day20 <- warming.sims.kt.5[20:55, ]
warming.sims.id.5.day20 <- warming.sims.id.5[20:55, ]
warming.sims.ac.5.day20 <- warming.sims.ac.5[20:55, ]
warming.sims.acv.5.day20 <- warming.sims.acv.5[20:55, ]
warming.sims.fdv.5.day20 <- warming.sims.fdv.5[20:55, ]
warming.sims.dt.5.day20 <- warming.sims.dt.5[20:55, ]

constant.sims.var.5.day20.tau <- c()
constant.sims.mean.5.day20.tau <- c()
constant.sims.cv.5.day20.tau <- c()
constant.sims.sk.5.day20.tau <- c()
constant.sims.kt.5.day20.tau <- c()
constant.sims.id.5.day20.tau <- c()
constant.sims.ac.5.day20.tau <- c()
constant.sims.acv.5.day20.tau <- c()
constant.sims.fdv.5.day20.tau <- c()
constant.sims.dt.5.day20.tau <- c()

warming.sims.var.5.day20.tau <- c()
warming.sims.mean.5.day20.tau <- c()
warming.sims.cv.5.day20.tau <- c()
warming.sims.sk.5.day20.tau <- c()
warming.sims.kt.5.day20.tau <- c()
warming.sims.id.5.day20.tau <- c()
warming.sims.ac.5.day20.tau <- c()
warming.sims.acv.5.day20.tau <- c()
warming.sims.fdv.5.day20.tau <- c()
warming.sims.dt.5.day20.tau <- c()

time.5 <- c(1:36)

for(i in 1:num.of.sims){
  constant.sims.mean.5.day20.tau[i] <- cor(constant.sims.mean.5.day20[ , i], time.5, method = "kendall")
  constant.sims.sk.5.day20.tau[i] <- cor(constant.sims.sk.5.day20[ , i], time.5, method = "kendall")
  constant.sims.kt.5.day20.tau[i] <- cor(constant.sims.kt.5.day20[ , i], time.5, method = "kendall")
  constant.sims.var.5.day20.tau[i] <- cor(constant.sims.var.5.day20[ , i], time.5, method = "kendall")
  constant.sims.cv.5.day20.tau[i] <- cor(constant.sims.cv.5.day20[ , i], time.5, method = "kendall")
  constant.sims.id.5.day20.tau[i] <- cor(constant.sims.id.5.day20[ , i], time.5, method = "kendall")
  constant.sims.fdv.5.day20.tau[i] <- cor(constant.sims.fdv.5.day20[ , i], time.5, method = "kendall")
  constant.sims.ac.5.day20.tau[i] <- cor(constant.sims.ac.5.day20[ , i], time.5, method = "kendall")
  constant.sims.acv.5.day20.tau[i] <- cor(constant.sims.acv.5.day20[ , i], time.5, method = "kendall")
  constant.sims.dt.5.day20.tau[i] <- cor(constant.sims.dt.5.day20[ , i], time.5, method = "kendall")
  
  warming.sims.mean.5.day20.tau[i] <- cor(warming.sims.mean.5.day20[ , i], time.5, method = "kendall")
  warming.sims.sk.5.day20.tau[i] <- cor(warming.sims.sk.5.day20[ , i], time.5, method = "kendall")
  warming.sims.kt.5.day20.tau[i] <- cor(warming.sims.kt.5.day20[ , i], time.5, method = "kendall")
  warming.sims.var.5.day20.tau[i] <- cor(warming.sims.var.5.day20[ , i], time.5, method = "kendall")
  warming.sims.cv.5.day20.tau[i] <- cor(warming.sims.cv.5.day20[ , i], time.5, method = "kendall")
  warming.sims.id.5.day20.tau[i] <- cor(warming.sims.id.5.day20[ , i], time.5, method = "kendall")
  warming.sims.fdv.5.day20.tau[i] <- cor(warming.sims.fdv.5.day20[ , i], time.5, method = "kendall")
  warming.sims.ac.5.day20.tau[i] <- cor(warming.sims.ac.5.day20[ , i], time.5, method = "kendall")
  warming.sims.acv.5.day20.tau[i] <- cor(warming.sims.acv.5.day20[ , i], time.5, method = "kendall")
  warming.sims.dt.5.day20.tau[i] <- cor(warming.sims.dt.5.day20[ , i], time.5, method = "kendall")
}

AUC_mean_5 <- calc_auc(warming.sims.mean.5.day20.tau, constant.sims.mean.5.day20.tau); AUC_mean_5
AUC_sk_5 <- calc_auc(warming.sims.sk.5.day20.tau, constant.sims.sk.5.day20.tau); AUC_sk_5
AUC_kt_5 <- calc_auc(warming.sims.kt.5.day20.tau, constant.sims.kt.5.day20.tau); AUC_kt_5
AUC_var_5 <- calc_auc(warming.sims.var.5.day20.tau, constant.sims.var.5.day20.tau); AUC_var_5
AUC_cv_5 <- calc_auc(warming.sims.cv.5.day20.tau, constant.sims.cv.5.day20.tau); AUC_cv_5
AUC_id_5 <- calc_auc(warming.sims.id.5.day20.tau, constant.sims.id.5.day20.tau); AUC_id_5
AUC_fdv_5 <- calc_auc(warming.sims.fdv.5.day20.tau, constant.sims.fdv.5.day20.tau); AUC_fdv_5
AUC_ac_5 <- calc_auc(warming.sims.ac.5.day20.tau, constant.sims.ac.5.day20.tau); AUC_ac_5
AUC_acv_5 <- calc_auc(warming.sims.acv.5.day20.tau, constant.sims.acv.5.day20.tau); AUC_acv_5
AUC_dt_5 <- calc_auc(warming.sims.dt.5.day20.tau, constant.sims.dt.5.day20.tau); AUC_dt_5
```

### Kendall's Tau and AUC (15 Day Sliding Window) \| Days 20 to 45

```{r}
constant.sims.var.15.day20 <- constant.sims.var.15[20:45, ]
constant.sims.mean.15.day20 <- constant.sims.mean.15[20:45, ]
constant.sims.cv.15.day20 <- constant.sims.cv.15[20:45, ]
constant.sims.sk.15.day20 <- constant.sims.sk.15[20:45, ]
constant.sims.kt.15.day20 <- constant.sims.kt.15[20:45, ]
constant.sims.id.15.day20 <- constant.sims.id.15[20:45, ]
constant.sims.ac.15.day20 <- constant.sims.ac.15[20:45, ]
constant.sims.acv.15.day20 <- constant.sims.acv.15[20:45, ]
constant.sims.fdv.15.day20 <- constant.sims.fdv.15[20:45, ]
constant.sims.dt.15.day20 <- constant.sims.dt.15[20:45, ]

warming.sims.var.15.day20 <- warming.sims.var.15[20:45, ]
warming.sims.mean.15.day20 <- warming.sims.mean.15[20:45, ]
warming.sims.cv.15.day20 <- warming.sims.cv.15[20:45, ]
warming.sims.sk.15.day20 <- warming.sims.sk.15[20:45, ]
warming.sims.kt.15.day20 <- warming.sims.kt.15[20:45, ]
warming.sims.id.15.day20 <- warming.sims.id.15[20:45, ]
warming.sims.ac.15.day20 <- warming.sims.ac.15[20:45, ]
warming.sims.acv.15.day20 <- warming.sims.acv.15[20:45, ]
warming.sims.fdv.15.day20 <- warming.sims.fdv.15[20:45, ]
warming.sims.dt.15.day20 <- warming.sims.dt.15[20:45, ]

constant.sims.var.15.day20.tau <- c()
constant.sims.mean.15.day20.tau <- c()
constant.sims.cv.15.day20.tau <- c()
constant.sims.sk.15.day20.tau <- c()
constant.sims.kt.15.day20.tau <- c()
constant.sims.id.15.day20.tau <- c()
constant.sims.ac.15.day20.tau <- c()
constant.sims.acv.15.day20.tau <- c()
constant.sims.fdv.15.day20.tau <- c()
constant.sims.dt.15.day20.tau <- c()

warming.sims.var.15.day20.tau <- c()
warming.sims.mean.15.day20.tau <- c()
warming.sims.cv.15.day20.tau <- c()
warming.sims.sk.15.day20.tau <- c()
warming.sims.kt.15.day20.tau <- c()
warming.sims.id.15.day20.tau <- c()
warming.sims.ac.15.day20.tau <- c()
warming.sims.acv.15.day20.tau <- c()
warming.sims.fdv.15.day20.tau <- c()
warming.sims.dt.15.day20.tau <- c()

time.15 <- c(1:26)

for(i in 1:num.of.sims){
  constant.sims.mean.15.day20.tau[i] <- cor(constant.sims.mean.15.day20[ , i], time.15, method = "kendall")
  constant.sims.sk.15.day20.tau[i] <- cor(constant.sims.sk.15.day20[ , i], time.15, method = "kendall")
  constant.sims.kt.15.day20.tau[i] <- cor(constant.sims.kt.15.day20[ , i], time.15, method = "kendall")
  constant.sims.var.15.day20.tau[i] <- cor(constant.sims.var.15.day20[ , i], time.15, method = "kendall")
  constant.sims.cv.15.day20.tau[i] <- cor(constant.sims.cv.15.day20[ , i], time.15, method = "kendall")
  constant.sims.id.15.day20.tau[i] <- cor(constant.sims.id.15.day20[ , i], time.15, method = "kendall")
  constant.sims.fdv.15.day20.tau[i] <- cor(constant.sims.fdv.15.day20[ , i], time.15, method = "kendall")
  constant.sims.ac.15.day20.tau[i] <- cor(constant.sims.ac.15.day20[ , i], time.15, method = "kendall")
  constant.sims.acv.15.day20.tau[i] <- cor(constant.sims.acv.15.day20[ , i], time.15, method = "kendall")
  constant.sims.dt.15.day20.tau[i] <- cor(constant.sims.dt.15.day20[ , i], time.15, method = "kendall")
  
  warming.sims.mean.15.day20.tau[i] <- cor(warming.sims.mean.15.day20[ , i], time.15, method = "kendall")
  warming.sims.sk.15.day20.tau[i] <- cor(warming.sims.sk.15.day20[ , i], time.15, method = "kendall")
  warming.sims.kt.15.day20.tau[i] <- cor(warming.sims.kt.15.day20[ , i], time.15, method = "kendall")
  warming.sims.var.15.day20.tau[i] <- cor(warming.sims.var.15.day20[ , i], time.15, method = "kendall")
  warming.sims.cv.15.day20.tau[i] <- cor(warming.sims.cv.15.day20[ , i], time.15, method = "kendall")
  warming.sims.id.15.day20.tau[i] <- cor(warming.sims.id.15.day20[ , i], time.15, method = "kendall")
  warming.sims.fdv.15.day20.tau[i] <- cor(warming.sims.fdv.15.day20[ , i], time.15, method = "kendall")
  warming.sims.ac.15.day20.tau[i] <- cor(warming.sims.ac.15.day20[ , i], time.15, method = "kendall")
  warming.sims.acv.15.day20.tau[i] <- cor(warming.sims.acv.15.day20[ , i], time.15, method = "kendall")
  warming.sims.dt.15.day20.tau[i] <- cor(warming.sims.dt.15.day20[ , i], time.15, method = "kendall")
}

AUC_mean_15 <- calc_auc(warming.sims.mean.15.day20.tau, constant.sims.mean.15.day20.tau); AUC_mean_15
AUC_sk_15 <- calc_auc(warming.sims.sk.15.day20.tau, constant.sims.sk.15.day20.tau); AUC_sk_15
AUC_kt_15 <- calc_auc(warming.sims.kt.15.day20.tau, constant.sims.kt.15.day20.tau); AUC_kt_15
AUC_var_15 <- calc_auc(warming.sims.var.15.day20.tau, constant.sims.var.15.day20.tau); AUC_var_15
AUC_cv_15 <- calc_auc(warming.sims.cv.15.day20.tau, constant.sims.cv.15.day20.tau); AUC_cv_15
AUC_id_15 <- calc_auc(warming.sims.id.15.day20.tau, constant.sims.id.15.day20.tau); AUC_id_15
AUC_fdv_15 <- calc_auc(warming.sims.fdv.15.day20.tau, constant.sims.fdv.15.day20.tau); AUC_fdv_15
AUC_ac_15 <- calc_auc(warming.sims.ac.15.day20.tau, constant.sims.ac.15.day20.tau); AUC_ac_15
AUC_acv_15 <- calc_auc(warming.sims.acv.15.day20.tau, constant.sims.acv.15.day20.tau); AUC_acv_15
AUC_dt_15 <- calc_auc(warming.sims.dt.15.day20.tau, constant.sims.dt.15.day20.tau); AUC_dt_15
```

### Summarized Outputs

```{r}
metrics <- c("mean", "skewness", "kurtosis", "variance", "cv", "id", "fdv", "cor", "acv", "dt")
AUC_5_restricted <- c(AUC_mean_5, AUC_sk_5, AUC_kt_5, AUC_var_5, AUC_cv_5, 
                      AUC_id_5, AUC_fdv_5, AUC_ac_5 ,AUC_acv_5, AUC_dt_5)
AUC_15_restricted <- c(AUC_mean_15, AUC_sk_15, AUC_kt_15, AUC_var_15, AUC_cv_15, 
                       AUC_id_15, AUC_fdv_15, AUC_ac_15 ,AUC_acv_15, AUC_dt_15)
AUC_restricted_df <- data.frame(metrics, AUC_5_restricted, AUC_15_restricted); AUC_restricted_df

plot(1:10, AUC_restricted_df$AUC_5_restricted, type = "p", 
     main = "Simulated Data (Days 20 to 55)", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "orchid", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)

plot(1:10, AUC_restricted_df$AUC_15_restricted, type = "p", 
     main = "Simulated Data (Days 20 to 45)", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "orchid", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)

## Medians
paste("Median Mean in Control Populations: ", median(constant.sims.mean.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Mean in Test Populations: ", median(warming.sims.mean.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Skewness in Control Populations: ", median(constant.sims.sk.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Skewness in Test Populations: ", median(warming.sims.sk.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Kurtosis in Control Populations: ", median(constant.sims.kt.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Kurtosis in Test Populations: ", median(warming.sims.kt.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Variance in Control Populations: ", median(constant.sims.var.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Variance in Test Populations: ", median(warming.sims.var.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Co-Efficient of Variation in Control Populations: ", median(constant.sims.cv.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Co-Efficient of Variation in Test Populations: ", median(warming.sims.cv.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Index of Dispersion in Control Populations: ", median(constant.sims.id.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Index of Dispersion in Test Populations: ", median(warming.sims.id.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median First Difference Variance in Control Populations: ", median(constant.sims.fdv.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median First Difference Variance in Test Populations: ", median(warming.sims.fdv.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Autocorrelation in Control Populations: ", median(constant.sims.ac.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Autocorrelation in Test Populations: ", median(warming.sims.ac.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Autocovariance in Control Populations: ", median(constant.sims.acv.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Autocovariance in Test Populations: ", median(warming.sims.acv.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Decay Time in Control Populations: ", median(constant.sims.dt.5.day20.tau, na.rm = TRUE), sep = "")
paste("Median Decay Time in Test Populations: ", median(warming.sims.dt.5.day20.tau, na.rm = TRUE), sep = "")

paste("Median Mean in Control Populations: ", median(constant.sims.mean.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Mean in Test Populations: ", median(warming.sims.mean.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Skewness in Control Populations: ", median(constant.sims.sk.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Skewness in Test Populations: ", median(warming.sims.sk.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Kurtosis in Control Populations: ", median(constant.sims.kt.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Kurtosis in Test Populations: ", median(warming.sims.kt.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Variance in Control Populations: ", median(constant.sims.var.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Variance in Test Populations: ", median(warming.sims.var.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Co-Efficient of Variation in Control Populations: ", median(constant.sims.cv.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Co-Efficient of Variation in Test Populations: ", median(warming.sims.cv.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Index of Dispersion in Control Populations: ", median(constant.sims.id.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Index of Dispersion in Test Populations: ", median(warming.sims.id.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median First Difference Variance in Control Populations: ", median(constant.sims.fdv.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median First Difference Variance in Test Populations: ", median(warming.sims.fdv.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Autocorrelation in Control Populations: ", median(constant.sims.ac.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Autocorrelation in Test Populations: ", median(warming.sims.ac.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Autocovariance in Control Populations: ", median(constant.sims.acv.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Autocovariance in Test Populations: ", median(warming.sims.acv.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Decay Time in Control Populations: ", median(constant.sims.dt.15.day20.tau, na.rm = TRUE), sep = "")
paste("Median Decay Time in Test Populations: ", median(warming.sims.dt.15.day20.tau, na.rm = TRUE), sep = "")
```

### Plotting

```{r}
cols <- c(rep("#1434A4", 512), rep("#C21E56", 512))

## Mean
sim.mean.15.day20.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.mean.15.day20.tau, warming.sims.mean.15.day20.tau))
med.const.sims.mean.15.day20.tau <- median(constant.sims.mean.15.day20.tau)
med.warm.sims.mean.15.day20.tau <- median(warming.sims.mean.15.day20.tau)
dens.mean <- ggplot(data = sim.mean.15.day20.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.mean.15.day20.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.mean.15.day20.tau, colour = "#C21E56") +
  ggtitle("Mean: 80% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## Variance
sim.var.15.day20.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.var.15.day20.tau, warming.sims.var.15.day20.tau))
med.const.sims.var.15.day20.tau <- median(constant.sims.var.15.day20.tau)
med.warm.sims.var.15.day20.tau <- median(warming.sims.var.15.day20.tau)
dens.var <- ggplot(data = sim.var.15.day20.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.var.15.day20.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.var.15.day20.tau, colour = "#C21E56") +
  ggtitle("Variance: 91% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## Skewness
sim.sk.15.day20.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.sk.15.day20.tau, warming.sims.sk.15.day20.tau))
med.const.sims.sk.15.day20.tau <- median(constant.sims.sk.15.day20.tau)
med.warm.sims.sk.15.day20.tau <- median(warming.sims.sk.15.day20.tau)
dens.sk <- ggplot(data = sim.sk.15.day20.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.sk.15.day20.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.sk.15.day20.tau, colour = "#C21E56") +
  ggtitle("Skewness: 94% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## Kurtosis
sim.kt.15.day20.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.kt.15.day20.tau, warming.sims.kt.15.day20.tau))
med.const.sims.kt.15.day20.tau <- median(constant.sims.kt.15.day20.tau)
med.warm.sims.kt.15.day20.tau <- median(warming.sims.kt.15.day20.tau)
dens.kt <- ggplot(data = sim.kt.15.day20.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.kt.15.day20.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.kt.15.day20.tau, colour = "#C21E56") +
  ggtitle("Kurtosis: 96% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## Co-Efficient of Variation
sim.cv.15.day20.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.cv.15.day20.tau, warming.sims.cv.15.day20.tau))
med.const.sims.cv.15.day20.tau <- median(constant.sims.cv.15.day20.tau)
med.warm.sims.cv.15.day20.tau <- median(warming.sims.cv.15.day20.tau)
dens.cv <- ggplot(data = sim.cv.15.day20.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.cv.15.day20.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.cv.15.day20.tau, colour = "#C21E56") +
  ggtitle("Co-Efficient of Variation: 97% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## Index of Dispersion
sim.id.15.day20.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.id.15.day20.tau, warming.sims.id.15.day20.tau))
med.const.sims.id.15.day20.tau <- median(constant.sims.id.15.day20.tau)
med.warm.sims.id.15.day20.tau <- median(warming.sims.id.15.day20.tau)
dens.id <- ggplot(data = sim.id.15.day20.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.id.15.day20.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.id.15.day20.tau, colour = "#C21E56") +
  ggtitle("Index of Dispersion: 94% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## First Difference Variance
sim.fdv.15.day20.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.fdv.15.day20.tau, warming.sims.fdv.15.day20.tau))
med.const.sims.fdv.15.day20.tau <- median(constant.sims.fdv.15.day20.tau)
med.warm.sims.fdv.15.day20.tau <- median(warming.sims.fdv.15.day20.tau)
dens.fdv <- ggplot(data = sim.fdv.15.day20.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.fdv.15.day20.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.fdv.15.day20.tau, colour = "#C21E56") +
  ggtitle("First Difference Variance: 96% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## Autocorrelation
sim.ac.15.day20.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.ac.15.day20.tau, warming.sims.ac.15.day20.tau))
med.const.sims.ac.15.day20.tau <- median(constant.sims.ac.15.day20.tau)
med.warm.sims.ac.15.day20.tau <- median(warming.sims.ac.15.day20.tau)
dens.ac <- ggplot(data = sim.ac.15.day20.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.ac.15.day20.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.ac.15.day20.tau, colour = "#C21E56") +
  ggtitle("Autocorrelation: 97% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## Autocovariance
sim.acv.15.day20.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.acv.15.day20.tau, warming.sims.acv.15.day20.tau))
med.const.sims.acv.15.day20.tau <- median(constant.sims.acv.15.day20.tau)
med.warm.sims.acv.15.day20.tau <- median(warming.sims.acv.15.day20.tau)
dens.acv <- ggplot(data = sim.acv.15.day20.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.acv.15.day20.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.acv.15.day20.tau, colour = "#C21E56") +
  ggtitle("Autocovariance: 92% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## Decay Time
sim.dt.15.day20.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.dt.15.day20.tau, warming.sims.dt.15.day20.tau))
med.const.sims.dt.15.day20.tau <- median(constant.sims.dt.15.day20.tau)
med.warm.sims.dt.15.day20.tau <- median(warming.sims.dt.15.day20.tau)
dens.dt <- ggplot(data = sim.dt.15.day20.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.dt.15.day20.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.dt.15.day20.tau, colour = "#C21E56") +
  ggtitle("Decay Time: 94% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

sim.taus.day20.plot <- plot_grid(dens.mean, dens.sk, dens.kt, dens.var, dens.cv,
                                 dens.id, dens.fdv, dens.ac, dens.acv, dens.dt,
                                 nrow = 2, align = "h"
                                 # , labels = "AUTO", label_fontfamily = "serif"
                                 )
```

### Kendall's Tau and AUC (5 Day Sliding Window) \| Days 30 to 55

```{r}
constant.sims.var.5.day30 <- constant.sims.var.5[30:55, ]
constant.sims.mean.5.day30 <- constant.sims.mean.5[30:55, ]
constant.sims.cv.5.day30 <- constant.sims.cv.5[30:55, ]
constant.sims.sk.5.day30 <- constant.sims.sk.5[30:55, ]
constant.sims.kt.5.day30 <- constant.sims.kt.5[30:55, ]
constant.sims.id.5.day30 <- constant.sims.id.5[30:55, ]
constant.sims.ac.5.day30 <- constant.sims.ac.5[30:55, ]
constant.sims.acv.5.day30 <- constant.sims.acv.5[30:55, ]
constant.sims.fdv.5.day30 <- constant.sims.fdv.5[30:55, ]
constant.sims.dt.5.day30 <- constant.sims.dt.5[30:55, ]

warming.sims.var.5.day30 <- warming.sims.var.5[30:55, ]
warming.sims.mean.5.day30 <- warming.sims.mean.5[30:55, ]
warming.sims.cv.5.day30 <- warming.sims.cv.5[30:55, ]
warming.sims.sk.5.day30 <- warming.sims.sk.5[30:55, ]
warming.sims.kt.5.day30 <- warming.sims.kt.5[30:55, ]
warming.sims.id.5.day30 <- warming.sims.id.5[30:55, ]
warming.sims.ac.5.day30 <- warming.sims.ac.5[30:55, ]
warming.sims.acv.5.day30 <- warming.sims.acv.5[30:55, ]
warming.sims.fdv.5.day30 <- warming.sims.fdv.5[30:55, ]
warming.sims.dt.5.day30 <- warming.sims.dt.5[30:55, ]

constant.sims.var.5.day30.tau <- c()
constant.sims.mean.5.day30.tau <- c()
constant.sims.cv.5.day30.tau <- c()
constant.sims.sk.5.day30.tau <- c()
constant.sims.kt.5.day30.tau <- c()
constant.sims.id.5.day30.tau <- c()
constant.sims.ac.5.day30.tau <- c()
constant.sims.acv.5.day30.tau <- c()
constant.sims.fdv.5.day30.tau <- c()
constant.sims.dt.5.day30.tau <- c()

warming.sims.var.5.day30.tau <- c()
warming.sims.mean.5.day30.tau <- c()
warming.sims.cv.5.day30.tau <- c()
warming.sims.sk.5.day30.tau <- c()
warming.sims.kt.5.day30.tau <- c()
warming.sims.id.5.day30.tau <- c()
warming.sims.ac.5.day30.tau <- c()
warming.sims.acv.5.day30.tau <- c()
warming.sims.fdv.5.day30.tau <- c()
warming.sims.dt.5.day30.tau <- c()

time.5 <- c(1:26)

for(i in 1:num.of.sims){
  constant.sims.mean.5.day30.tau[i] <- cor(constant.sims.mean.5.day30[ , i], time.5, method = "kendall")
  constant.sims.sk.5.day30.tau[i] <- cor(constant.sims.sk.5.day30[ , i], time.5, method = "kendall")
  constant.sims.kt.5.day30.tau[i] <- cor(constant.sims.kt.5.day30[ , i], time.5, method = "kendall")
  constant.sims.var.5.day30.tau[i] <- cor(constant.sims.var.5.day30[ , i], time.5, method = "kendall")
  constant.sims.cv.5.day30.tau[i] <- cor(constant.sims.cv.5.day30[ , i], time.5, method = "kendall")
  constant.sims.id.5.day30.tau[i] <- cor(constant.sims.id.5.day30[ , i], time.5, method = "kendall")
  constant.sims.fdv.5.day30.tau[i] <- cor(constant.sims.fdv.5.day30[ , i], time.5, method = "kendall")
  constant.sims.ac.5.day30.tau[i] <- cor(constant.sims.ac.5.day30[ , i], time.5, method = "kendall")
  constant.sims.acv.5.day30.tau[i] <- cor(constant.sims.acv.5.day30[ , i], time.5, method = "kendall")
  constant.sims.dt.5.day30.tau[i] <- cor(constant.sims.dt.5.day30[ , i], time.5, method = "kendall")
  
  warming.sims.mean.5.day30.tau[i] <- cor(warming.sims.mean.5.day30[ , i], time.5, method = "kendall")
  warming.sims.sk.5.day30.tau[i] <- cor(warming.sims.sk.5.day30[ , i], time.5, method = "kendall")
  warming.sims.kt.5.day30.tau[i] <- cor(warming.sims.kt.5.day30[ , i], time.5, method = "kendall")
  warming.sims.var.5.day30.tau[i] <- cor(warming.sims.var.5.day30[ , i], time.5, method = "kendall")
  warming.sims.cv.5.day30.tau[i] <- cor(warming.sims.cv.5.day30[ , i], time.5, method = "kendall")
  warming.sims.id.5.day30.tau[i] <- cor(warming.sims.id.5.day30[ , i], time.5, method = "kendall")
  warming.sims.fdv.5.day30.tau[i] <- cor(warming.sims.fdv.5.day30[ , i], time.5, method = "kendall")
  warming.sims.ac.5.day30.tau[i] <- cor(warming.sims.ac.5.day30[ , i], time.5, method = "kendall")
  warming.sims.acv.5.day30.tau[i] <- cor(warming.sims.acv.5.day30[ , i], time.5, method = "kendall")
  warming.sims.dt.5.day30.tau[i] <- cor(warming.sims.dt.5.day30[ , i], time.5, method = "kendall")
}

AUC_mean_5 <- calc_auc(warming.sims.mean.5.day30.tau, constant.sims.mean.5.day30.tau); AUC_mean_5
AUC_sk_5 <- calc_auc(warming.sims.sk.5.day30.tau, constant.sims.sk.5.day30.tau); AUC_sk_5
AUC_kt_5 <- calc_auc(warming.sims.kt.5.day30.tau, constant.sims.kt.5.day30.tau); AUC_kt_5
AUC_var_5 <- calc_auc(warming.sims.var.5.day30.tau, constant.sims.var.5.day30.tau); AUC_var_5
AUC_cv_5 <- calc_auc(warming.sims.cv.5.day30.tau, constant.sims.cv.5.day30.tau); AUC_cv_5
AUC_id_5 <- calc_auc(warming.sims.id.5.day30.tau, constant.sims.id.5.day30.tau); AUC_id_5
AUC_fdv_5 <- calc_auc(warming.sims.fdv.5.day30.tau, constant.sims.fdv.5.day30.tau); AUC_fdv_5
AUC_ac_5 <- calc_auc(warming.sims.ac.5.day30.tau, constant.sims.ac.5.day30.tau); AUC_ac_5
AUC_acv_5 <- calc_auc(warming.sims.acv.5.day30.tau, constant.sims.acv.5.day30.tau); AUC_acv_5
AUC_dt_5 <- calc_auc(warming.sims.dt.5.day30.tau, constant.sims.dt.5.day30.tau); AUC_dt_5
```

### Kendall's Tau and AUC (15 Day Sliding Window) \| Days 30 to 45

```{r}
constant.sims.var.15.day30 <- constant.sims.var.15[30:45, ]
constant.sims.mean.15.day30 <- constant.sims.mean.15[30:45, ]
constant.sims.cv.15.day30 <- constant.sims.cv.15[30:45, ]
constant.sims.sk.15.day30 <- constant.sims.sk.15[30:45, ]
constant.sims.kt.15.day30 <- constant.sims.kt.15[30:45, ]
constant.sims.id.15.day30 <- constant.sims.id.15[30:45, ]
constant.sims.ac.15.day30 <- constant.sims.ac.15[30:45, ]
constant.sims.acv.15.day30 <- constant.sims.acv.15[30:45, ]
constant.sims.fdv.15.day30 <- constant.sims.fdv.15[30:45, ]
constant.sims.dt.15.day30 <- constant.sims.dt.15[30:45, ]

warming.sims.var.15.day30 <- warming.sims.var.15[30:45, ]
warming.sims.mean.15.day30 <- warming.sims.mean.15[30:45, ]
warming.sims.cv.15.day30 <- warming.sims.cv.15[30:45, ]
warming.sims.sk.15.day30 <- warming.sims.sk.15[30:45, ]
warming.sims.kt.15.day30 <- warming.sims.kt.15[30:45, ]
warming.sims.id.15.day30 <- warming.sims.id.15[30:45, ]
warming.sims.ac.15.day30 <- warming.sims.ac.15[30:45, ]
warming.sims.acv.15.day30 <- warming.sims.acv.15[30:45, ]
warming.sims.fdv.15.day30 <- warming.sims.fdv.15[30:45, ]
warming.sims.dt.15.day30 <- warming.sims.dt.15[30:45, ]

constant.sims.var.15.day30.tau <- c()
constant.sims.mean.15.day30.tau <- c()
constant.sims.cv.15.day30.tau <- c()
constant.sims.sk.15.day30.tau <- c()
constant.sims.kt.15.day30.tau <- c()
constant.sims.id.15.day30.tau <- c()
constant.sims.ac.15.day30.tau <- c()
constant.sims.acv.15.day30.tau <- c()
constant.sims.fdv.15.day30.tau <- c()
constant.sims.dt.15.day30.tau <- c()

warming.sims.var.15.day30.tau <- c()
warming.sims.mean.15.day30.tau <- c()
warming.sims.cv.15.day30.tau <- c()
warming.sims.sk.15.day30.tau <- c()
warming.sims.kt.15.day30.tau <- c()
warming.sims.id.15.day30.tau <- c()
warming.sims.ac.15.day30.tau <- c()
warming.sims.acv.15.day30.tau <- c()
warming.sims.fdv.15.day30.tau <- c()
warming.sims.dt.15.day30.tau <- c()

time.15 <- c(1:16)

for(i in 1:num.of.sims){
  constant.sims.mean.15.day30.tau[i] <- cor(constant.sims.mean.15.day30[ , i], time.15, method = "kendall")
  constant.sims.sk.15.day30.tau[i] <- cor(constant.sims.sk.15.day30[ , i], time.15, method = "kendall")
  constant.sims.kt.15.day30.tau[i] <- cor(constant.sims.kt.15.day30[ , i], time.15, method = "kendall")
  constant.sims.var.15.day30.tau[i] <- cor(constant.sims.var.15.day30[ , i], time.15, method = "kendall")
  constant.sims.cv.15.day30.tau[i] <- cor(constant.sims.cv.15.day30[ , i], time.15, method = "kendall")
  constant.sims.id.15.day30.tau[i] <- cor(constant.sims.id.15.day30[ , i], time.15, method = "kendall")
  constant.sims.fdv.15.day30.tau[i] <- cor(constant.sims.fdv.15.day30[ , i], time.15, method = "kendall")
  constant.sims.ac.15.day30.tau[i] <- cor(constant.sims.ac.15.day30[ , i], time.15, method = "kendall")
  constant.sims.acv.15.day30.tau[i] <- cor(constant.sims.acv.15.day30[ , i], time.15, method = "kendall")
  constant.sims.dt.15.day30.tau[i] <- cor(constant.sims.dt.15.day30[ , i], time.15, method = "kendall")
  
  warming.sims.mean.15.day30.tau[i] <- cor(warming.sims.mean.15.day30[ , i], time.15, method = "kendall")
  warming.sims.sk.15.day30.tau[i] <- cor(warming.sims.sk.15.day30[ , i], time.15, method = "kendall")
  warming.sims.kt.15.day30.tau[i] <- cor(warming.sims.kt.15.day30[ , i], time.15, method = "kendall")
  warming.sims.var.15.day30.tau[i] <- cor(warming.sims.var.15.day30[ , i], time.15, method = "kendall")
  warming.sims.cv.15.day30.tau[i] <- cor(warming.sims.cv.15.day30[ , i], time.15, method = "kendall")
  warming.sims.id.15.day30.tau[i] <- cor(warming.sims.id.15.day30[ , i], time.15, method = "kendall")
  warming.sims.fdv.15.day30.tau[i] <- cor(warming.sims.fdv.15.day30[ , i], time.15, method = "kendall")
  warming.sims.ac.15.day30.tau[i] <- cor(warming.sims.ac.15.day30[ , i], time.15, method = "kendall")
  warming.sims.acv.15.day30.tau[i] <- cor(warming.sims.acv.15.day30[ , i], time.15, method = "kendall")
  warming.sims.dt.15.day30.tau[i] <- cor(warming.sims.dt.15.day30[ , i], time.15, method = "kendall")
}

AUC_mean_15 <- calc_auc(warming.sims.mean.15.day30.tau, constant.sims.mean.15.day30.tau); AUC_mean_15
AUC_sk_15 <- calc_auc(warming.sims.sk.15.day30.tau, constant.sims.sk.15.day30.tau); AUC_sk_15
AUC_kt_15 <- calc_auc(warming.sims.kt.15.day30.tau, constant.sims.kt.15.day30.tau); AUC_kt_15
AUC_var_15 <- calc_auc(warming.sims.var.15.day30.tau, constant.sims.var.15.day30.tau); AUC_var_15
AUC_cv_15 <- calc_auc(warming.sims.cv.15.day30.tau, constant.sims.cv.15.day30.tau); AUC_cv_15
AUC_id_15 <- calc_auc(warming.sims.id.15.day30.tau, constant.sims.id.15.day30.tau); AUC_id_15
AUC_fdv_15 <- calc_auc(warming.sims.fdv.15.day30.tau, constant.sims.fdv.15.day30.tau); AUC_fdv_15
AUC_ac_15 <- calc_auc(warming.sims.ac.15.day30.tau, constant.sims.ac.15.day30.tau); AUC_ac_15
AUC_acv_15 <- calc_auc(warming.sims.acv.15.day30.tau, constant.sims.acv.15.day30.tau); AUC_acv_15
AUC_dt_15 <- calc_auc(warming.sims.dt.15.day30.tau, constant.sims.dt.15.day30.tau); AUC_dt_15
```

### Summarized Outputs

```{r}
metrics <- c("mean", "skewness", "kurtosis", "variance", "cv", "id", "fdv", "cor", "acv", "dt")
AUC_5_restricted <- c(AUC_mean_5, AUC_sk_5, AUC_kt_5, AUC_var_5, AUC_cv_5, 
                      AUC_id_5, AUC_fdv_5, AUC_ac_5 ,AUC_acv_5, AUC_dt_5)
AUC_15_restricted <- c(AUC_mean_15, AUC_sk_15, AUC_kt_15, AUC_var_15, AUC_cv_15, 
                       AUC_id_15, AUC_fdv_15, AUC_ac_15 ,AUC_acv_15, AUC_dt_15)
AUC_restricted_df <- data.frame(metrics, AUC_5_restricted, AUC_15_restricted); AUC_restricted_df

plot(1:10, AUC_restricted_df$AUC_5_restricted, type = "p", 
     main = "Simulated Data (Days 30 to 55)", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "orchid", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)

plot(1:10, AUC_restricted_df$AUC_15_restricted, type = "p", 
     main = "Simulated Data (Days 30 to 45)", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "orchid", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)

## Medians
paste("Median Mean in Control Populations: ", median(constant.sims.mean.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Mean in Test Populations: ", median(warming.sims.mean.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Skewness in Control Populations: ", median(constant.sims.sk.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Skewness in Test Populations: ", median(warming.sims.sk.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Kurtosis in Control Populations: ", median(constant.sims.kt.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Kurtosis in Test Populations: ", median(warming.sims.kt.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Variance in Control Populations: ", median(constant.sims.var.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Variance in Test Populations: ", median(warming.sims.var.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Co-Efficient of Variation in Control Populations: ", median(constant.sims.cv.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Co-Efficient of Variation in Test Populations: ", median(warming.sims.cv.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Index of Dispersion in Control Populations: ", median(constant.sims.id.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Index of Dispersion in Test Populations: ", median(warming.sims.id.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median First Difference Variance in Control Populations: ", median(constant.sims.fdv.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median First Difference Variance in Test Populations: ", median(warming.sims.fdv.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Autocorrelation in Control Populations: ", median(constant.sims.ac.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Autocorrelation in Test Populations: ", median(warming.sims.ac.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Autocovariance in Control Populations: ", median(constant.sims.acv.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Autocovariance in Test Populations: ", median(warming.sims.acv.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Decay Time in Control Populations: ", median(constant.sims.dt.5.day30.tau, na.rm = TRUE), sep = "")
paste("Median Decay Time in Test Populations: ", median(warming.sims.dt.5.day30.tau, na.rm = TRUE), sep = "")

paste("Median Mean in Control Populations: ", median(constant.sims.mean.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Mean in Test Populations: ", median(warming.sims.mean.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Skewness in Control Populations: ", median(constant.sims.sk.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Skewness in Test Populations: ", median(warming.sims.sk.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Kurtosis in Control Populations: ", median(constant.sims.kt.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Kurtosis in Test Populations: ", median(warming.sims.kt.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Variance in Control Populations: ", median(constant.sims.var.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Variance in Test Populations: ", median(warming.sims.var.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Co-Efficient of Variation in Control Populations: ", median(constant.sims.cv.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Co-Efficient of Variation in Test Populations: ", median(warming.sims.cv.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Index of Dispersion in Control Populations: ", median(constant.sims.id.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Index of Dispersion in Test Populations: ", median(warming.sims.id.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median First Difference Variance in Control Populations: ", median(constant.sims.fdv.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median First Difference Variance in Test Populations: ", median(warming.sims.fdv.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Autocorrelation in Control Populations: ", median(constant.sims.ac.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Autocorrelation in Test Populations: ", median(warming.sims.ac.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Autocovariance in Control Populations: ", median(constant.sims.acv.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Autocovariance in Test Populations: ", median(warming.sims.acv.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Decay Time in Control Populations: ", median(constant.sims.dt.15.day30.tau, na.rm = TRUE), sep = "")
paste("Median Decay Time in Test Populations: ", median(warming.sims.dt.15.day30.tau, na.rm = TRUE), sep = "")
```

### Plotting

```{r}
cols <- c(rep("#1434A4", 512), rep("#C21E56", 512))

## Mean
sim.mean.15.day30.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.mean.15.day30.tau, warming.sims.mean.15.day30.tau))
med.const.sims.mean.15.day30.tau <- median(constant.sims.mean.15.day30.tau)
med.warm.sims.mean.15.day30.tau <- median(warming.sims.mean.15.day30.tau)
dens.mean <- ggplot(data = sim.mean.15.day30.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.mean.15.day30.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.mean.15.day30.tau, colour = "#C21E56") +
  ggtitle("Mean: 79% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## Variance
sim.var.15.day30.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.var.15.day30.tau, warming.sims.var.15.day30.tau))
med.const.sims.var.15.day30.tau <- median(constant.sims.var.15.day30.tau)
med.warm.sims.var.15.day30.tau <- median(warming.sims.var.15.day30.tau)
dens.var <- ggplot(data = sim.var.15.day30.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.var.15.day30.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.var.15.day30.tau, colour = "#C21E56") +
  ggtitle("Variance: 91% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## Skewness
sim.sk.15.day30.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.sk.15.day30.tau, warming.sims.sk.15.day30.tau))
med.const.sims.sk.15.day30.tau <- median(constant.sims.sk.15.day30.tau)
med.warm.sims.sk.15.day30.tau <- median(warming.sims.sk.15.day30.tau)
dens.sk <- ggplot(data = sim.sk.15.day30.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.sk.15.day30.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.sk.15.day30.tau, colour = "#C21E56") +
  ggtitle("Skewness: 95% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## Kurtosis
sim.kt.15.day30.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.kt.15.day30.tau, warming.sims.kt.15.day30.tau))
med.const.sims.kt.15.day30.tau <- median(constant.sims.kt.15.day30.tau)
med.warm.sims.kt.15.day30.tau <- median(warming.sims.kt.15.day30.tau)
dens.kt <- ggplot(data = sim.kt.15.day30.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.kt.15.day30.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.kt.15.day30.tau, colour = "#C21E56") +
  ggtitle("Kurtosis: 96% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## Co-Efficient of Variation
sim.cv.15.day30.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.cv.15.day30.tau, warming.sims.cv.15.day30.tau))
med.const.sims.cv.15.day30.tau <- median(constant.sims.cv.15.day30.tau)
med.warm.sims.cv.15.day30.tau <- median(warming.sims.cv.15.day30.tau)
dens.cv <- ggplot(data = sim.cv.15.day30.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.cv.15.day30.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.cv.15.day30.tau, colour = "#C21E56") +
  ggtitle("Co-Efficient of Variation: 94% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## Index of Dispersion
sim.id.15.day30.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.id.15.day30.tau, warming.sims.id.15.day30.tau))
med.const.sims.id.15.day30.tau <- median(constant.sims.id.15.day30.tau)
med.warm.sims.id.15.day30.tau <- median(warming.sims.id.15.day30.tau)
dens.id <- ggplot(data = sim.id.15.day30.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.id.15.day30.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.id.15.day30.tau, colour = "#C21E56") +
  ggtitle("Index of Dispersion: 93% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## First Difference Variance
sim.fdv.15.day30.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.fdv.15.day30.tau, warming.sims.fdv.15.day30.tau))
med.const.sims.fdv.15.day30.tau <- median(constant.sims.fdv.15.day30.tau)
med.warm.sims.fdv.15.day30.tau <- median(warming.sims.fdv.15.day30.tau)
dens.fdv <- ggplot(data = sim.fdv.15.day30.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.fdv.15.day30.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.fdv.15.day30.tau, colour = "#C21E56") +
  ggtitle("First Difference Variance: 95% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## Autocorrelation
sim.ac.15.day30.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.ac.15.day30.tau, warming.sims.ac.15.day30.tau))
med.const.sims.ac.15.day30.tau <- median(constant.sims.ac.15.day30.tau)
med.warm.sims.ac.15.day30.tau <- median(warming.sims.ac.15.day30.tau)
dens.ac <- ggplot(data = sim.ac.15.day30.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.ac.15.day30.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.ac.15.day30.tau, colour = "#C21E56") +
  ggtitle("Autocorrelation: 96% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## Autocovariance
sim.acv.15.day30.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.acv.15.day30.tau, warming.sims.acv.15.day30.tau))
med.const.sims.acv.15.day30.tau <- median(constant.sims.acv.15.day30.tau)
med.warm.sims.acv.15.day30.tau <- median(warming.sims.acv.15.day30.tau)
dens.acv <- ggplot(data = sim.acv.15.day30.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.acv.15.day30.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.acv.15.day30.tau, colour = "#C21E56") +
  ggtitle("Autocovariance: 93% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

## Decay Time
sim.dt.15.day30.tau <- data.frame(Treatment = c(rep("Constant", 1000), rep("Warming", 1000)), k.tau = c(constant.sims.dt.15.day30.tau, warming.sims.dt.15.day30.tau))
med.const.sims.dt.15.day30.tau <- median(constant.sims.dt.15.day30.tau)
med.warm.sims.dt.15.day30.tau <- median(warming.sims.dt.15.day30.tau)
dens.dt <- ggplot(data = sim.dt.15.day30.tau, aes(x = k.tau, group = Treatment, fill = Treatment)) +
  geom_density(alpha = 0.4, colour = cols) +
  scale_fill_manual(values = c("#1434A4", "#C21E56")) +
  geom_vline(xintercept = med.const.sims.dt.15.day30.tau, colour = "#1434A4") +
  geom_vline(xintercept = med.warm.sims.dt.15.day30.tau, colour = "#C21E56") +
  ggtitle("Decay Time: 95% Overlap") +
  xlab(" ") +
  ylab(" ") +
  xlim(-1, 1) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(legend.position = "none")

sim.taus.day30.plot <- plot_grid(dens.mean, dens.sk, dens.kt, dens.var, dens.cv,
                                 dens.id, dens.fdv, dens.ac, dens.acv, dens.dt,
                                 nrow = 2, align = "h"
                                 # , labels = "AUTO", label_fontfamily = "serif"
                                 )
```

## Empirical Data

```{r}
data <- read.csv("Kirk_Data/Epidemics_Data_Dryad/Kirk_et_al_epidemics_sampling_data.csv")
data.sum <- summaryBy(status ~ population + treatment + day, data = data, FUN = sum)
sub1 <- subset(data.sum, population == 1)
sub2 <- subset(data.sum, population == 2)
sub3 <- subset(data.sum, population == 3)
sub4 <- subset(data.sum, population == 4)
sub5 <- subset(data.sum, population == 5)
sub6 <- subset(data.sum, population == 6)
sub7 <- subset(data.sum, population == 7)
sub8 <- subset(data.sum, population == 8)
```

### Calculating Metrics (15 Day Sliding Window)

```{r}
## Standard deviation
pop1.sd <- SlidingWindow(FUN = "sd", data = sub1$status.sum, window = 5, step = 1)
pop2.sd <- SlidingWindow(FUN = "sd", data = sub2$status.sum, window = 5, step = 1)
pop3.sd <- SlidingWindow(FUN = "sd", data = sub3$status.sum, window = 5, step = 1)
pop4.sd <- SlidingWindow(FUN = "sd", data = sub4$status.sum, window = 5, step = 1)
pop5.sd <- SlidingWindow(FUN = "sd", data = sub5$status.sum, window = 5, step = 1)
pop6.sd <- SlidingWindow(FUN = "sd", data = sub6$status.sum, window = 5, step = 1)
pop7.sd <- SlidingWindow(FUN = "sd", data = sub7$status.sum, window = 5, step = 1)
pop8.sd <- SlidingWindow(FUN = "sd", data = sub8$status.sum, window = 5, step = 1)

## Mean
pop1.mean <- SlidingWindow(FUN = "mean", data = sub1$status.sum, window = 5, step = 1)
pop2.mean <- SlidingWindow(FUN = "mean", data = sub2$status.sum, window = 5, step = 1)
pop3.mean <- SlidingWindow(FUN = "mean", data = sub3$status.sum, window = 5, step = 1)
pop4.mean <- SlidingWindow(FUN = "mean", data = sub4$status.sum, window = 5, step = 1)
pop5.mean <- SlidingWindow(FUN = "mean", data = sub5$status.sum, window = 5, step = 1)
pop6.mean <- SlidingWindow(FUN = "mean", data = sub6$status.sum, window = 5, step = 1)
pop7.mean <- SlidingWindow(FUN = "mean", data = sub7$status.sum, window = 5, step = 1)
pop8.mean <- SlidingWindow(FUN = "mean", data = sub8$status.sum, window = 5, step = 1)

## Skewness
pop1.sk <- SlidingWindow(FUN = "skewness", data = sub1$status.sum, window = 5, step = 1)
pop2.sk <- SlidingWindow(FUN = "skewness", data = sub2$status.sum, window = 5, step = 1)
pop3.sk <- SlidingWindow(FUN = "skewness", data = sub3$status.sum, window = 5, step = 1)
pop4.sk <- SlidingWindow(FUN = "skewness", data = sub4$status.sum, window = 5, step = 1)
pop5.sk <- SlidingWindow(FUN = "skewness", data = sub5$status.sum, window = 5, step = 1)
pop6.sk <- SlidingWindow(FUN = "skewness", data = sub6$status.sum, window = 5, step = 1)
pop7.sk <- SlidingWindow(FUN = "skewness", data = sub7$status.sum, window = 5, step = 1)
pop8.sk <- SlidingWindow(FUN = "skewness", data = sub8$status.sum, window = 5, step = 1)

## Kurtosis
pop1.kt <- SlidingWindow(FUN = "kurtosis", data = sub1$status.sum, window = 5, step = 1)
pop2.kt <- SlidingWindow(FUN = "kurtosis", data = sub2$status.sum, window = 5, step = 1)
pop3.kt <- SlidingWindow(FUN = "kurtosis", data = sub3$status.sum, window = 5, step = 1)
pop4.kt <- SlidingWindow(FUN = "kurtosis", data = sub4$status.sum, window = 5, step = 1)
pop5.kt <- SlidingWindow(FUN = "kurtosis", data = sub5$status.sum, window = 5, step = 1)
pop6.kt <- SlidingWindow(FUN = "kurtosis", data = sub6$status.sum, window = 5, step = 1)
pop7.kt <- SlidingWindow(FUN = "kurtosis", data = sub7$status.sum, window = 5, step = 1)
pop8.kt <- SlidingWindow(FUN = "kurtosis", data = sub8$status.sum, window = 5, step = 1)

## Variance
pop1.var <- pop1.sd^2
pop2.var <- pop2.sd^2
pop3.var <- pop3.sd^2
pop4.var <- pop4.sd^2
pop5.var <- pop5.sd^2
pop6.var <- pop6.sd^2
pop7.var <- pop7.sd^2
pop8.var <- pop8.sd^2

## Co-Efficient of Variation
pop1.cv <- pop1.sd/(pop1.mean + 1e-10)
pop2.cv <- pop2.sd/(pop2.mean + 1e-10)
pop3.cv <- pop3.sd/(pop3.mean + 1e-10)
pop4.cv <- pop4.sd/(pop4.mean + 1e-10)
pop5.cv <- pop5.sd/(pop5.mean + 1e-10)
pop6.cv <- pop6.sd/(pop6.mean + 1e-10)
pop7.cv <- pop7.sd/(pop7.mean + 1e-10)
pop8.cv <- pop8.sd/(pop8.mean + 1e-10)

## Index of Dispersion
pop1.id <- pop1.var/(pop1.mean + 1e-10)
pop2.id <- pop2.var/(pop2.mean + 1e-10)
pop3.id <- pop3.var/(pop3.mean + 1e-10)
pop4.id <- pop4.var/(pop4.mean + 1e-10)
pop5.id <- pop5.var/(pop5.mean + 1e-10)
pop6.id <- pop6.var/(pop6.mean + 1e-10)
pop7.id <- pop7.var/(pop7.mean + 1e-10)
pop8.id <- pop8.var/(pop8.mean + 1e-10)

## First Difference Variance
pop1.fdv <- c()
pop2.fdv <- c()
pop3.fdv <- c()
pop4.fdv <- c()
pop5.fdv <- c()
pop6.fdv <- c()
pop7.fdv <- c()
pop8.fdv <- c()
for(i in 2:35){
pop1.fdv[i - 1] <- pop1.var[i] - pop1.var[i - 1]
pop2.fdv[i - 1] <- pop2.var[i] - pop2.var[i - 1]
pop3.fdv[i - 1] <- pop3.var[i] - pop3.var[i - 1]
pop4.fdv[i - 1] <- pop4.var[i] - pop4.var[i - 1]
pop5.fdv[i - 1] <- pop5.var[i] - pop5.var[i - 1]
pop6.fdv[i - 1] <- pop6.var[i] - pop6.var[i - 1]
pop7.fdv[i - 1] <- pop7.var[i] - pop7.var[i - 1]
pop8.fdv[i - 1] <- pop8.var[i] - pop8.var[i - 1]
}

## Autocorrelation and Autocovariance
pop1.ac <- c()
pop1.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  segment <- sub1$status.sum[c(window.start:window.end)]
  pop1.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop1.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}
pop2.ac <- c()
pop2.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  segment <- sub2$status.sum[c(window.start:window.end)]
  pop2.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop2.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}
pop3.ac <- c()
pop3.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  segment <- sub3$status.sum[c(window.start:window.end)]
  pop3.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop3.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}
pop4.ac <- c()
pop4.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  segment <- sub4$status.sum[c(window.start:window.end)]
  pop4.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop4.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}
pop5.ac <- c()
pop5.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  segment <- sub5$status.sum[c(window.start:window.end)]
  pop5.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop5.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}
pop6.ac <- c()
pop6.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  segment <- sub6$status.sum[c(window.start:window.end)]
  pop6.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop6.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}
pop7.ac <- c()
pop7.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  segment <- sub7$status.sum[c(window.start:window.end)]
  pop7.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop7.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}
pop8.ac <- c()
pop8.acv <- c()
for(i in 1:35){
  window.start <- i
  window.end <- i + 4
  segment <- sub8$status.sum[c(window.start:window.end)]
  pop8.ac[i] <- acf(segment, plot = FALSE, type = "correlation")[[1]][2]
  pop8.acv[i] <- acf(segment, plot = FALSE, type = "covariance")[[1]][2]
}

## Decay Time
pop1.dt <- c()
pop2.dt <- c()
pop3.dt <- c()
pop4.dt <- c()
pop5.dt <- c()
pop6.dt <- c()
pop7.dt <- c()
pop8.dt <- c()
for(i in 1:35){
  suppressWarnings(pop1.dt[i] <- -i/log(pop1.ac[i]))
  suppressWarnings(pop2.dt[i] <- -i/log(pop2.ac[i]))
  suppressWarnings(pop3.dt[i] <- -i/log(pop3.ac[i]))
  suppressWarnings(pop4.dt[i] <- -i/log(pop4.ac[i]))
  suppressWarnings(pop5.dt[i] <- -i/log(pop5.ac[i]))
  suppressWarnings(pop6.dt[i] <- -i/log(pop6.ac[i]))
  suppressWarnings(pop7.dt[i] <- -i/log(pop7.ac[i]))
  suppressWarnings(pop8.dt[i] <- -i/log(pop8.ac[i]))
}
```

### Kendall's Tau and AUC (15 Day Sliding Window) \| Days 21 to 60

```{r}
time.15 <- seq(1, 9, 1)

constant.mean.tau.day21 <- c(cor(pop1.mean[7:15], time.15, method = "kendall"), cor(pop2.mean[7:15], time.15, method = "kendall"),
                      cor(pop3.mean[7:15], time.15, method = "kendall"), cor(pop4.mean[7:15], time.15, method = "kendall"))
warming.mean.tau.day21 <- c(cor(pop5.mean[7:15], time.15, method = "kendall"), cor(pop6.mean[7:15], time.15, method = "kendall"),
                      cor(pop7.mean[7:15], time.15, method = "kendall"), cor(pop8.mean[7:15], time.15, method = "kendall"))

constant.sk.tau.day21 <- c(cor(pop1.sk[7:15], time.15, method = "kendall"), cor(pop2.sk[7:15], time.15, method = "kendall"),
                          cor(pop3.sk[7:15], time.15, method = "kendall"), cor(pop4.sk[7:15], time.15, method = "kendall"))
warming.sk.tau.day21 <- c(cor(pop5.sk[7:15], time.15, method = "kendall"), cor(pop6.sk[7:15], time.15, method = "kendall"),
                          cor(pop7.sk[7:15], time.15, method = "kendall"), cor(pop8.sk[7:15], time.15, method = "kendall"))

constant.kt.tau.day21 <- c(cor(pop1.kt[7:15], time.15, method = "kendall"), cor(pop2.kt[7:15], time.15, method = "kendall"),
                          cor(pop3.kt[7:15], time.15, method = "kendall"), cor(pop4.kt[7:15], time.15, method = "kendall"))
warming.kt.tau.day21 <- c(cor(pop5.kt[7:15], time.15, method = "kendall"), cor(pop6.kt[7:15], time.15, method = "kendall"),
                          cor(pop7.kt[7:15], time.15, method = "kendall"), cor(pop8.kt[7:15], time.15, method = "kendall"))

constant.var.tau.day21 <- c(cor(pop1.var[7:15], time.15, method = "kendall"), cor(pop2.var[7:15], time.15, method = "kendall"),
                          cor(pop3.var[7:15], time.15, method = "kendall"), cor(pop4.var[7:15], time.15, method = "kendall"))
warming.var.tau.day21<- c(cor(pop5.var[7:15], time.15, method = "kendall"), cor(pop6.var[7:15], time.15, method = "kendall"),
                         cor(pop7.var[7:15], time.15, method = "kendall"), cor(pop8.var[7:15], time.15, method = "kendall"))

constant.cv.tau.day21 <- c(cor(pop1.cv[7:15], time.15, method = "kendall"), cor(pop2.cv[7:15], time.15, method = "kendall"),
                    cor(pop3.cv[7:15], time.15, method = "kendall"), cor(pop4.cv[7:15], time.15, method = "kendall"))
warming.cv.tau.day21 <- c(cor(pop5.cv[7:15], time.15, method = "kendall"), cor(pop6.cv[7:15], time.15, method = "kendall"),
                    cor(pop7.cv[7:15], time.15, method = "kendall"), cor(pop8.cv[7:15], time.15, method = "kendall"))

constant.id.tau.day21 <- c(cor(pop1.id[7:15], time.15, method = "kendall"), cor(pop2.id[7:15], time.15, method = "kendall"),
                    cor(pop3.id[7:15], time.15, method = "kendall"), cor(pop4.id[7:15], time.15, method = "kendall"))
warming.id.tau.day21 <- c(cor(pop5.id[7:15], time.15, method = "kendall"), cor(pop6.id[7:15], time.15, method = "kendall"),
                    cor(pop7.id[7:15], time.15, method = "kendall"), cor(pop8.id[7:15], time.15, method = "kendall"))

constant.fdv.tau.day21 <- c(cor(pop1.fdv[7:15], time.15, method = "kendall"), cor(pop2.fdv[7:15], time.15, method = "kendall"),
                                     cor(pop3.fdv[7:15], time.15, method = "kendall"), cor(pop4.fdv[7:15], time.15, method = "kendall"))
warming.fdv.tau.day21 <- c(cor(pop5.fdv[7:15], time.15, method = "kendall"), cor(pop6.fdv[7:15], time.15, method = "kendall"),
                                     cor(pop7.fdv[7:15], time.15, method = "kendall"), cor(pop8.fdv[7:15], time.15, method = "kendall"))

constant.ac.tau.day21 <- c(cor(pop1.ac[7:15], time.15, method = "kendall"), cor(pop2.ac[7:15], time.15, method = "kendall"),
                    cor(pop3.ac[7:15], time.15, method = "kendall"), cor(pop4.ac[7:15], time.15, method = "kendall"))
warming.ac.tau.day21 <- c(cor(pop5.ac[7:15], time.15, method = "kendall"), cor(pop6.ac[7:15], time.15, method = "kendall"),
                    cor(pop7.ac[7:15], time.15, method = "kendall"), cor(pop8.ac[7:15], time.15, method = "kendall"))

constant.acv.tau.day21 <- c(cor(pop1.acv[7:15], time.15, method = "kendall"), cor(pop2.acv[7:15], time.15, method = "kendall"),
                     cor(pop3.acv[7:15], time.15, method = "kendall"), cor(pop4.acv[7:15], time.15, method = "kendall"))
warming.acv.tau.day21 <- c(cor(pop5.acv[7:15], time.15, method = "kendall"), cor(pop6.acv[7:15], time.15, method = "kendall"),
                     cor(pop7.acv[7:15], time.15, method = "kendall"), cor(pop8.acv[7:15], time.15, method = "kendall"))

constant.dt.tau.day21 <- c(cor(pop1.dt[7:15], time.15, method = "kendall"), cor(pop2.dt[7:15], time.15, method = "kendall"),
                            cor(pop3.dt[7:15], time.15, method = "kendall"), cor(pop4.dt[7:15], time.15, method = "kendall"))
warming.dt.tau.day21 <- c(cor(pop5.dt[7:15], time.15, method = "kendall"), cor(pop6.dt[7:15], time.15, method = "kendall"),
                            cor(pop7.dt[7:15], time.15, method = "kendall"), cor(pop8.dt[7:15], time.15, method = "kendall"))

AUC_mean_data <- calc_auc(warming.mean.tau.day21, constant.mean.tau.day21); AUC_mean_data
AUC_sk_data <- calc_auc(warming.sk.tau.day21, constant.sk.tau.day21); AUC_sk_data
AUC_kt_data <- calc_auc(warming.kt.tau.day21, constant.kt.tau.day21); AUC_kt_data
AUC_var_data <- calc_auc(warming.var.tau.day21, constant.var.tau.day21); AUC_var_data
AUC_cv_data <- calc_auc(warming.cv.tau.day21, constant.cv.tau.day21); AUC_cv_data
AUC_id_data <- calc_auc(warming.id.tau.day21, constant.id.tau.day21); AUC_id_data
AUC_fdv_data <- calc_auc(warming.fdv.tau.day21, constant.fdv.tau.day21); AUC_fdv_data
AUC_ac_data <- calc_auc(warming.ac.tau.day21, constant.ac.tau.day21); AUC_ac_data
AUC_acv_data <- calc_auc(warming.acv.tau.day21, constant.acv.tau.day21); AUC_acv_data
AUC_dt_data <- calc_auc(warming.dt.tau.day21, constant.dt.tau.day21); AUC_dt_data
```

### Summarized Outputs

```{r}
metrics <- c("mean", "skewness", "kurtosis", "variance", "cv", "id", "fdv", "ac", "acv", "dt")
AUC_data_restrict <- c(AUC_mean_data, AUC_sk_data, AUC_kt_data, AUC_var_data, AUC_cv_data, AUC_id_data, AUC_fdv_data, AUC_ac_data, AUC_acv_data, AUC_dt_data)
AUC_df_restrict <- data.frame(metrics, AUC_data_restrict)

metrics_full <- c("Mean", "Skewness", "Kurtosis", "Variance", "Covariance", "Index of Dispersion", "First Difference Variance", "Autocorrelation", "Autocovariance", "Decay Time")
plot(1:10, AUC_df_restrict$AUC_data, type = "p", 
     main = "Empirical Data, Days 21 to 45", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "orchid", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)
```

### Plotting
```{r}
emp.ktaus <- data.frame(
  mean = c(constant.mean.tau.day21, warming.mean.tau.day21),
  sk = c(constant.sk.tau.day21, warming.sk.tau.day21),
  kt = c(constant.kt.tau.day21, warming.kt.tau.day21),
  var = c(constant.var.tau.day21, warming.var.tau.day21),
  cv = c(constant.cv.tau.day21, warming.cv.tau.day21),
  id = c(constant.id.tau.day21, warming.id.tau.day21),
  fdv = c(constant.fdv.tau.day21, warming.fdv.tau.day21),
  ac = c(constant.ac.tau.day21, warming.ac.tau.day21),
  acv = c(constant.acv.tau.day21, warming.acv.tau.day21),
  dt = c(constant.dt.tau.day21, warming.dt.tau.day21)
)
rownames(emp.ktaus) <- c("constant.1", "constant.2", "constant.3", "constant.4", "warming.1", "warming.2", "warming.3", "warming.4")

x.labs <- c("Mean", "Skewness", "Kurtosis", 
                              "Variance", "Co-Efficient of Variation", "Index of Dispersion",
                              "First Difference Variance", "Autocorrelation", "Autocovariance",
                              "Decay Time")
par(mar = c(6.5, 4, 4, 4)) # Order: 'bottom', 'left', 'top', 'right'.
plot(x = 1, type = "n", xaxt = "n", yaxt = "n", 
     main = "Experimental Data", xlab = "", ylab = "Kendall's Tau",
     xlim = c(1, 10), ylim = c(-1, 1))
text(1:10, emp.ktaus[1, ], label = 1, col = alpha("#1434A4", 0.5))
text(1:10, emp.ktaus[2, ], label = 2, col = alpha("#1434A4", 0.5))
text(1:10, emp.ktaus[3, ], label = 3, col = alpha("#1434A4", 0.5))
text(1:10, emp.ktaus[4, ], label = 4, col = alpha("#1434A4", 0.5))
text(1:10, emp.ktaus[5, ], label = 5, col = alpha("#C21E56", 0.5))
text(1:10, emp.ktaus[6, ], label = 6, col = alpha("#C21E56", 0.5))
text(1:10, emp.ktaus[7, ], label = 7, col = alpha("#C21E56", 0.5))
text(1:10, emp.ktaus[8, ], label = 8, col = alpha("#C21E56", 0.5))
abline(h = 0, lty = 3)
axis(2, at = c(-1, -0.5, 0, 0.5, 1))
text(cex = 1, x = 1:10, y = -1.4, x.labs, xpd = TRUE, srt = 35)
legend(8, 1.1, legend = c("Constant Tempertaure", "Rising Temperature"),
       pch = 19, col = c("#1434A4", "#C21E56"), cex = 0.8, bty = "n")
```

### Kendall's Tau and AUC (15 Day Sliding Window) \| Days 30 to 60

```{r}
time.15 <- seq(1, 6, 1)

constant.mean.tau.day30 <- c(cor(pop1.mean[10:15], time.15, method = "kendall"), cor(pop2.mean[10:15], time.15, method = "kendall"),
                      cor(pop3.mean[10:15], time.15, method = "kendall"), cor(pop4.mean[10:15], time.15, method = "kendall"))
warming.mean.tau.day30 <- c(cor(pop5.mean[10:15], time.15, method = "kendall"), cor(pop6.mean[10:15], time.15, method = "kendall"),
                      cor(pop7.mean[10:15], time.15, method = "kendall"), cor(pop8.mean[10:15], time.15, method = "kendall"))

constant.sk.tau.day30 <- c(cor(pop1.sk[10:15], time.15, method = "kendall"), cor(pop2.sk[10:15], time.15, method = "kendall"),
                          cor(pop3.sk[10:15], time.15, method = "kendall"), cor(pop4.sk[10:15], time.15, method = "kendall"))
warming.sk.tau.day30 <- c(cor(pop5.sk[10:15], time.15, method = "kendall"), cor(pop6.sk[10:15], time.15, method = "kendall"),
                          cor(pop7.sk[10:15], time.15, method = "kendall"), cor(pop8.sk[10:15], time.15, method = "kendall"))

constant.kt.tau.day30 <- c(cor(pop1.kt[10:15], time.15, method = "kendall"), cor(pop2.kt[10:15], time.15, method = "kendall"),
                          cor(pop3.kt[10:15], time.15, method = "kendall"), cor(pop4.kt[10:15], time.15, method = "kendall"))
warming.kt.tau.day30 <- c(cor(pop5.kt[10:15], time.15, method = "kendall"), cor(pop6.kt[10:15], time.15, method = "kendall"),
                          cor(pop7.kt[10:15], time.15, method = "kendall"), cor(pop8.kt[10:15], time.15, method = "kendall"))

constant.var.tau.day30 <- c(cor(pop1.var[10:15], time.15, method = "kendall"), cor(pop2.var[10:15], time.15, method = "kendall"),
                          cor(pop3.var[10:15], time.15, method = "kendall"), cor(pop4.var[10:15], time.15, method = "kendall"))
warming.var.tau.day30<- c(cor(pop5.var[10:15], time.15, method = "kendall"), cor(pop6.var[10:15], time.15, method = "kendall"),
                         cor(pop7.var[10:15], time.15, method = "kendall"), cor(pop8.var[10:15], time.15, method = "kendall"))

constant.cv.tau.day30 <- c(cor(pop1.cv[10:15], time.15, method = "kendall"), cor(pop2.cv[10:15], time.15, method = "kendall"),
                    cor(pop3.cv[10:15], time.15, method = "kendall"), cor(pop4.cv[10:15], time.15, method = "kendall"))
warming.cv.tau.day30 <- c(cor(pop5.cv[10:15], time.15, method = "kendall"), cor(pop6.cv[10:15], time.15, method = "kendall"),
                    cor(pop7.cv[10:15], time.15, method = "kendall"), cor(pop8.cv[10:15], time.15, method = "kendall"))

constant.id.tau.day30 <- c(cor(pop1.id[10:15], time.15, method = "kendall"), cor(pop2.id[10:15], time.15, method = "kendall"),
                    cor(pop3.id[10:15], time.15, method = "kendall"), cor(pop4.id[10:15], time.15, method = "kendall"))
warming.id.tau.day30 <- c(cor(pop5.id[10:15], time.15, method = "kendall"), cor(pop6.id[10:15], time.15, method = "kendall"),
                    cor(pop7.id[10:15], time.15, method = "kendall"), cor(pop8.id[10:15], time.15, method = "kendall"))

constant.fdv.tau.day30 <- c(cor(pop1.fdv[10:15], time.15, method = "kendall"), cor(pop2.fdv[10:15], time.15, method = "kendall"),
                                     cor(pop3.fdv[10:15], time.15, method = "kendall"), cor(pop4.fdv[10:15], time.15, method = "kendall"))
warming.fdv.tau.day30 <- c(cor(pop5.fdv[10:15], time.15, method = "kendall"), cor(pop6.fdv[10:15], time.15, method = "kendall"),
                                     cor(pop7.fdv[10:15], time.15, method = "kendall"), cor(pop8.fdv[10:15], time.15, method = "kendall"))

constant.ac.tau.day30 <- c(cor(pop1.ac[10:15], time.15, method = "kendall"), cor(pop2.ac[10:15], time.15, method = "kendall"),
                    cor(pop3.ac[10:15], time.15, method = "kendall"), cor(pop4.ac[10:15], time.15, method = "kendall"))
warming.ac.tau.day30 <- c(cor(pop5.ac[10:15], time.15, method = "kendall"), cor(pop6.ac[10:15], time.15, method = "kendall"),
                    cor(pop7.ac[10:15], time.15, method = "kendall"), cor(pop8.ac[10:15], time.15, method = "kendall"))

constant.acv.tau.day30 <- c(cor(pop1.acv[10:15], time.15, method = "kendall"), cor(pop2.acv[10:15], time.15, method = "kendall"),
                     cor(pop3.acv[10:15], time.15, method = "kendall"), cor(pop4.acv[10:15], time.15, method = "kendall"))
warming.acv.tau.day30 <- c(cor(pop5.acv[10:15], time.15, method = "kendall"), cor(pop6.acv[10:15], time.15, method = "kendall"),
                     cor(pop7.acv[10:15], time.15, method = "kendall"), cor(pop8.acv[10:15], time.15, method = "kendall"))

constant.dt.tau.day30 <- c(cor(pop1.dt[10:15], time.15, method = "kendall"), cor(pop2.dt[10:15], time.15, method = "kendall"),
                            cor(pop3.dt[10:15], time.15, method = "kendall"), cor(pop4.dt[10:15], time.15, method = "kendall"))
warming.dt.tau.day30 <- c(cor(pop5.dt[10:15], time.15, method = "kendall"), cor(pop6.dt[10:15], time.15, method = "kendall"),
                            cor(pop7.dt[10:15], time.15, method = "kendall"), cor(pop8.dt[10:15], time.15, method = "kendall"))

AUC_mean_data <- calc_auc(warming.mean.tau.day30, constant.mean.tau.day30); AUC_mean_data
AUC_sk_data <- calc_auc(warming.sk.tau.day30, constant.sk.tau.day30); AUC_sk_data
AUC_kt_data <- calc_auc(warming.kt.tau.day30, constant.kt.tau.day30); AUC_kt_data
AUC_var_data <- calc_auc(warming.var.tau.day30, constant.var.tau.day30); AUC_var_data
AUC_cv_data <- calc_auc(warming.cv.tau.day30, constant.cv.tau.day30); AUC_cv_data
AUC_id_data <- calc_auc(warming.id.tau.day30, constant.id.tau.day30); AUC_id_data
AUC_fdv_data <- calc_auc(warming.fdv.tau.day30, constant.fdv.tau.day30); AUC_fdv_data
AUC_ac_data <- calc_auc(warming.ac.tau.day30, constant.ac.tau.day30); AUC_ac_data
AUC_acv_data <- calc_auc(warming.acv.tau.day30, constant.acv.tau.day30); AUC_acv_data
AUC_dt_data <- calc_auc(warming.dt.tau.day30, constant.dt.tau.day30); AUC_dt_data
```

### Summarized Outputs

```{r}
metrics <- c("mean", "skewness", "kurtosis", "variance", "cv", "id", "fdv", "ac", "acv", "dt")
AUC_data_restrict <- c(AUC_mean_data, AUC_sk_data, AUC_kt_data, AUC_var_data, AUC_cv_data, AUC_id_data, AUC_fdv_data, AUC_ac_data, AUC_acv_data, AUC_dt_data)
AUC_df_restrict <- data.frame(metrics, AUC_data_restrict)

metrics_full <- c("Mean", "Skewness", "Kurtosis", "Variance", "Covariance", "Index of Dispersion", "First Difference Variance", "Autocorrelation", "Autocovariance", "Decay Time")
plot(1:10, AUC_df_restrict$AUC_data, type = "p", 
     main = "Empirical Data, Days 30 to 45", xaxt = "none", xlab = "", ylim = c(0, 1), ylab = "AUC",
     col = "orchid", pch = 19)
abline(h = 0.5, lty = 3)
axis(side = 1, at = 1:10, labels = metrics_full, las = 2)
```

### Plotting
```{r}
emp.ktaus <- data.frame(
  mean = c(constant.mean.tau.day30, warming.mean.tau.day30),
  sk = c(constant.sk.tau.day30, warming.sk.tau.day30),
  kt = c(constant.kt.tau.day30, warming.kt.tau.day30),
  var = c(constant.var.tau.day30, warming.var.tau.day30),
  cv = c(constant.cv.tau.day30, warming.cv.tau.day30),
  id = c(constant.id.tau.day30, warming.id.tau.day30),
  fdv = c(constant.fdv.tau.day30, warming.fdv.tau.day30),
  ac = c(constant.ac.tau.day30, warming.ac.tau.day30),
  acv = c(constant.acv.tau.day30, warming.acv.tau.day30),
  dt = c(constant.dt.tau.day30, warming.dt.tau.day30)
)
rownames(emp.ktaus) <- c("constant.1", "constant.2", "constant.3", "constant.4", "warming.1", "warming.2", "warming.3", "warming.4")

x.labs <- c("Mean", "Skewness", "Kurtosis", 
                              "Variance", "Co-Efficient of Variation", "Index of Dispersion",
                              "First Difference Variance", "Autocorrelation", "Autocovariance",
                              "Decay Time")
par(mar = c(6.5, 4, 4, 4)) # Order: 'bottom', 'left', 'top', 'right'.
plot(x = 1, type = "n", xaxt = "n", yaxt = "n", 
     main = "Experimental Data", xlab = "", ylab = "Kendall's Tau",
     xlim = c(1, 10), ylim = c(-1, 1))
text(1:10, emp.ktaus[1, ], label = 1, col = alpha("#1434A4", 0.5))
text(1:10, emp.ktaus[2, ], label = 2, col = alpha("#1434A4", 0.5))
text(1:10, emp.ktaus[3, ], label = 3, col = alpha("#1434A4", 0.5))
text(1:10, emp.ktaus[4, ], label = 4, col = alpha("#1434A4", 0.5))
text(1:10, emp.ktaus[5, ], label = 5, col = alpha("#C21E56", 0.5))
text(1:10, emp.ktaus[6, ], label = 6, col = alpha("#C21E56", 0.5))
text(1:10, emp.ktaus[7, ], label = 7, col = alpha("#C21E56", 0.5))
text(1:10, emp.ktaus[8, ], label = 8, col = alpha("#C21E56", 0.5))
abline(h = 0, lty = 3)
axis(2, at = c(-1, -0.5, 0, 0.5, 1))
text(cex = 1, x = 1:10, y = -1.4, x.labs, xpd = TRUE, srt = 35)
legend(8, 1.1, legend = c("Constant Tempertaure", "Rising Temperature"),
       pch = 19, col = c("#1434A4", "#C21E56"), cex = 0.8, bty = "n")
```
